calculate Bachi's index of age heaping

Bachi's index involves applying the Whipple method repeatedly to determine the extent of preference for each final digit. Similarly to Myers', it equals the sum of the positive deviations from 10 percent. It has a theoretical range from 0 to 90, and 10 is the expected value for each digit. Two Implementations: one following the PASEX spreadsheet SINGAGE (pasex = TRUE), with ages hard-coded, and another with flexible upper and lower age bounds, but that does not match the PASEX implementation.

Bachi(Value, Age, ageMin = 30, ageMax = 79, pasex = FALSE)

Arguments

Value
numeric. A vector of demographic counts by single age.

Age
numeric. A vector of ages corresponding to the lower integer bound of the counts.

ageMin
integer. The lowest age included in calculations. Default 25.

ageMax
integer. The upper age bound used for calculations. Default 65.

pasex
logical. Whether or not reproduce the specific age weightings in the PASEX spreadsheet. Default FALSE.

Value

The value of the index.

Details

ageMax is an inclusive upper bound, treated as interval. If you want ages 30 to 79, then give ageMin = 30 and ageMax = 79, not 80. These are only heeded if pasex = FALSE.

References

United States Census Bureau (2017). “Population Analysis System (PAS) Software.” https://www.census.gov/data/software/pas.html, https://www.census.gov/data/software/pas.html. Bachi R (1951). “The tendency to round off age returns: measurement and correction.” Bulletin of the International Statistical Institute, 33(4), 195--222. Shryock HS, Siegel JS, Larmon EA (1973). The methods and materials of demography. US Bureau of the Census.

Examples

Age <- 0:99

Bachi(pop1m_pasex, Age, ageMin = 20, ageMax = 79, pasex = TRUE) # reproduces PASEX SINGAGE

#> [1] 28.96891

Bachi(pop1m_pasex, Age, ageMin = 20, ageMax = 79) # default simpler

#> [1] 26.30127