calculate Noumbissi's digit heaping index

Noumbissi's method improves on Whipple's method by extending its basic principle to all ten digits. It compares single terminal digit numerators to denominators consisting in 5-year age groups centered on the terminal digit of age in question. For example, 9, 19, 29, etc are all of digit 9.

Noumbissi(Value, Age, ageMin = 20, ageMax = 64, digit = 0)

Arguments

Value
numeric. A vector of demographic counts by single age.

Age
numeric. A vector of ages corresponding to the lower integer bound of the counts.

ageMin
integer. The lowest age included in calculations. Default 20.

ageMax
integer. The upper age bound used for calculations. Default 64.

digit
integer. Any digit 0-9. Default 0.

Value

The value of the index.

Details

ageMin and ageMax are applied to numerator ages, not denominators. Denominators are always 5-year age groups centered on the digit in question, and these therefore stretch into ages a bit higher or lower than the numerator ages. ageMax is an inclusive upper bound, treated as interval. If you want ages 20 to 64, then give ageMin = 20 and ageMax = 64, not 65.

References

Noumbissi A (1992). “L'indice de Whipple modifié: une application aux données du Cameroun, de la Suède et de la Belgique.” Population (french edition), 1038--1041. Spoorenberg T, Dutreuilh C (2007). “Quality of age reporting: extension and application of the modified Whipple's index.” Population, 62(4), 729--741.

Examples

Age <- 0:99
Noumbissi(pop1m_pasex, Age, digit = 0) # 2.32

#> [1] 2.324269

Noumbissi(pop1m_pasex, Age, digit = 1) # 0.55

#> [1] 0.5502185

Noumbissi(pop1m_pasex, Age, digit = 2) # 0.73

#> [1] 0.7298055

Noumbissi(pop1m_pasex, Age, digit = 3) # 0.76

#> [1] 0.7610771

Noumbissi(pop1m_pasex, Age, digit = 4) # 0.49

#> [1] 0.5158693

Noumbissi(pop1m_pasex, Age, digit = 5) # 2.08

#> [1] 2.040742

Noumbissi(pop1m_pasex, Age, digit = 6) # 0.66

#> [1] 0.6597471

Noumbissi(pop1m_pasex, Age, digit = 7) # 1.08  7 looks good!

#> [1] 1.077396

Noumbissi(pop1m_pasex, Age, digit = 8) # 0.57

#> [1] 0.5700284

Noumbissi(pop1m_pasex, Age, digit = 9) # 0.59

#> [1] 0.5895384