creates stable standard based on optimizing the growth rate

The stationary standard, nLx is transformed into a stable standard by optimizing a growth rate, r such that the stable standard matches observed population counts in selected age groups. Usually the ages used for fitting are wide age groups in older ages preceding the open age group. The standard output by this function is used by OPAG to creat the standard used to redistribute counts over older age groups up to a specified open age group, such as 100.

OPAG_fit_stable_standard(
  Pop_fit,
  Age_fit,
  AgeInt_fit,
  nLx,
  Age_nLx,
  AgeInt_nLx,
  method = "uniform",
  continuous = TRUE
)

Arguments

Pop_fit
numeric vector of at least two population counts to use for fitting

Age_fit
integer vector of lower bounds for age groups of Pop_fit

AgeInt_fit
integer vector of widths of age groups of Pop_fit

nLx
numeric vector of stable population standard

Age_nLx
integer vector of lower bounds for age groups of nLx

AgeInt_nLx
optional integer vector of widths of age groups of nLx, inferred if not given.

method
character. Graduation method, default "uniform". "mono" or "pclm" would also be good choices.

continuous
logical. If TRUE we use the growth adjustment. e^(-age*r). If FALSE we assume r is geometric growth, and we use (1+r)^age for the growth adjustment.

Value

list constaining

Standard numeric vector, the transformed nLx to be used for redistribution in OPAG()

r_opt the output of optimize(), where min is the growth parameter, r

Details

The arguments method and continous don't have much leverage on the result. In short, the stable population transformation is done by ungrouping nLx to single ages (if it isn't already), and method controls which graduation method is used for this, where "uniform", "mono", "pclm" are the reasonable choices at this writing. In single ages, the difference between using a geometric r versus continuous r are quite small for this task.

Examples

Pop_fit    <- c(85000,37000)
Age_fit    <- c(70,80)
AgeInt_fit <- c(10,10)
nLx        <- downloadnLx(NULL, "Spain","female",1971)

#> Downloading nLx data for Spain, years 1971, gender female

#> New names:
#> * NA -> ...1

Age_nLx    <- names2age(nLx)

# India Males, 1991
Pop        <- smooth_age_5(pop1m_ind,
                           Age = 0:100,
                           method = "Arriaga")
Pop80      <- groupOAG(Pop, names2age(Pop), 80)
Age        <- names2age(Pop80)
AgeInt     <- age2int(Age, OAvalue = 1)

nLx        <- downloadnLx(NULL, "India","male",1991)

#> Downloading nLx data for India, years 1991, gender male

#> New names:
#> * NA -> ...1

Age_nLx    <- names2age(nLx)
AgeInt_nLx <- age2int(Age_nLx,OAvalue = 1)

Pop_fit    <- groupAges(Pop80, Age, N = 10)[c("60","70")]
Age_fit    <- c(60,70)
AgeInt_fit <- c(10,10)

Standard <- OPAG_fit_stable_standard(
             Pop_fit,
             Age_fit,
             AgeInt_fit,
             nLx = nLx,
             Age_nLx = Age_nLx,
             AgeInt_nLx = AgeInt_nLx,
             method = "uniform",
             continuous = TRUE)

# A visual comparison:
nL60 <- rescale_vector(nLx[Age_nLx >= 60])
St60p <- rescale_vector( Standard$Standard[Age_nLx >= 60] )
ages_plot <- seq(60,100,by=5)
if (FALSE) {
plot(ages_plot,nL60, type = 'l')
lines(ages_plot, St60p, col = "blue")
}