Proportionally adjust population or death counts by age to a given total.

The adjustment factor for each age is the ratio of the desired counts total divided by the initial counts total. This adjustment factor is used to adjust each age group to the new given total. This comes from the PAS spreadsheet called ADJAGE.

adjustAge(Value, DesiredTotal)

Arguments

Value
numeric. A vector of population or death counts by age.

DesiredTotal
integer. An integer giving the desired total counts.

Value

A vector of the adjusted counts by age.

Details

The age group structure of the output is the same as that of the input. This function does not adjust for unknown age groups.

References

United States Census Bureau (2017). “Population Analysis System (PAS) Software.” https://www.census.gov/data/software/pas.html, https://www.census.gov/data/software/pas.html.

Examples

Males      <- c(423, 1654, 1523, 1475, 1482, 1356, 1321, 1298, 1121,
                1023, 974, 968, 859, 824, 756, 723, 652, 622)
Females    <- c(401, 1608, 1498, 1450, 1465, 1345, 1338, 1295, 1156,
                1098, 1036, 987, 867, 882, 755, 746, 823, 875)
DesiredTotalMale       <- 23657
DesiredTotalFemale     <- 24764
adjustAge(Males, DesiredTotalMale)

#>  [1]  525.1869 2053.5676 1890.9211 1831.3254 1840.0165 1683.5778 1640.1227
#>  [8] 1611.5664 1391.8073 1270.1328 1209.2956 1201.8461 1066.5143 1023.0591
#> [15]  938.6319  897.6599  809.5079  772.2606

adjustAge(Females, DesiredTotalFemale)

#>  [1]  506.0058 2029.0707 1890.2661 1829.6968 1848.6247 1697.2015 1688.3685
#>  [8] 1634.1085 1458.7100 1385.5221 1307.2868 1245.4557 1094.0325 1112.9604
#> [15]  952.7042  941.3475 1038.5107 1104.1274

adjustAge(Males+Females, DesiredTotalMale+DesiredTotalFemale)

#>  [1] 1031.539 4083.593 3781.893 3661.714 3689.255 3381.295 3328.717 3246.094
#>  [9] 2850.503 2655.212 2516.255 2447.402 2160.724 2135.687 1891.572 1838.994
#> [17] 1846.505 1874.046

if (FALSE) {
plot(adjustAge(Males, DesiredTotalMale), t ='l', col='red',
    ylim = c(400,2100), ylab = 'The counts', xlab = 'Age groups')
lines(Males, t = 'l', col='black')
legend(15,2000, legend = c('Original','Adjusted'),
      col=c('black','red'),
      lty = 1)
}