force a (count) vector to abridged ages

This is a robustness utility, in place to avoid annoying hang-ups in LTAbr(). If data are given in non-standard ages, they are forced to standard abrdiged ages on the fly. Really this should happen prior to calling lt_abridged()

age_abridge_force(Value, AgeInt, Age)

Arguments

Value
numeric vector, presumably counts in grouped ages

AgeInt
integer vector, age interval widths

Age
integer vector, lower bounds of age groups

Details

This should be able to group up and group down as needed. graduate_mono() is used below the hood. pclm() or graduate_uniform() out to be flexible enough to do the same.

See also

Examples

V1        <- pop1m_ind
Age       <- c(0,1,3,seq(5,100,5))
AgeInt    <- c(1,2,2,rep(5,19),1)
Value     <- tapply(V1,rep(Age,times=AgeInt), sum)

is_abridged(Age)

#> [1] FALSE

age_abridge_force(Value, AgeInt, Age)

#>        0        1        5       10       15       20       25       30 
#>  9350830 43009822 57419164 51947630 42231074 37514223 34546587 29917765 
#>       35       40       45       50       55       60       65       70 
#> 27558300 22842245 18954561 16904890 10941747 11907237  6493630  5535950 
#>       75       80       85       90       95      100 
#>  2102284  2094961   601000   421930   134360    72373