The ordinary modified Beers splitting methods

This method offers both ordinary and modified Beers splitting, with an optional Demographic Analysis & Population Projection System Software adjustment johnson for ages under 10.

beers(popmat, Age = as.integer(rownames(as.matrix(popmat))),
  OAG = TRUE, method = "mod", johnson = FALSE)

Arguments

popmat
numeric. Matrix of age-period population counts in 5-year age groups with integer-labeled margins (age in rows and year in columns).

Age
numeric. A vector of ages corresponding to the lower integer bound of the counts. Detected from row names of popmat if missing.

OAG
logical. Whether or not the final age group open. Default TRUE.

method
character. Valid values are "mod" or "ord". Default "mod".

johnson
logical. Whether or not to adjust young ages according to the Demographic Analysis & Population Projection System Software method. Default FALSE.

Value

An age-period matrix of split population counts with the same number of columns as popmat, and single ages in rows.

Details

Ages should refer to lower age bounds. The rows of popmat must be labelled with ages unless Age is given separately. There must be at least six 5-year age groups (including the open group, 5 otherwise). One year of data will work as well, as long as it is given as or coercible to a single-column matrix. If you want the johnson adjustment then popmat must contain a single-year estimate of the population count in age 0. That means popmat must come either as standard abridged or single age data.

If the highest age does not end in a 0 or 5, and OAG == TRUE, then the open age will be grouped down to the next highest age ending in 0 or 5. If the highest age does not end in a 0 or 5, and OAG == FALSE, then results extend to single ages covering the entire 5-year age group.

References

Beers H (1945). “Modified-interpolation formulas that minimize fourth differences.” Record of the American Institute of Actuaries, 34(69), 14--20. Shryock HS, Siegel JS, Larmon EA (1973). The methods and materials of demography. US Bureau of the Census. Siegel Jacob S, Swanson David A (eds.) (2004). The Methods and Materials of Demography, 2 edition. Elsevier Academic Press, California, USA, San Diego, USA. Stover J, Johnson P, Zaba B, Zwahlen M, Dabis F, Ekpini R (2008). “The Spectrum projection package: improvements in estimating mortality, ART needs, PMTCT impact and uncertainty bounds.” Sexually Transmitted Infections, 84(Suppl 1), i24--i30.

Examples

p5 <- pop5_mat
head(p5) # this is the entire matrix

#>     1950  1951  1952  1953  1954
#> 0  54170 57424 60272 62727 64816
#> 5  44775 44475 44780 45681 47137
#> 10 42142 41752 41804 42101 42508
#> 15 38464 39628 40229 40474 40532
#> 20 34406 34757 35155 35599 36083
#> 25 30386 30605 30978 31439 31940

p1 <- beers(p5, OAG = FALSE)
head(p1)

#>        1950      1951      1952     1953     1954
#> 0 11952.235 13054.267 13941.546 14603.87 15046.71
#> 1 11325.417 12173.033 12881.400 13449.94 13885.44
#> 2 10766.300 11388.299 11937.826 12420.70 12843.69
#> 3 10274.882 10700.066 11110.827 11516.16 11921.45
#> 4  9851.165 10108.335 10400.401 10736.33 11118.72
#> 5  9494.302  9612.476  9805.927 10080.56 10434.97

# note some negatives in high ages
tail(p1)

#>        1950    1951    1952    1953    1954
#> 99   0.3184  0.1410 -0.0487 -0.1860 -0.1520
#> 100 -0.0886 -0.2388 -0.4086 -0.4984 -0.4184
#> 101 -0.2724 -0.3250 -0.4433 -0.4860 -0.3520
#> 102 -0.2281 -0.1056 -0.1390 -0.1368  0.0572
#> 103  0.0443  0.4194  0.5043  0.5492  0.8092
#> 104  0.5448  1.2500  1.4866  1.5720  1.9040

colSums(p1) - colSums(p5)

#> 1950 1951 1952 1953 1954 
#>    0    0    0    0    0 

# another case, starting with single ages
# note beers() groups ages.
Value        <- pop1m_ind
Age          <- 0:100
names(Value) <- Age
ord1 <-  beers(Value, OAG = TRUE, method = "ord")
mod1 <- beers(Value, OAG = TRUE, method = "mod")
if (FALSE) {
plot(Age,Value,
ylab = 'The counts', xlab = 'Age groups')
lines(Age, ord1, t='l', col='blue')
lines(Age, mod1, t = 'l', col ='red', lty =2)
legend(80,15000000,
      legend = c('Ordinary',
                 'Modified'),
      col=c('blue', 'red'),
      lty = c(1,2))
}

# notice this negative value. Yuck!
tail(mod1)

#>          [,1]
#> 95   59273.54
#> 96   47761.30
#> 97   31560.53
#> 98   10671.23
#> 99  -14906.60
#> 100  72373.00

# this replaces ages 90+, guaranteed no negatives.
graduate_mono_closeout(Value, pops = mod1, OAG = TRUE)

#> Error in graduate_mono_closeout(Value, pops = mod1, OAG = TRUE): argument "Age" is missing, with no default

# Note: there are no kludges built into beers() to handle such cases.
# these ought to be handled by wrappers as appropriate.

# This replicates Johnson_2016_BEERSP.XLS, males
M <- c(184499,752124-184499,582662,463534,369976,286946,235867,
                199561,172133,151194,131502,113439,95614,
                78777,60157,40960,21318,25451)
dim(M)      <- c(length(M),1)
rownames(M) <- c(0,1,seq(5,80,by=5))
Age0        <- 184499
johnson     <- beers(
                         popmat = M,
                         OAG = TRUE,
                             method = "ord",
                             johnson = TRUE)