The basic Grabill age-splitting method

This method uses Grabill's redistribution of middle ages and blends into Sprague estimated single-age population counts for the first and final ten ages. Open age groups are preserved, as are annual totals.

grabill(popmat, Age = as.integer(rownames(as.matrix(popmat))),
  OAG = TRUE)

Arguments

popmat
numeric. Matrix of age-period population counts in 5-year age groups with integer-labeled margins (age in rows and year in columns).

Age
integer. A vector of ages corresponding to the lower integer bound of the counts. Detected from row names of popmat if missing.

OAG
logical. Whether or not the final age group open. Default TRUE.

Value

An age-period matrix of split population counts with the same number of columns as popmat, and single ages in rows.

Details

Dimension labelling is necessary. There must be at least six age groups (including the open group). One year of data will work as well, as long as it's given as a single-column matrix. Data may be given in either single or grouped ages.

If the highest age does not end in a 0 or 5, and OAG == TRUE, then the open age will be grouped down to the next highest age ending in 0 or 5. If the highest age does not end in a 0 or 5, and OAG == FALSE, then results extend to single ages covering the entire 5-year age group.

References

Shryock HS, Siegel JS, Larmon EA (1973). The methods and materials of demography. US Bureau of the Census.

Examples

p5 <- pop5_mat
head(p5) # this is the entire matrix

#>     1950  1951  1952  1953  1954
#> 0  54170 57424 60272 62727 64816
#> 5  44775 44475 44780 45681 47137
#> 10 42142 41752 41804 42101 42508
#> 15 38464 39628 40229 40474 40532
#> 20 34406 34757 35155 35599 36083
#> 25 30386 30605 30978 31439 31940

p1g <- grabill(p5)
head(p1g); tail(p1g)

#>        1950      1951      1952      1953     1954
#> 0 12171.749 13334.190 14267.068 14941.309 15352.26
#> 1 11378.972 12241.735 12961.457 13532.900 13960.38
#> 2 10714.699 11322.949 11862.101 12342.137 12772.05
#> 3 10166.274 10562.348 10951.183 11350.568 11770.56
#> 4  9721.429  9945.055 10211.645 10540.549 10939.95
#> 5  9373.506  9462.950  9633.993  9902.438 10272.29

#>          1950      1951      1952      1953      1954
#> 95  5.8921418  7.132792  7.364145  6.803432  6.470227
#> 96  4.5129521  5.534307  5.669743  5.150564  4.823490
#> 97  3.4262982  3.769919  3.697894  3.346066  3.170257
#> 98  2.3206541  1.440895  1.032992  1.000963  1.152927
#> 99  0.8946524 -1.842648 -2.732157 -2.265021 -1.577096
#> 100 0.0000000  1.000000  1.000000  1.000000  2.000000

colSums(p1g) - colSums(p5)

#> 1950 1951 1952 1953 1954 
#>    0    0    0    0    0 

p1s <- sprague(p5)
if (FALSE) {
plot(seq(0,100,by=5),p5[,1]/5,type = "s", col = "gray", xlab = "Age", ylab = "Count")
lines(0:100, p1g[,1], col = "red", lwd = 2)
lines(0:100, p1s[,1], col = "blue", lty = 2, lwd =2)
legend("topright",
        lty = c(1,1,2),
        col = c("gray","red","blue"),
        lwd = c(1,2,1),
        legend = c("grouped","Grabill", "Sprague"))
}

# also works for single ages:
names(pop1m_ind) <- 0:100
grab1 <- grabill(pop1m_ind)
if (FALSE) {
plot(0:100, pop1m_ind)
lines(0:100, c(grab1))
}