Graduate grouped data

A wrapper function for several graduation methods, primarily for count data ("sprague", "beers(ord)", "beers(mod)", "mono" (Monotonic spline), "uniform", "pclm"), but also with one ("pclm") with an option for graduating rates if both event counts and population at risk are available.

graduate(Value, Age, AgeInt = age2int(Age), OAG = TRUE,
  OAnew = max(Age), method = c("sprague", "beers(ord)", "beers(mod)",
  "pclm", "mono", "uniform"), keep0 = FALSE, ...)

Arguments

Value
numeric vector, presumably counts in grouped ages

Age
integer vector, lower bounds of age groups

AgeInt
integer vector, age interval widths

OAG
logical, default = TRUE is the final age group open?

OAnew
integer, optional new open age, higher than max(Age). See details.

method
character, either "sprague", "beers(ord)"), "beers(mod)"), "mono"), "uniform"), or "pclm"

keep0
logical. Default FALSE. If available, should the value in the infant age group be maintained, and ages 1-4 constrained?

...
extra arguments passed to beers() or graduate_pclm()

Details

"sprague", "beers(ord)", "beers(mod)" methods require original data to be in uniform five-year age groups. If they are not (for example, the infant group is separate) then they are grouped to uniform width prior to splitting. If you want to keep the original infant count in output, then specify keep0 = TRUE. In this case, it is imputed, and ages 1-4 are rescaled, which may introduce a discontinuity in results from age 4 to 5. keep0 = TRUE may also be desired along with method = "pclm".

Some methods are constrained, others not, and others are optionally constrained. If this is required, then this function can be followed up with rescaleAgeGroups(), which may have the effect of breaking continuity in smooth results. This is inconsequential for downstream demography, but if this aesthetic side effect is undesired, then try one of the constrained methods: "sprague", "mono", "pclm" (with control = list(lambda = 1/1e7) specified or similar).

Beers may either be ordinary "beers(ord)" or modified "beers(mod)", and either can pass on the optional argument johnson = TRUE if desired (this has a different distribution pattern for young ages, FALSE by default). If method = "beers" is given, then "beers(ord)" is used.

This wrapper standardizes some inconsistencies in how open ages are dealt with. For example, with the "pclm" method, the last age group can be redistributed over a specified interval implied by increase OAnew beyond the range of Age. To get this same behavior from "mono", or "uniform" specify OAG = FALSE along with an appropriately high OAnew (or integer final value of AgeInt.

OAnew cannot be higher than max(Age)+4 for "sprague" or "beers" methods. For "uniform","mono","pclm" it can be higher than this, and in each case the open age group is completely redistributed within this range, meaning it's not really open anymore.

References

Pascariu MD, Danko MJ, Sch昼㸶ley J, Rizzi S (2018). “ungroup: An R package for efficient estimation of smooth distributions from coarsely binned data.” Age, 10, 0. Rizzi S, Gampe J, Eilers PH (2015). “Efficient estimation of smooth distributions from coarsely grouped data.” American journal of epidemiology, 182(2), 138--147. Sprague TB (1880). “Explanation of a new formula for interpolation.” Journal of the Institute of Actuaries, 22(4), 270--285. Shryock HS, Siegel JS, Larmon EA (1973). The methods and materials of demography. US Bureau of the Census. Siegel Jacob S, Swanson David A (eds.) (2004). The Methods and Materials of Demography, 2 edition. Elsevier Academic Press, California, USA, San Diego, USA. Beers H (1945). “Modified-interpolation formulas that minimize fourth differences.” Record of the American Institute of Actuaries, 34(69), 14--20.

See also

Examples

Value <- pop5_mat[, 1]
Value <- c(10000,44170,Value[-1])
Age   <- sort(c(1,seq(0,100,by=5)))

graduate(Value, Age, method = "sprague")

#>           [,1]
#> 0   12172.4912
#> 1   11380.4480
#> 2   10717.4480
#> 3   10171.0000
#> 4    9728.6128
#> 5    9377.7952
#> 6    9106.0560
#> 7    8900.9040
#> 8    8749.8480
#> 9    8640.3968
#> 10   8573.6144
#> 11   8550.5648
#> 12   8490.9808
#> 13   8355.2608
#> 14   8171.5792
#> 15   8008.3808
#> 16   7852.7792
#> 17   7694.2592
#> 18   7534.6752
#> 19   7373.9056
#> 20   7210.8208
#> 21   7046.6624
#> 22   6881.6064
#> 23   6715.9664
#> 24   6550.9440
#> 25   6386.5216
#> 26   6221.6160
#> 27   6065.6560
#> 28   5922.9920
#> 29   5789.2144
#> 30   5656.4784
#> 31   5527.4512
#> 32   5393.8752
#> 33   5251.2032
#> 34   5103.9920
#> 35   4958.4544
#> 36   4809.8784
#> 37   4676.3904
#> 38   4566.5184
#> 39   4469.7584
#> 40   4372.0480
#> 41   4282.1632
#> 42   4161.7152
#> 43   3992.0912
#> 44   3793.9824
#> 45   3600.3712
#> 46   3397.0752
#> 47   3237.6592
#> 48   3149.6752
#> 49   3104.2192
#> 50   3052.7728
#> 51   3014.5648
#> 52   2923.2688
#> 53   2744.0448
#> 54   2513.3488
#> 55   2294.2640
#> 56   2064.3184
#> 57   1897.9424
#> 58   1835.3744
#> 59   1836.1008
#> 60   1826.9504
#> 61   1831.6544
#> 62   1779.1824
#> 63   1629.7344
#> 64   1422.4784
#> 65   1231.8448
#> 66   1037.5024
#> 67    891.8304
#> 68    826.6464
#> 69    813.1760
#> 70    792.0960
#> 71    776.7936
#> 72    745.8176
#> 73    683.0176
#> 74    601.2752
#> 75    529.6992
#> 76    462.5600
#> 77    401.2800
#> 78    349.5120
#> 79    304.9488
#> 80    261.4336
#> 81    219.9616
#> 82    183.1696
#> 83    151.6816
#> 84    124.7536
#> 85    100.7200
#> 86     79.5568
#> 87     61.8448
#> 88     47.5888
#> 89     36.2896
#> 90     27.1456
#> 91     19.9600
#> 92     14.5360
#> 93     10.5760
#> 94      7.7824
#> 95      5.8576
#> 96      4.5040
#> 97      3.4240
#> 98      2.3200
#> 99      0.8944
#> 100     0.0000

graduate(Value, Age, method = "sprague", keep0=FALSE)

#>           [,1]
#> 0   12172.4912
#> 1   11380.4480
#> 2   10717.4480
#> 3   10171.0000
#> 4    9728.6128
#> 5    9377.7952
#> 6    9106.0560
#> 7    8900.9040
#> 8    8749.8480
#> 9    8640.3968
#> 10   8573.6144
#> 11   8550.5648
#> 12   8490.9808
#> 13   8355.2608
#> 14   8171.5792
#> 15   8008.3808
#> 16   7852.7792
#> 17   7694.2592
#> 18   7534.6752
#> 19   7373.9056
#> 20   7210.8208
#> 21   7046.6624
#> 22   6881.6064
#> 23   6715.9664
#> 24   6550.9440
#> 25   6386.5216
#> 26   6221.6160
#> 27   6065.6560
#> 28   5922.9920
#> 29   5789.2144
#> 30   5656.4784
#> 31   5527.4512
#> 32   5393.8752
#> 33   5251.2032
#> 34   5103.9920
#> 35   4958.4544
#> 36   4809.8784
#> 37   4676.3904
#> 38   4566.5184
#> 39   4469.7584
#> 40   4372.0480
#> 41   4282.1632
#> 42   4161.7152
#> 43   3992.0912
#> 44   3793.9824
#> 45   3600.3712
#> 46   3397.0752
#> 47   3237.6592
#> 48   3149.6752
#> 49   3104.2192
#> 50   3052.7728
#> 51   3014.5648
#> 52   2923.2688
#> 53   2744.0448
#> 54   2513.3488
#> 55   2294.2640
#> 56   2064.3184
#> 57   1897.9424
#> 58   1835.3744
#> 59   1836.1008
#> 60   1826.9504
#> 61   1831.6544
#> 62   1779.1824
#> 63   1629.7344
#> 64   1422.4784
#> 65   1231.8448
#> 66   1037.5024
#> 67    891.8304
#> 68    826.6464
#> 69    813.1760
#> 70    792.0960
#> 71    776.7936
#> 72    745.8176
#> 73    683.0176
#> 74    601.2752
#> 75    529.6992
#> 76    462.5600
#> 77    401.2800
#> 78    349.5120
#> 79    304.9488
#> 80    261.4336
#> 81    219.9616
#> 82    183.1696
#> 83    151.6816
#> 84    124.7536
#> 85    100.7200
#> 86     79.5568
#> 87     61.8448
#> 88     47.5888
#> 89     36.2896
#> 90     27.1456
#> 91     19.9600
#> 92     14.5360
#> 93     10.5760
#> 94      7.7824
#> 95      5.8576
#> 96      4.5040
#> 97      3.4240
#> 98      2.3200
#> 99      0.8944
#> 100     0.0000

graduate(Value, Age, method = "beers(ord)")

#>           [,1]
#> 0   11932.7336
#> 1   11342.3240
#> 2   10788.6128
#> 3   10280.2888
#> 4    9826.0408
#> 5    9435.4048
#> 6    9117.0696
#> 7    8877.1824
#> 8    8717.6544
#> 9    8627.6888
#> 10   8582.9336
#> 11   8547.1760
#> 12   8479.1200
#> 13   8351.8720
#> 14   8180.8984
#> 15   8008.1091
#> 16   7852.8780
#> 17   7694.6050
#> 18   7534.7740
#> 19   7373.6339
#> 20   7210.9429
#> 21   7046.6180
#> 22   6881.4510
#> 23   6715.9220
#> 24   6551.0661
#> 25   6385.3171
#> 26   6222.0540
#> 27   6067.1890
#> 28   5923.4300
#> 29   5788.0099
#> 30   5657.7302
#> 31   5526.9960
#> 32   5392.2820
#> 33   5250.7480
#> 34   5105.2438
#> 35   4955.8375
#> 36   4810.8300
#> 37   4679.7210
#> 38   4567.4700
#> 39   4467.1415
#> 40   4377.4523
#> 41   4280.1980
#> 42   4154.8370
#> 43   3990.1260
#> 44   3799.3867
#> 45   3592.6085
#> 46   3399.8980
#> 47   3247.5390
#> 48   3152.4980
#> 49   3096.4565
#> 50   3062.5760
#> 51   3011.0000
#> 52   2910.7920
#> 53   2740.4800
#> 54   2523.1520
#> 55   2283.1606
#> 56   2068.3560
#> 57   1912.0740
#> 58   1839.4120
#> 59   1824.9974
#> 60   1837.8085
#> 61   1827.7060
#> 62   1765.3630
#> 63   1625.7860
#> 64   1433.3365
#> 65   1223.5134
#> 66   1040.5320
#> 67    902.4340
#> 68    829.6760
#> 69    804.8446
#> 70    796.0879
#> 71    775.3420
#> 72    740.7370
#> 73    681.5660
#> 74    605.2671
#> 75    528.8797
#> 76    462.8580
#> 77    402.3230
#> 78    349.8100
#> 79    304.1293
#> 80    261.2455
#> 81    220.0300
#> 82    183.4090
#> 83    151.7500
#> 84    124.5655
#> 85    100.6507
#> 86     79.5820
#> 87     61.9330
#> 88     47.6140
#> 89     36.2203
#> 90     27.2401
#> 91     20.1994
#> 92     14.7124
#> 93     10.4941
#> 94      7.3540
#> 95      5.1331
#> 96      3.6913
#> 97      2.8948
#> 98      2.6035
#> 99      2.6773
#> 100     0.0000

graduate(Value, Age, method = "beers(ord)", keep0=TRUE)

#>           [,1]
#> 0   10000.0000
#> 1   11861.3370
#> 2   11282.2886
#> 3   10750.7042
#> 4   10275.6703
#> 5    9435.4048
#> 6    9117.0696
#> 7    8877.1824
#> 8    8717.6544
#> 9    8627.6888
#> 10   8582.9336
#> 11   8547.1760
#> 12   8479.1200
#> 13   8351.8720
#> 14   8180.8984
#> 15   8008.1091
#> 16   7852.8780
#> 17   7694.6050
#> 18   7534.7740
#> 19   7373.6339
#> 20   7210.9429
#> 21   7046.6180
#> 22   6881.4510
#> 23   6715.9220
#> 24   6551.0661
#> 25   6385.3171
#> 26   6222.0540
#> 27   6067.1890
#> 28   5923.4300
#> 29   5788.0099
#> 30   5657.7302
#> 31   5526.9960
#> 32   5392.2820
#> 33   5250.7480
#> 34   5105.2438
#> 35   4955.8375
#> 36   4810.8300
#> 37   4679.7210
#> 38   4567.4700
#> 39   4467.1415
#> 40   4377.4523
#> 41   4280.1980
#> 42   4154.8370
#> 43   3990.1260
#> 44   3799.3867
#> 45   3592.6085
#> 46   3399.8980
#> 47   3247.5390
#> 48   3152.4980
#> 49   3096.4565
#> 50   3062.5760
#> 51   3011.0000
#> 52   2910.7920
#> 53   2740.4800
#> 54   2523.1520
#> 55   2283.1606
#> 56   2068.3560
#> 57   1912.0740
#> 58   1839.4120
#> 59   1824.9974
#> 60   1837.8085
#> 61   1827.7060
#> 62   1765.3630
#> 63   1625.7860
#> 64   1433.3365
#> 65   1223.5134
#> 66   1040.5320
#> 67    902.4340
#> 68    829.6760
#> 69    804.8446
#> 70    796.0879
#> 71    775.3420
#> 72    740.7370
#> 73    681.5660
#> 74    605.2671
#> 75    528.8797
#> 76    462.8580
#> 77    402.3230
#> 78    349.8100
#> 79    304.1293
#> 80    261.2455
#> 81    220.0300
#> 82    183.4090
#> 83    151.7500
#> 84    124.5655
#> 85    100.6507
#> 86     79.5820
#> 87     61.9330
#> 88     47.6140
#> 89     36.2203
#> 90     27.2401
#> 91     20.1994
#> 92     14.7124
#> 93     10.4941
#> 94      7.3540
#> 95      5.1331
#> 96      3.6913
#> 97      2.8948
#> 98      2.6035
#> 99      2.6773
#> 100     0.0000

graduate(Value, Age, method = "beers(ord)", keep0=TRUE, johnson = TRUE)

#>           [,1]
#> 0   10000.0000
#> 1   11342.3240
#> 2   11462.1335
#> 3   11023.6236
#> 4   10341.9189
#> 5    9651.0947
#> 6    9097.9694
#> 7    8762.3499
#> 8    8635.8972
#> 9    8627.6888
#> 10   8582.9336
#> 11   8547.1760
#> 12   8479.1200
#> 13   8351.8720
#> 14   8180.8984
#> 15   8008.1091
#> 16   7852.8780
#> 17   7694.6050
#> 18   7534.7740
#> 19   7373.6339
#> 20   7210.9429
#> 21   7046.6180
#> 22   6881.4510
#> 23   6715.9220
#> 24   6551.0661
#> 25   6385.3171
#> 26   6222.0540
#> 27   6067.1890
#> 28   5923.4300
#> 29   5788.0099
#> 30   5657.7302
#> 31   5526.9960
#> 32   5392.2820
#> 33   5250.7480
#> 34   5105.2438
#> 35   4955.8375
#> 36   4810.8300
#> 37   4679.7210
#> 38   4567.4700
#> 39   4467.1415
#> 40   4377.4523
#> 41   4280.1980
#> 42   4154.8370
#> 43   3990.1260
#> 44   3799.3867
#> 45   3592.6085
#> 46   3399.8980
#> 47   3247.5390
#> 48   3152.4980
#> 49   3096.4565
#> 50   3062.5760
#> 51   3011.0000
#> 52   2910.7920
#> 53   2740.4800
#> 54   2523.1520
#> 55   2283.1606
#> 56   2068.3560
#> 57   1912.0740
#> 58   1839.4120
#> 59   1824.9974
#> 60   1837.8085
#> 61   1827.7060
#> 62   1765.3630
#> 63   1625.7860
#> 64   1433.3365
#> 65   1223.5134
#> 66   1040.5320
#> 67    902.4340
#> 68    829.6760
#> 69    804.8446
#> 70    796.0879
#> 71    775.3420
#> 72    740.7370
#> 73    681.5660
#> 74    605.2671
#> 75    528.8797
#> 76    462.8580
#> 77    402.3230
#> 78    349.8100
#> 79    304.1293
#> 80    261.2455
#> 81    220.0300
#> 82    183.4090
#> 83    151.7500
#> 84    124.5655
#> 85    100.6507
#> 86     79.5820
#> 87     61.9330
#> 88     47.6140
#> 89     36.2203
#> 90     27.2401
#> 91     20.1994
#> 92     14.7124
#> 93     10.4941
#> 94      7.3540
#> 95      5.1331
#> 96      3.6913
#> 97      2.8948
#> 98      2.6035
#> 99      2.6773
#> 100     0.0000

graduate(Value, Age, method = "beers(mod)")

#>           [,1]
#> 0   11952.2354
#> 1   11325.4173
#> 2   10766.2996
#> 3   10274.8823
#> 4    9851.1654
#> 5    9494.3017
#> 6    9201.8161
#> 7    8969.8051
#> 8    8790.9099
#> 9    8653.8682
#> 10   8546.7035
#> 11   8455.2302
#> 12   8364.7480
#> 13   8264.2114
#> 14   8146.8109
#> 15   8009.2966
#> 16   7855.8211
#> 17   7697.9395
#> 18   7537.0673
#> 19   7374.4965
#> 20   7210.6243
#> 21   7045.7255
#> 22   6879.9525
#> 23   6714.4613
#> 24   6550.4634
#> 25   6389.8558
#> 26   6233.6499
#> 27   6081.9715
#> 28   5935.0481
#> 29   5792.5597
#> 30   5653.0084
#> 31   5514.9350
#> 32   5376.9190
#> 33   5238.6834
#> 34   5100.5202
#> 35   4965.5251
#> 36   4835.6769
#> 37   4711.8375
#> 38   4593.0579
#> 39   4477.1996
#> 40   4357.2582
#> 41   4228.5099
#> 42   4088.5115
#> 43   3937.6585
#> 44   3778.8029
#> 45   3621.7683
#> 46   3474.4487
#> 47   3342.8085
#> 48   3227.5557
#> 49   3125.8698
#> 50   3025.7427
#> 51   2916.8358
#> 52   2790.4800
#> 53   2645.7098
#> 54   2486.0157
#> 55   2325.0249
#> 56   2175.3008
#> 57   2048.3430
#> 58   1946.4600
#> 59   1866.9133
#> 60   1796.7846
#> 61   1722.9549
#> 62   1632.1045
#> 63   1521.2719
#> 64   1392.4311
#> 65   1255.2308
#> 66   1121.3870
#> 67   1004.6830
#> 68    909.3898
#> 69    835.9914
#> 70    780.7211
#> 71    736.2617
#> 72    691.7455
#> 73    643.7115
#> 74    590.5132
#> 75    532.0976
#> 76    471.0073
#> 77    412.3805
#> 78    357.4547
#> 79    307.0949
#> 80    261.9439
#> 81    221.8201
#> 82    185.7175
#> 83    153.5851
#> 84    125.2864
#> 85    100.8658
#> 86     80.1571
#> 87     62.7835
#> 88     48.3745
#> 89     36.5281
#> 90     26.8663
#> 91     19.0789
#> 92     12.9361
#> 93      8.2390
#> 94      4.8817
#> 95      2.7712
#> 96      1.8763
#> 97      2.1907
#> 98      3.7144
#> 99      6.4474
#> 100     0.0000

graduate(Value, Age, method = "beers(mod)", keep0=TRUE)

#>           [,1]
#> 0   10000.0000
#> 1   11849.1277
#> 2   11264.1552
#> 3   10750.0138
#> 4   10306.7033
#> 5    9494.3017
#> 6    9201.8161
#> 7    8969.8051
#> 8    8790.9099
#> 9    8653.8682
#> 10   8546.7035
#> 11   8455.2302
#> 12   8364.7480
#> 13   8264.2114
#> 14   8146.8109
#> 15   8009.2966
#> 16   7855.8211
#> 17   7697.9395
#> 18   7537.0673
#> 19   7374.4965
#> 20   7210.6243
#> 21   7045.7255
#> 22   6879.9525
#> 23   6714.4613
#> 24   6550.4634
#> 25   6389.8558
#> 26   6233.6499
#> 27   6081.9715
#> 28   5935.0481
#> 29   5792.5597
#> 30   5653.0084
#> 31   5514.9350
#> 32   5376.9190
#> 33   5238.6834
#> 34   5100.5202
#> 35   4965.5251
#> 36   4835.6769
#> 37   4711.8375
#> 38   4593.0579
#> 39   4477.1996
#> 40   4357.2582
#> 41   4228.5099
#> 42   4088.5115
#> 43   3937.6585
#> 44   3778.8029
#> 45   3621.7683
#> 46   3474.4487
#> 47   3342.8085
#> 48   3227.5557
#> 49   3125.8698
#> 50   3025.7427
#> 51   2916.8358
#> 52   2790.4800
#> 53   2645.7098
#> 54   2486.0157
#> 55   2325.0249
#> 56   2175.3008
#> 57   2048.3430
#> 58   1946.4600
#> 59   1866.9133
#> 60   1796.7846
#> 61   1722.9549
#> 62   1632.1045
#> 63   1521.2719
#> 64   1392.4311
#> 65   1255.2308
#> 66   1121.3870
#> 67   1004.6830
#> 68    909.3898
#> 69    835.9914
#> 70    780.7211
#> 71    736.2617
#> 72    691.7455
#> 73    643.7115
#> 74    590.5132
#> 75    532.0976
#> 76    471.0073
#> 77    412.3805
#> 78    357.4547
#> 79    307.0949
#> 80    261.9439
#> 81    221.8201
#> 82    185.7175
#> 83    153.5851
#> 84    125.2864
#> 85    100.8658
#> 86     80.1571
#> 87     62.7835
#> 88     48.3745
#> 89     36.5281
#> 90     26.8663
#> 91     19.0789
#> 92     12.9361
#> 93      8.2390
#> 94      4.8817
#> 95      2.7712
#> 96      1.8763
#> 97      2.1907
#> 98      3.7144
#> 99      6.4474
#> 100     0.0000

graduate(Value, Age, method = "beers(mod)", keep0=TRUE, johnson = TRUE)

#>           [,1]
#> 0   10000.0000
#> 1   11325.4173
#> 2   11441.7180
#> 3   11025.7228
#> 4   10377.1420
#> 5    9717.5927
#> 6    9184.3461
#> 7    8850.5817
#> 8    8704.3123
#> 9    8653.8682
#> 10   8546.7035
#> 11   8455.2302
#> 12   8364.7480
#> 13   8264.2114
#> 14   8146.8109
#> 15   8009.2966
#> 16   7855.8211
#> 17   7697.9395
#> 18   7537.0673
#> 19   7374.4965
#> 20   7210.6243
#> 21   7045.7255
#> 22   6879.9525
#> 23   6714.4613
#> 24   6550.4634
#> 25   6389.8558
#> 26   6233.6499
#> 27   6081.9715
#> 28   5935.0481
#> 29   5792.5597
#> 30   5653.0084
#> 31   5514.9350
#> 32   5376.9190
#> 33   5238.6834
#> 34   5100.5202
#> 35   4965.5251
#> 36   4835.6769
#> 37   4711.8375
#> 38   4593.0579
#> 39   4477.1996
#> 40   4357.2582
#> 41   4228.5099
#> 42   4088.5115
#> 43   3937.6585
#> 44   3778.8029
#> 45   3621.7683
#> 46   3474.4487
#> 47   3342.8085
#> 48   3227.5557
#> 49   3125.8698
#> 50   3025.7427
#> 51   2916.8358
#> 52   2790.4800
#> 53   2645.7098
#> 54   2486.0157
#> 55   2325.0249
#> 56   2175.3008
#> 57   2048.3430
#> 58   1946.4600
#> 59   1866.9133
#> 60   1796.7846
#> 61   1722.9549
#> 62   1632.1045
#> 63   1521.2719
#> 64   1392.4311
#> 65   1255.2308
#> 66   1121.3870
#> 67   1004.6830
#> 68    909.3898
#> 69    835.9914
#> 70    780.7211
#> 71    736.2617
#> 72    691.7455
#> 73    643.7115
#> 74    590.5132
#> 75    532.0976
#> 76    471.0073
#> 77    412.3805
#> 78    357.4547
#> 79    307.0949
#> 80    261.9439
#> 81    221.8201
#> 82    185.7175
#> 83    153.5851
#> 84    125.2864
#> 85    100.8658
#> 86     80.1571
#> 87     62.7835
#> 88     48.3745
#> 89     36.5281
#> 90     26.8663
#> 91     19.0789
#> 92     12.9361
#> 93      8.2390
#> 94      4.8817
#> 95      2.7712
#> 96      1.8763
#> 97      2.1907
#> 98      3.7144
#> 99      6.4474
#> 100     0.0000

graduate(Value, Age, method = "mono")

#>         0         1         2         3         4         5         6         7 
#> 10000.000 10945.000 11401.250 11270.625 10553.125  9665.128  9122.756  8767.692 
#>         8         9        10        11        12        13        14        15 
#>  8599.936  8619.488  8655.760  8567.160  8453.480  8314.720  8150.880  7993.120 
#>        16        17        18        19        20        21        22        23 
#>  7852.080  7701.920  7542.640  7374.240  7205.232  7042.304  6880.288  6719.184 
#>        24        25        26        27        28        29        30        31 
#>  6558.992  6389.728  6219.856  6063.592  5920.936  5791.888  5662.824  5524.688 
#>        32        33        34        35        36        37        38        39 
#>  5386.576  5248.488  5110.424  4963.192  4815.944  4682.448  4562.704  4456.712 
#>        40        41        42        43        44        45        46        47 
#>  4370.464  4275.048  4150.016  3995.368  3811.104  3596.888  3402.416  3252.872 
#>        48        49        50        51        52        53        54        55 
#>  3148.256  3088.568  3062.144  3005.768  2899.496  2743.328  2537.264  2285.088 
#>        56        57        58        59        60        61        62        63 
#>  2066.176  1916.432  1835.856  1824.448  1849.056  1827.552  1752.024  1622.472 
#>        64        65        66        67        68        69        70        71 
#>  1438.896  1215.528  1028.176   900.512   832.536   824.248   821.544   779.048 
#>        72        73        74        75        76        77        78        79 
#>   728.176   668.928   601.304   526.576   457.432   398.944   351.112   313.936 
#>        80        81        82        83        84        85        86        87 
#>   268.888   216.736   176.392   147.856   131.128   108.496    77.992    56.344 
#>        88        89        90        91        92        93        94        95 
#>    43.552    39.616    32.752    19.984    11.608     7.624     8.032     7.432 
#>        96        97        98        99       100 
#>     3.904     1.888     1.384     2.392     0.000 

graduate(Value, Age, method = "mono", keep0=TRUE)

#>         0         1         2         3         4         5         6         7 
#> 10000.000 10945.000 11401.250 11270.625 10553.125  9665.128  9122.756  8767.692 
#>         8         9        10        11        12        13        14        15 
#>  8599.936  8619.488  8655.760  8567.160  8453.480  8314.720  8150.880  7993.120 
#>        16        17        18        19        20        21        22        23 
#>  7852.080  7701.920  7542.640  7374.240  7205.232  7042.304  6880.288  6719.184 
#>        24        25        26        27        28        29        30        31 
#>  6558.992  6389.728  6219.856  6063.592  5920.936  5791.888  5662.824  5524.688 
#>        32        33        34        35        36        37        38        39 
#>  5386.576  5248.488  5110.424  4963.192  4815.944  4682.448  4562.704  4456.712 
#>        40        41        42        43        44        45        46        47 
#>  4370.464  4275.048  4150.016  3995.368  3811.104  3596.888  3402.416  3252.872 
#>        48        49        50        51        52        53        54        55 
#>  3148.256  3088.568  3062.144  3005.768  2899.496  2743.328  2537.264  2285.088 
#>        56        57        58        59        60        61        62        63 
#>  2066.176  1916.432  1835.856  1824.448  1849.056  1827.552  1752.024  1622.472 
#>        64        65        66        67        68        69        70        71 
#>  1438.896  1215.528  1028.176   900.512   832.536   824.248   821.544   779.048 
#>        72        73        74        75        76        77        78        79 
#>   728.176   668.928   601.304   526.576   457.432   398.944   351.112   313.936 
#>        80        81        82        83        84        85        86        87 
#>   268.888   216.736   176.392   147.856   131.128   108.496    77.992    56.344 
#>        88        89        90        91        92        93        94        95 
#>    43.552    39.616    32.752    19.984    11.608     7.624     8.032     7.432 
#>        96        97        98        99       100 
#>     3.904     1.888     1.384     2.392     0.000 

graduate(Value, Age, method = "uniform")

#>       0       1       2       3       4       5       6       7       8       9 
#> 10000.0 11042.5 11042.5 11042.5 11042.5  8955.0  8955.0  8955.0  8955.0  8955.0 
#>      10      11      12      13      14      15      16      17      18      19 
#>  8428.4  8428.4  8428.4  8428.4  8428.4  7692.8  7692.8  7692.8  7692.8  7692.8 
#>      20      21      22      23      24      25      26      27      28      29 
#>  6881.2  6881.2  6881.2  6881.2  6881.2  6077.2  6077.2  6077.2  6077.2  6077.2 
#>      30      31      32      33      34      35      36      37      38      39 
#>  5386.6  5386.6  5386.6  5386.6  5386.6  4696.2  4696.2  4696.2  4696.2  4696.2 
#>      40      41      42      43      44      45      46      47      48      49 
#>  4120.4  4120.4  4120.4  4120.4  4120.4  3297.8  3297.8  3297.8  3297.8  3297.8 
#>      50      51      52      53      54      55      56      57      58      59 
#>  2849.6  2849.6  2849.6  2849.6  2849.6  1985.6  1985.6  1985.6  1985.6  1985.6 
#>      60      61      62      63      64      65      66      67      68      69 
#>  1698.0  1698.0  1698.0  1698.0  1698.0   960.2   960.2   960.2   960.2   960.2 
#>      70      71      72      73      74      75      76      77      78      79 
#>   719.8   719.8   719.8   719.8   719.8   409.6   409.6   409.6   409.6   409.6 
#>      80      81      82      83      84      85      86      87      88      89 
#>   188.2   188.2   188.2   188.2   188.2    65.2    65.2    65.2    65.2    65.2 
#>      90      91      92      93      94      95      96      97      98      99 
#>    16.0    16.0    16.0    16.0    16.0     3.4     3.4     3.4     3.4     3.4 
#>     100 
#>     0.0 

graduate(Value, Age, method = "pclm")

#>            0            1            2            3            4            5 
#> 10022.206166 10787.283969 11254.932558 11275.885637 10807.951423 10074.821802 
#>            6            7            8            9           10           11 
#>  9328.988744  8736.949692  8394.290484  8273.419215  8318.185349  8432.064234 
#>           12           13           14           15           16           17 
#>  8510.164397  8497.044472  8366.765819  8151.826916  7903.088275  7662.042021 
#>           18           19           20           21           22           23 
#>  7459.547647  7296.414764  7161.468632  7035.158814  6896.995478  6740.484441 
#>           24           25           26           27           28           29 
#>  6566.732478  6387.045674  6215.809170  6060.717420  5925.005956  5801.821192 
#>           30           31           32           33           34           35 
#>  5679.678101  5548.089015  5398.665318  5234.736857  5066.280112  4908.164621 
#>           36           37           38           39           40           41 
#>  4776.277248  4674.567054  4598.978778  4531.954680  4447.598715  4325.845342 
#>           42           43           44           45           46           47 
#>  4153.831514  3942.404243  3717.818156  3511.254515  3353.612943  3254.021670 
#>           48           49           50           51           52           53 
#>  3206.436723  3185.639206  3149.061720  3061.490648  2899.845840  2676.397236 
#>           54           55           56           57           58           59 
#>  2430.916307  2205.094665  2036.287307  1935.376747  1895.248589  1894.350582 
#>           60           61           62           63           64           65 
#>  1891.972238  1851.214007  1745.251330  1577.505630  1381.551282  1192.256352 
#>           66           67           68           69           70           71 
#>  1038.626281   929.281154   859.265602   818.951293   792.503364   766.385428 
#>           72           73           74           75           76           77 
#>   729.252757   676.036518   610.318839   538.450939   468.333489   404.619332 
#>           78           79           80           81           82           83 
#>   348.592064   299.924684   257.028970   218.668992   184.007104   152.690098 
#>           84           85           86           87           88           89 
#>   124.764601   100.309353    79.377465    61.866165    47.547799    36.084595 
#>           90           91           92           93           94           95 
#>    27.084347    20.136711    14.848708    10.871851     7.909705     5.722058 
#>           96           97           98           99          100 
#>     4.118689     2.951838     2.108558     1.502782     1.069736 

graduate(Value, Age, method = "pclm", keep0=TRUE)

#>            0            1            2            3            4            5 
#> 10000.000000 10798.027337 11266.141671 11287.115617 10818.715375 10074.821802 
#>            6            7            8            9           10           11 
#>  9328.988744  8736.949692  8394.290484  8273.419215  8318.185349  8432.064234 
#>           12           13           14           15           16           17 
#>  8510.164397  8497.044472  8366.765819  8151.826916  7903.088275  7662.042021 
#>           18           19           20           21           22           23 
#>  7459.547647  7296.414764  7161.468632  7035.158814  6896.995478  6740.484441 
#>           24           25           26           27           28           29 
#>  6566.732478  6387.045674  6215.809170  6060.717420  5925.005956  5801.821192 
#>           30           31           32           33           34           35 
#>  5679.678101  5548.089015  5398.665318  5234.736857  5066.280112  4908.164621 
#>           36           37           38           39           40           41 
#>  4776.277248  4674.567054  4598.978778  4531.954680  4447.598715  4325.845342 
#>           42           43           44           45           46           47 
#>  4153.831514  3942.404243  3717.818156  3511.254515  3353.612943  3254.021670 
#>           48           49           50           51           52           53 
#>  3206.436723  3185.639206  3149.061720  3061.490648  2899.845840  2676.397236 
#>           54           55           56           57           58           59 
#>  2430.916307  2205.094665  2036.287307  1935.376747  1895.248589  1894.350582 
#>           60           61           62           63           64           65 
#>  1891.972238  1851.214007  1745.251330  1577.505630  1381.551282  1192.256352 
#>           66           67           68           69           70           71 
#>  1038.626281   929.281154   859.265602   818.951293   792.503364   766.385428 
#>           72           73           74           75           76           77 
#>   729.252757   676.036518   610.318839   538.450939   468.333489   404.619332 
#>           78           79           80           81           82           83 
#>   348.592064   299.924684   257.028970   218.668992   184.007104   152.690098 
#>           84           85           86           87           88           89 
#>   124.764601   100.309353    79.377465    61.866165    47.547799    36.084595 
#>           90           91           92           93           94           95 
#>    27.084347    20.136711    14.848708    10.871851     7.909705     5.722058 
#>           96           97           98           99          100 
#>     4.118689     2.951838     2.108558     1.502782     1.069736 

# pclm can also graduate rates if both
# numerators and denominators are on hand:
Exposures <- c(100958,466275,624134,559559,446736,370653,301862,249409,
               247473,223014,172260,149338,127242,105715,79614,53660,
               31021,16805,8000,4000,2000,1000)

Deaths <- c(8674,1592,618,411,755,1098,1100,1357,
            1335,3257,2200,4023,2167,4578,2956,4212,
            2887,2351,1500,900,500,300)
Age    <- c(0, 1, seq(5, 100, by = 5))
AgeInt <- c(diff(Age), NA)

# exclude infants for better fit.
mx    <- graduate(
           Value = Deaths[-1], Age = Age[-1],
           AgeInt = AgeInt[-1], OAG = TRUE,
           OAnew = 110, offset = Exposures[-1],
           method = "pclm")
mx_sm <- graduate(
           Value = Deaths[-1], Age = Age[-1],
           AgeInt = AgeInt[-1], OAG = TRUE,
           OAnew = 110, offset = Exposures[-1],
           method = "pclm", control = list(lambda = 1e7))

if (FALSE) {
plot(Age,
     Deaths / Exposures,
     type = 's', log = 'y',
     main = "Underlying data have differential heaping on 0s and 5s")
lines(1:110, mx)
lines(1:110, mx_sm, col = "blue")
legend("bottomright",
       col = c("black","blue"),
       lty = c(1, 1),
       legend = c("lambda optimized (almost constrained)",
                  "higher lambda = smoother")
       )
  }