Adjust ages under 10 using a modification of Beers

Assuming we have an external estimate of age 0, this method refits to the ordinary Beers single age results, remaining constrained to the original 5-year age groups and smoothly blending into ages greater than 10.

graduate_beers_johnson(Age0, pop5, pop1)

Arguments

Age0
numeric. An estimate of age 0.

pop5
numeric. Matrix of age-period population counts in 5-year age groups.

pop1
numeric. Matrix of age-period population using Beers ordinary (or some other) method.

Value

A matrix of single age population estimates.

Details

This has not been tested using pop1 as generated from other methods, such as the Beers modified, Sprague, or Grabill methods. Called internally by beers().

References

Stover J, Johnson P, Zaba B, Zwahlen M, Dabis F, Ekpini R (2008). “The Spectrum projection package: improvements in estimating mortality, ART needs, PMTCT impact and uncertainty bounds.” Sexually Transmitted Infections, 84(Suppl 1), i24--i30.