Create the Grabill coefficient matrix.

The resulting coefficient matrix is based on the number of rows in popmat where we assume that each row of data is a 5-year age group and the final row is an open age group to be preserved as such.

graduate_grabill_expand(Value, Age, OAG = TRUE)

Arguments

Value
numeric vector, presumably counts in grouped ages

Age
integer vector, lower bounds of age groups

OAG
logical, default = TRUE is the final age group open?

Details

The Value vector is really just a placeholder in this case. This function is a utility called by the Grabill family of functions, where it is most convenient to just pass in the same matrix being used in those calculations to determine the layout of the coefficient matrix. Note that these coefficients do not constrain population counts to their year totals. This function is called by grabill(), which ensures matching marginals by 1) blending boundary ages into the Sprague estimated population, and 2) a second constraint on the middle age groups to enforce matching sums.

References

Shryock HS, Siegel JS, Larmon EA (1973). The methods and materials of demography. US Bureau of the Census.

Examples

a5 <- as.integer(rownames(pop5_mat))
graduate_grabill_expand(pop5_mat[,1], Age = a5, OAG = TRUE)

#>           [,1]    [,2]    [,3]    [,4]    [,5]    [,6]    [,7]    [,8]    [,9]
#>   [1,]  0.0826  0.0256 -0.0009  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>   [2,]  0.0903  0.0377 -0.0002  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>   [3,]  0.0932  0.0519  0.0015  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>   [4,]  0.0903  0.0673  0.0049  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>   [5,]  0.0826  0.0816  0.0111  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>   [6,]  0.0816  0.0826  0.0256 -0.0009  0.0000  0.0000  0.0000  0.0000  0.0000
#>   [7,]  0.0673  0.0903  0.0377 -0.0002  0.0000  0.0000  0.0000  0.0000  0.0000
#>   [8,]  0.0519  0.0932  0.0519  0.0015  0.0000  0.0000  0.0000  0.0000  0.0000
#>   [9,]  0.0377  0.0903  0.0673  0.0049  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [10,]  0.0256  0.0826  0.0816  0.0111  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [11,]  0.0111  0.0816  0.0826  0.0256 -0.0009  0.0000  0.0000  0.0000  0.0000
#>  [12,]  0.0049  0.0673  0.0903  0.0377 -0.0002  0.0000  0.0000  0.0000  0.0000
#>  [13,]  0.0015  0.0519  0.0932  0.0519  0.0015  0.0000  0.0000  0.0000  0.0000
#>  [14,] -0.0002  0.0377  0.0903  0.0673  0.0049  0.0000  0.0000  0.0000  0.0000
#>  [15,] -0.0009  0.0256  0.0826  0.0816  0.0111  0.0000  0.0000  0.0000  0.0000
#>  [16,]  0.0000  0.0111  0.0816  0.0826  0.0256 -0.0009  0.0000  0.0000  0.0000
#>  [17,]  0.0000  0.0049  0.0673  0.0903  0.0377 -0.0002  0.0000  0.0000  0.0000
#>  [18,]  0.0000  0.0015  0.0519  0.0932  0.0519  0.0015  0.0000  0.0000  0.0000
#>  [19,]  0.0000 -0.0002  0.0377  0.0903  0.0673  0.0049  0.0000  0.0000  0.0000
#>  [20,]  0.0000 -0.0009  0.0256  0.0826  0.0816  0.0111  0.0000  0.0000  0.0000
#>  [21,]  0.0000  0.0000  0.0111  0.0816  0.0826  0.0256 -0.0009  0.0000  0.0000
#>  [22,]  0.0000  0.0000  0.0049  0.0673  0.0903  0.0377 -0.0002  0.0000  0.0000
#>  [23,]  0.0000  0.0000  0.0015  0.0519  0.0932  0.0519  0.0015  0.0000  0.0000
#>  [24,]  0.0000  0.0000 -0.0002  0.0377  0.0903  0.0673  0.0049  0.0000  0.0000
#>  [25,]  0.0000  0.0000 -0.0009  0.0256  0.0826  0.0816  0.0111  0.0000  0.0000
#>  [26,]  0.0000  0.0000  0.0000  0.0111  0.0816  0.0826  0.0256 -0.0009  0.0000
#>  [27,]  0.0000  0.0000  0.0000  0.0049  0.0673  0.0903  0.0377 -0.0002  0.0000
#>  [28,]  0.0000  0.0000  0.0000  0.0015  0.0519  0.0932  0.0519  0.0015  0.0000
#>  [29,]  0.0000  0.0000  0.0000 -0.0002  0.0377  0.0903  0.0673  0.0049  0.0000
#>  [30,]  0.0000  0.0000  0.0000 -0.0009  0.0256  0.0826  0.0816  0.0111  0.0000
#>  [31,]  0.0000  0.0000  0.0000  0.0000  0.0111  0.0816  0.0826  0.0256 -0.0009
#>  [32,]  0.0000  0.0000  0.0000  0.0000  0.0049  0.0673  0.0903  0.0377 -0.0002
#>  [33,]  0.0000  0.0000  0.0000  0.0000  0.0015  0.0519  0.0932  0.0519  0.0015
#>  [34,]  0.0000  0.0000  0.0000  0.0000 -0.0002  0.0377  0.0903  0.0673  0.0049
#>  [35,]  0.0000  0.0000  0.0000  0.0000 -0.0009  0.0256  0.0826  0.0816  0.0111
#>  [36,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0111  0.0816  0.0826  0.0256
#>  [37,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0049  0.0673  0.0903  0.0377
#>  [38,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0015  0.0519  0.0932  0.0519
#>  [39,]  0.0000  0.0000  0.0000  0.0000  0.0000 -0.0002  0.0377  0.0903  0.0673
#>  [40,]  0.0000  0.0000  0.0000  0.0000  0.0000 -0.0009  0.0256  0.0826  0.0816
#>  [41,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0111  0.0816  0.0826
#>  [42,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0049  0.0673  0.0903
#>  [43,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0015  0.0519  0.0932
#>  [44,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000 -0.0002  0.0377  0.0903
#>  [45,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000 -0.0009  0.0256  0.0826
#>  [46,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0111  0.0816
#>  [47,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0049  0.0673
#>  [48,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0015  0.0519
#>  [49,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000 -0.0002  0.0377
#>  [50,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000 -0.0009  0.0256
#>  [51,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0111
#>  [52,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0049
#>  [53,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0015
#>  [54,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000 -0.0002
#>  [55,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000 -0.0009
#>  [56,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [57,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [58,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [59,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [60,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [61,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [62,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [63,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [64,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [65,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [66,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [67,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [68,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [69,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [70,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [71,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [72,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [73,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [74,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [75,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [76,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [77,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [78,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [79,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [80,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [81,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [82,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [83,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [84,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [85,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [86,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [87,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [88,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [89,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [90,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [91,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [92,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [93,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [94,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [95,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [96,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [97,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [98,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [99,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#> [100,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#> [101,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>          [,10]   [,11]   [,12]   [,13]   [,14]   [,15]   [,16]   [,17]   [,18]
#>   [1,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>   [2,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>   [3,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>   [4,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>   [5,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>   [6,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>   [7,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>   [8,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>   [9,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [10,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [11,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [12,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [13,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [14,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [15,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [16,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [17,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [18,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [19,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [20,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [21,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [22,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [23,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [24,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [25,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [26,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [27,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [28,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [29,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [30,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [31,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [32,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [33,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [34,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [35,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [36,] -0.0009  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [37,] -0.0002  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [38,]  0.0015  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [39,]  0.0049  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [40,]  0.0111  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [41,]  0.0256 -0.0009  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [42,]  0.0377 -0.0002  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [43,]  0.0519  0.0015  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [44,]  0.0673  0.0049  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [45,]  0.0816  0.0111  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [46,]  0.0826  0.0256 -0.0009  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [47,]  0.0903  0.0377 -0.0002  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [48,]  0.0932  0.0519  0.0015  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [49,]  0.0903  0.0673  0.0049  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [50,]  0.0826  0.0816  0.0111  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [51,]  0.0816  0.0826  0.0256 -0.0009  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [52,]  0.0673  0.0903  0.0377 -0.0002  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [53,]  0.0519  0.0932  0.0519  0.0015  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [54,]  0.0377  0.0903  0.0673  0.0049  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [55,]  0.0256  0.0826  0.0816  0.0111  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [56,]  0.0111  0.0816  0.0826  0.0256 -0.0009  0.0000  0.0000  0.0000  0.0000
#>  [57,]  0.0049  0.0673  0.0903  0.0377 -0.0002  0.0000  0.0000  0.0000  0.0000
#>  [58,]  0.0015  0.0519  0.0932  0.0519  0.0015  0.0000  0.0000  0.0000  0.0000
#>  [59,] -0.0002  0.0377  0.0903  0.0673  0.0049  0.0000  0.0000  0.0000  0.0000
#>  [60,] -0.0009  0.0256  0.0826  0.0816  0.0111  0.0000  0.0000  0.0000  0.0000
#>  [61,]  0.0000  0.0111  0.0816  0.0826  0.0256 -0.0009  0.0000  0.0000  0.0000
#>  [62,]  0.0000  0.0049  0.0673  0.0903  0.0377 -0.0002  0.0000  0.0000  0.0000
#>  [63,]  0.0000  0.0015  0.0519  0.0932  0.0519  0.0015  0.0000  0.0000  0.0000
#>  [64,]  0.0000 -0.0002  0.0377  0.0903  0.0673  0.0049  0.0000  0.0000  0.0000
#>  [65,]  0.0000 -0.0009  0.0256  0.0826  0.0816  0.0111  0.0000  0.0000  0.0000
#>  [66,]  0.0000  0.0000  0.0111  0.0816  0.0826  0.0256 -0.0009  0.0000  0.0000
#>  [67,]  0.0000  0.0000  0.0049  0.0673  0.0903  0.0377 -0.0002  0.0000  0.0000
#>  [68,]  0.0000  0.0000  0.0015  0.0519  0.0932  0.0519  0.0015  0.0000  0.0000
#>  [69,]  0.0000  0.0000 -0.0002  0.0377  0.0903  0.0673  0.0049  0.0000  0.0000
#>  [70,]  0.0000  0.0000 -0.0009  0.0256  0.0826  0.0816  0.0111  0.0000  0.0000
#>  [71,]  0.0000  0.0000  0.0000  0.0111  0.0816  0.0826  0.0256 -0.0009  0.0000
#>  [72,]  0.0000  0.0000  0.0000  0.0049  0.0673  0.0903  0.0377 -0.0002  0.0000
#>  [73,]  0.0000  0.0000  0.0000  0.0015  0.0519  0.0932  0.0519  0.0015  0.0000
#>  [74,]  0.0000  0.0000  0.0000 -0.0002  0.0377  0.0903  0.0673  0.0049  0.0000
#>  [75,]  0.0000  0.0000  0.0000 -0.0009  0.0256  0.0826  0.0816  0.0111  0.0000
#>  [76,]  0.0000  0.0000  0.0000  0.0000  0.0111  0.0816  0.0826  0.0256 -0.0009
#>  [77,]  0.0000  0.0000  0.0000  0.0000  0.0049  0.0673  0.0903  0.0377 -0.0002
#>  [78,]  0.0000  0.0000  0.0000  0.0000  0.0015  0.0519  0.0932  0.0519  0.0015
#>  [79,]  0.0000  0.0000  0.0000  0.0000 -0.0002  0.0377  0.0903  0.0673  0.0049
#>  [80,]  0.0000  0.0000  0.0000  0.0000 -0.0009  0.0256  0.0826  0.0816  0.0111
#>  [81,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0111  0.0816  0.0826  0.0256
#>  [82,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0049  0.0673  0.0903  0.0377
#>  [83,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0015  0.0519  0.0932  0.0519
#>  [84,]  0.0000  0.0000  0.0000  0.0000  0.0000 -0.0002  0.0377  0.0903  0.0673
#>  [85,]  0.0000  0.0000  0.0000  0.0000  0.0000 -0.0009  0.0256  0.0826  0.0816
#>  [86,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0111  0.0816  0.0826
#>  [87,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0049  0.0673  0.0903
#>  [88,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0015  0.0519  0.0932
#>  [89,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000 -0.0002  0.0377  0.0903
#>  [90,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000 -0.0009  0.0256  0.0826
#>  [91,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0111  0.0816
#>  [92,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0049  0.0673
#>  [93,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0015  0.0519
#>  [94,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000 -0.0002  0.0377
#>  [95,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000 -0.0009  0.0256
#>  [96,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0111
#>  [97,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0049
#>  [98,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0015
#>  [99,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000 -0.0002
#> [100,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000 -0.0009
#> [101,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>          [,19]   [,20] [,21]
#>   [1,]  0.0000  0.0000     0
#>   [2,]  0.0000  0.0000     0
#>   [3,]  0.0000  0.0000     0
#>   [4,]  0.0000  0.0000     0
#>   [5,]  0.0000  0.0000     0
#>   [6,]  0.0000  0.0000     0
#>   [7,]  0.0000  0.0000     0
#>   [8,]  0.0000  0.0000     0
#>   [9,]  0.0000  0.0000     0
#>  [10,]  0.0000  0.0000     0
#>  [11,]  0.0000  0.0000     0
#>  [12,]  0.0000  0.0000     0
#>  [13,]  0.0000  0.0000     0
#>  [14,]  0.0000  0.0000     0
#>  [15,]  0.0000  0.0000     0
#>  [16,]  0.0000  0.0000     0
#>  [17,]  0.0000  0.0000     0
#>  [18,]  0.0000  0.0000     0
#>  [19,]  0.0000  0.0000     0
#>  [20,]  0.0000  0.0000     0
#>  [21,]  0.0000  0.0000     0
#>  [22,]  0.0000  0.0000     0
#>  [23,]  0.0000  0.0000     0
#>  [24,]  0.0000  0.0000     0
#>  [25,]  0.0000  0.0000     0
#>  [26,]  0.0000  0.0000     0
#>  [27,]  0.0000  0.0000     0
#>  [28,]  0.0000  0.0000     0
#>  [29,]  0.0000  0.0000     0
#>  [30,]  0.0000  0.0000     0
#>  [31,]  0.0000  0.0000     0
#>  [32,]  0.0000  0.0000     0
#>  [33,]  0.0000  0.0000     0
#>  [34,]  0.0000  0.0000     0
#>  [35,]  0.0000  0.0000     0
#>  [36,]  0.0000  0.0000     0
#>  [37,]  0.0000  0.0000     0
#>  [38,]  0.0000  0.0000     0
#>  [39,]  0.0000  0.0000     0
#>  [40,]  0.0000  0.0000     0
#>  [41,]  0.0000  0.0000     0
#>  [42,]  0.0000  0.0000     0
#>  [43,]  0.0000  0.0000     0
#>  [44,]  0.0000  0.0000     0
#>  [45,]  0.0000  0.0000     0
#>  [46,]  0.0000  0.0000     0
#>  [47,]  0.0000  0.0000     0
#>  [48,]  0.0000  0.0000     0
#>  [49,]  0.0000  0.0000     0
#>  [50,]  0.0000  0.0000     0
#>  [51,]  0.0000  0.0000     0
#>  [52,]  0.0000  0.0000     0
#>  [53,]  0.0000  0.0000     0
#>  [54,]  0.0000  0.0000     0
#>  [55,]  0.0000  0.0000     0
#>  [56,]  0.0000  0.0000     0
#>  [57,]  0.0000  0.0000     0
#>  [58,]  0.0000  0.0000     0
#>  [59,]  0.0000  0.0000     0
#>  [60,]  0.0000  0.0000     0
#>  [61,]  0.0000  0.0000     0
#>  [62,]  0.0000  0.0000     0
#>  [63,]  0.0000  0.0000     0
#>  [64,]  0.0000  0.0000     0
#>  [65,]  0.0000  0.0000     0
#>  [66,]  0.0000  0.0000     0
#>  [67,]  0.0000  0.0000     0
#>  [68,]  0.0000  0.0000     0
#>  [69,]  0.0000  0.0000     0
#>  [70,]  0.0000  0.0000     0
#>  [71,]  0.0000  0.0000     0
#>  [72,]  0.0000  0.0000     0
#>  [73,]  0.0000  0.0000     0
#>  [74,]  0.0000  0.0000     0
#>  [75,]  0.0000  0.0000     0
#>  [76,]  0.0000  0.0000     0
#>  [77,]  0.0000  0.0000     0
#>  [78,]  0.0000  0.0000     0
#>  [79,]  0.0000  0.0000     0
#>  [80,]  0.0000  0.0000     0
#>  [81,] -0.0009  0.0000     0
#>  [82,] -0.0002  0.0000     0
#>  [83,]  0.0015  0.0000     0
#>  [84,]  0.0049  0.0000     0
#>  [85,]  0.0111  0.0000     0
#>  [86,]  0.0256 -0.0009     0
#>  [87,]  0.0377 -0.0002     0
#>  [88,]  0.0519  0.0015     0
#>  [89,]  0.0673  0.0049     0
#>  [90,]  0.0816  0.0111     0
#>  [91,]  0.0826  0.0256     0
#>  [92,]  0.0903  0.0377     0
#>  [93,]  0.0932  0.0519     0
#>  [94,]  0.0903  0.0673     0
#>  [95,]  0.0826  0.0816     0
#>  [96,]  0.0816  0.0826     0
#>  [97,]  0.0673  0.0903     0
#>  [98,]  0.0519  0.0932     0
#>  [99,]  0.0377  0.0903     0
#> [100,]  0.0256  0.0826     0
#> [101,]  0.0000  0.0000     1

graduate_grabill_expand(pop5_mat[,1], Age = a5, OAG = FALSE)

#>           [,1]    [,2]    [,3]    [,4]    [,5]    [,6]    [,7]    [,8]    [,9]
#>   [1,]  0.0826  0.0256 -0.0009  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>   [2,]  0.0903  0.0377 -0.0002  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>   [3,]  0.0932  0.0519  0.0015  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>   [4,]  0.0903  0.0673  0.0049  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>   [5,]  0.0826  0.0816  0.0111  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>   [6,]  0.0816  0.0826  0.0256 -0.0009  0.0000  0.0000  0.0000  0.0000  0.0000
#>   [7,]  0.0673  0.0903  0.0377 -0.0002  0.0000  0.0000  0.0000  0.0000  0.0000
#>   [8,]  0.0519  0.0932  0.0519  0.0015  0.0000  0.0000  0.0000  0.0000  0.0000
#>   [9,]  0.0377  0.0903  0.0673  0.0049  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [10,]  0.0256  0.0826  0.0816  0.0111  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [11,]  0.0111  0.0816  0.0826  0.0256 -0.0009  0.0000  0.0000  0.0000  0.0000
#>  [12,]  0.0049  0.0673  0.0903  0.0377 -0.0002  0.0000  0.0000  0.0000  0.0000
#>  [13,]  0.0015  0.0519  0.0932  0.0519  0.0015  0.0000  0.0000  0.0000  0.0000
#>  [14,] -0.0002  0.0377  0.0903  0.0673  0.0049  0.0000  0.0000  0.0000  0.0000
#>  [15,] -0.0009  0.0256  0.0826  0.0816  0.0111  0.0000  0.0000  0.0000  0.0000
#>  [16,]  0.0000  0.0111  0.0816  0.0826  0.0256 -0.0009  0.0000  0.0000  0.0000
#>  [17,]  0.0000  0.0049  0.0673  0.0903  0.0377 -0.0002  0.0000  0.0000  0.0000
#>  [18,]  0.0000  0.0015  0.0519  0.0932  0.0519  0.0015  0.0000  0.0000  0.0000
#>  [19,]  0.0000 -0.0002  0.0377  0.0903  0.0673  0.0049  0.0000  0.0000  0.0000
#>  [20,]  0.0000 -0.0009  0.0256  0.0826  0.0816  0.0111  0.0000  0.0000  0.0000
#>  [21,]  0.0000  0.0000  0.0111  0.0816  0.0826  0.0256 -0.0009  0.0000  0.0000
#>  [22,]  0.0000  0.0000  0.0049  0.0673  0.0903  0.0377 -0.0002  0.0000  0.0000
#>  [23,]  0.0000  0.0000  0.0015  0.0519  0.0932  0.0519  0.0015  0.0000  0.0000
#>  [24,]  0.0000  0.0000 -0.0002  0.0377  0.0903  0.0673  0.0049  0.0000  0.0000
#>  [25,]  0.0000  0.0000 -0.0009  0.0256  0.0826  0.0816  0.0111  0.0000  0.0000
#>  [26,]  0.0000  0.0000  0.0000  0.0111  0.0816  0.0826  0.0256 -0.0009  0.0000
#>  [27,]  0.0000  0.0000  0.0000  0.0049  0.0673  0.0903  0.0377 -0.0002  0.0000
#>  [28,]  0.0000  0.0000  0.0000  0.0015  0.0519  0.0932  0.0519  0.0015  0.0000
#>  [29,]  0.0000  0.0000  0.0000 -0.0002  0.0377  0.0903  0.0673  0.0049  0.0000
#>  [30,]  0.0000  0.0000  0.0000 -0.0009  0.0256  0.0826  0.0816  0.0111  0.0000
#>  [31,]  0.0000  0.0000  0.0000  0.0000  0.0111  0.0816  0.0826  0.0256 -0.0009
#>  [32,]  0.0000  0.0000  0.0000  0.0000  0.0049  0.0673  0.0903  0.0377 -0.0002
#>  [33,]  0.0000  0.0000  0.0000  0.0000  0.0015  0.0519  0.0932  0.0519  0.0015
#>  [34,]  0.0000  0.0000  0.0000  0.0000 -0.0002  0.0377  0.0903  0.0673  0.0049
#>  [35,]  0.0000  0.0000  0.0000  0.0000 -0.0009  0.0256  0.0826  0.0816  0.0111
#>  [36,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0111  0.0816  0.0826  0.0256
#>  [37,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0049  0.0673  0.0903  0.0377
#>  [38,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0015  0.0519  0.0932  0.0519
#>  [39,]  0.0000  0.0000  0.0000  0.0000  0.0000 -0.0002  0.0377  0.0903  0.0673
#>  [40,]  0.0000  0.0000  0.0000  0.0000  0.0000 -0.0009  0.0256  0.0826  0.0816
#>  [41,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0111  0.0816  0.0826
#>  [42,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0049  0.0673  0.0903
#>  [43,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0015  0.0519  0.0932
#>  [44,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000 -0.0002  0.0377  0.0903
#>  [45,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000 -0.0009  0.0256  0.0826
#>  [46,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0111  0.0816
#>  [47,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0049  0.0673
#>  [48,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0015  0.0519
#>  [49,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000 -0.0002  0.0377
#>  [50,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000 -0.0009  0.0256
#>  [51,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0111
#>  [52,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0049
#>  [53,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0015
#>  [54,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000 -0.0002
#>  [55,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000 -0.0009
#>  [56,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [57,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [58,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [59,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [60,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [61,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [62,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [63,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [64,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [65,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [66,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [67,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [68,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [69,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [70,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [71,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [72,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [73,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [74,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [75,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [76,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [77,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [78,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [79,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [80,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [81,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [82,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [83,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [84,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [85,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [86,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [87,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [88,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [89,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [90,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [91,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [92,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [93,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [94,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [95,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [96,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [97,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [98,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [99,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#> [100,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#> [101,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#> [102,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#> [103,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#> [104,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#> [105,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>          [,10]   [,11]   [,12]   [,13]   [,14]   [,15]   [,16]   [,17]   [,18]
#>   [1,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>   [2,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>   [3,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>   [4,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>   [5,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>   [6,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>   [7,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>   [8,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>   [9,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [10,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [11,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [12,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [13,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [14,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [15,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [16,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [17,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [18,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [19,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [20,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [21,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [22,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [23,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [24,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [25,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [26,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [27,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [28,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [29,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [30,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [31,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [32,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [33,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [34,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [35,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [36,] -0.0009  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [37,] -0.0002  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [38,]  0.0015  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [39,]  0.0049  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [40,]  0.0111  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [41,]  0.0256 -0.0009  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [42,]  0.0377 -0.0002  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [43,]  0.0519  0.0015  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [44,]  0.0673  0.0049  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [45,]  0.0816  0.0111  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [46,]  0.0826  0.0256 -0.0009  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [47,]  0.0903  0.0377 -0.0002  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [48,]  0.0932  0.0519  0.0015  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [49,]  0.0903  0.0673  0.0049  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [50,]  0.0826  0.0816  0.0111  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [51,]  0.0816  0.0826  0.0256 -0.0009  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [52,]  0.0673  0.0903  0.0377 -0.0002  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [53,]  0.0519  0.0932  0.0519  0.0015  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [54,]  0.0377  0.0903  0.0673  0.0049  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [55,]  0.0256  0.0826  0.0816  0.0111  0.0000  0.0000  0.0000  0.0000  0.0000
#>  [56,]  0.0111  0.0816  0.0826  0.0256 -0.0009  0.0000  0.0000  0.0000  0.0000
#>  [57,]  0.0049  0.0673  0.0903  0.0377 -0.0002  0.0000  0.0000  0.0000  0.0000
#>  [58,]  0.0015  0.0519  0.0932  0.0519  0.0015  0.0000  0.0000  0.0000  0.0000
#>  [59,] -0.0002  0.0377  0.0903  0.0673  0.0049  0.0000  0.0000  0.0000  0.0000
#>  [60,] -0.0009  0.0256  0.0826  0.0816  0.0111  0.0000  0.0000  0.0000  0.0000
#>  [61,]  0.0000  0.0111  0.0816  0.0826  0.0256 -0.0009  0.0000  0.0000  0.0000
#>  [62,]  0.0000  0.0049  0.0673  0.0903  0.0377 -0.0002  0.0000  0.0000  0.0000
#>  [63,]  0.0000  0.0015  0.0519  0.0932  0.0519  0.0015  0.0000  0.0000  0.0000
#>  [64,]  0.0000 -0.0002  0.0377  0.0903  0.0673  0.0049  0.0000  0.0000  0.0000
#>  [65,]  0.0000 -0.0009  0.0256  0.0826  0.0816  0.0111  0.0000  0.0000  0.0000
#>  [66,]  0.0000  0.0000  0.0111  0.0816  0.0826  0.0256 -0.0009  0.0000  0.0000
#>  [67,]  0.0000  0.0000  0.0049  0.0673  0.0903  0.0377 -0.0002  0.0000  0.0000
#>  [68,]  0.0000  0.0000  0.0015  0.0519  0.0932  0.0519  0.0015  0.0000  0.0000
#>  [69,]  0.0000  0.0000 -0.0002  0.0377  0.0903  0.0673  0.0049  0.0000  0.0000
#>  [70,]  0.0000  0.0000 -0.0009  0.0256  0.0826  0.0816  0.0111  0.0000  0.0000
#>  [71,]  0.0000  0.0000  0.0000  0.0111  0.0816  0.0826  0.0256 -0.0009  0.0000
#>  [72,]  0.0000  0.0000  0.0000  0.0049  0.0673  0.0903  0.0377 -0.0002  0.0000
#>  [73,]  0.0000  0.0000  0.0000  0.0015  0.0519  0.0932  0.0519  0.0015  0.0000
#>  [74,]  0.0000  0.0000  0.0000 -0.0002  0.0377  0.0903  0.0673  0.0049  0.0000
#>  [75,]  0.0000  0.0000  0.0000 -0.0009  0.0256  0.0826  0.0816  0.0111  0.0000
#>  [76,]  0.0000  0.0000  0.0000  0.0000  0.0111  0.0816  0.0826  0.0256 -0.0009
#>  [77,]  0.0000  0.0000  0.0000  0.0000  0.0049  0.0673  0.0903  0.0377 -0.0002
#>  [78,]  0.0000  0.0000  0.0000  0.0000  0.0015  0.0519  0.0932  0.0519  0.0015
#>  [79,]  0.0000  0.0000  0.0000  0.0000 -0.0002  0.0377  0.0903  0.0673  0.0049
#>  [80,]  0.0000  0.0000  0.0000  0.0000 -0.0009  0.0256  0.0826  0.0816  0.0111
#>  [81,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0111  0.0816  0.0826  0.0256
#>  [82,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0049  0.0673  0.0903  0.0377
#>  [83,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0015  0.0519  0.0932  0.0519
#>  [84,]  0.0000  0.0000  0.0000  0.0000  0.0000 -0.0002  0.0377  0.0903  0.0673
#>  [85,]  0.0000  0.0000  0.0000  0.0000  0.0000 -0.0009  0.0256  0.0826  0.0816
#>  [86,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0111  0.0816  0.0826
#>  [87,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0049  0.0673  0.0903
#>  [88,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0015  0.0519  0.0932
#>  [89,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000 -0.0002  0.0377  0.0903
#>  [90,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000 -0.0009  0.0256  0.0826
#>  [91,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0111  0.0816
#>  [92,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0049  0.0673
#>  [93,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0015  0.0519
#>  [94,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000 -0.0002  0.0377
#>  [95,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000 -0.0009  0.0256
#>  [96,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0111
#>  [97,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0049
#>  [98,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0015
#>  [99,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000 -0.0002
#> [100,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000 -0.0009
#> [101,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#> [102,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#> [103,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#> [104,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#> [105,]  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000  0.0000
#>          [,19]   [,20]   [,21]
#>   [1,]  0.0000  0.0000  0.0000
#>   [2,]  0.0000  0.0000  0.0000
#>   [3,]  0.0000  0.0000  0.0000
#>   [4,]  0.0000  0.0000  0.0000
#>   [5,]  0.0000  0.0000  0.0000
#>   [6,]  0.0000  0.0000  0.0000
#>   [7,]  0.0000  0.0000  0.0000
#>   [8,]  0.0000  0.0000  0.0000
#>   [9,]  0.0000  0.0000  0.0000
#>  [10,]  0.0000  0.0000  0.0000
#>  [11,]  0.0000  0.0000  0.0000
#>  [12,]  0.0000  0.0000  0.0000
#>  [13,]  0.0000  0.0000  0.0000
#>  [14,]  0.0000  0.0000  0.0000
#>  [15,]  0.0000  0.0000  0.0000
#>  [16,]  0.0000  0.0000  0.0000
#>  [17,]  0.0000  0.0000  0.0000
#>  [18,]  0.0000  0.0000  0.0000
#>  [19,]  0.0000  0.0000  0.0000
#>  [20,]  0.0000  0.0000  0.0000
#>  [21,]  0.0000  0.0000  0.0000
#>  [22,]  0.0000  0.0000  0.0000
#>  [23,]  0.0000  0.0000  0.0000
#>  [24,]  0.0000  0.0000  0.0000
#>  [25,]  0.0000  0.0000  0.0000
#>  [26,]  0.0000  0.0000  0.0000
#>  [27,]  0.0000  0.0000  0.0000
#>  [28,]  0.0000  0.0000  0.0000
#>  [29,]  0.0000  0.0000  0.0000
#>  [30,]  0.0000  0.0000  0.0000
#>  [31,]  0.0000  0.0000  0.0000
#>  [32,]  0.0000  0.0000  0.0000
#>  [33,]  0.0000  0.0000  0.0000
#>  [34,]  0.0000  0.0000  0.0000
#>  [35,]  0.0000  0.0000  0.0000
#>  [36,]  0.0000  0.0000  0.0000
#>  [37,]  0.0000  0.0000  0.0000
#>  [38,]  0.0000  0.0000  0.0000
#>  [39,]  0.0000  0.0000  0.0000
#>  [40,]  0.0000  0.0000  0.0000
#>  [41,]  0.0000  0.0000  0.0000
#>  [42,]  0.0000  0.0000  0.0000
#>  [43,]  0.0000  0.0000  0.0000
#>  [44,]  0.0000  0.0000  0.0000
#>  [45,]  0.0000  0.0000  0.0000
#>  [46,]  0.0000  0.0000  0.0000
#>  [47,]  0.0000  0.0000  0.0000
#>  [48,]  0.0000  0.0000  0.0000
#>  [49,]  0.0000  0.0000  0.0000
#>  [50,]  0.0000  0.0000  0.0000
#>  [51,]  0.0000  0.0000  0.0000
#>  [52,]  0.0000  0.0000  0.0000
#>  [53,]  0.0000  0.0000  0.0000
#>  [54,]  0.0000  0.0000  0.0000
#>  [55,]  0.0000  0.0000  0.0000
#>  [56,]  0.0000  0.0000  0.0000
#>  [57,]  0.0000  0.0000  0.0000
#>  [58,]  0.0000  0.0000  0.0000
#>  [59,]  0.0000  0.0000  0.0000
#>  [60,]  0.0000  0.0000  0.0000
#>  [61,]  0.0000  0.0000  0.0000
#>  [62,]  0.0000  0.0000  0.0000
#>  [63,]  0.0000  0.0000  0.0000
#>  [64,]  0.0000  0.0000  0.0000
#>  [65,]  0.0000  0.0000  0.0000
#>  [66,]  0.0000  0.0000  0.0000
#>  [67,]  0.0000  0.0000  0.0000
#>  [68,]  0.0000  0.0000  0.0000
#>  [69,]  0.0000  0.0000  0.0000
#>  [70,]  0.0000  0.0000  0.0000
#>  [71,]  0.0000  0.0000  0.0000
#>  [72,]  0.0000  0.0000  0.0000
#>  [73,]  0.0000  0.0000  0.0000
#>  [74,]  0.0000  0.0000  0.0000
#>  [75,]  0.0000  0.0000  0.0000
#>  [76,]  0.0000  0.0000  0.0000
#>  [77,]  0.0000  0.0000  0.0000
#>  [78,]  0.0000  0.0000  0.0000
#>  [79,]  0.0000  0.0000  0.0000
#>  [80,]  0.0000  0.0000  0.0000
#>  [81,] -0.0009  0.0000  0.0000
#>  [82,] -0.0002  0.0000  0.0000
#>  [83,]  0.0015  0.0000  0.0000
#>  [84,]  0.0049  0.0000  0.0000
#>  [85,]  0.0111  0.0000  0.0000
#>  [86,]  0.0256 -0.0009  0.0000
#>  [87,]  0.0377 -0.0002  0.0000
#>  [88,]  0.0519  0.0015  0.0000
#>  [89,]  0.0673  0.0049  0.0000
#>  [90,]  0.0816  0.0111  0.0000
#>  [91,]  0.0826  0.0256 -0.0009
#>  [92,]  0.0903  0.0377 -0.0002
#>  [93,]  0.0932  0.0519  0.0015
#>  [94,]  0.0903  0.0673  0.0049
#>  [95,]  0.0826  0.0816  0.0111
#>  [96,]  0.0816  0.0826  0.0256
#>  [97,]  0.0673  0.0903  0.0377
#>  [98,]  0.0519  0.0932  0.0519
#>  [99,]  0.0377  0.0903  0.0673
#> [100,]  0.0256  0.0826  0.0816
#> [101,]  0.0111  0.0816  0.0826
#> [102,]  0.0049  0.0673  0.0903
#> [103,]  0.0015  0.0519  0.0932
#> [104,] -0.0002  0.0377  0.0903
#> [105,] -0.0009  0.0256  0.0826