Estimate LC with limited data params

Estimate LC with limited data from a matrix of rates (age by dates).

interp_lc_lim_estimate(M, dates_in, dates_out, SVD = F)

Arguments

M
numeric. Matrix with many rows as ages and columns as dates_in.

dates_in
numeric. Vector of dates with input rates.

dates_out
numeric. Vector of dates for estimate a set of rates.

SVD
logical. Use Singular Value Decomposition for estimate b and k or Maximum Likelihood Estimation. Default FALSE for Maximum Likelihood Estimation.

Details

SVD for ax and bx. Fit a simmple linear model for k and interp/extrapolate for objective dates.

References

Li N, Lee R, Tuljapurkar S (2004). “Using the Lee-Carter Method to Forecast Mortality for Populations with Limited Data\(\ast\).” Int. Stat. Rev., 72(1), 19--36. ISSN 0306-7734, doi: 10.1111/j.1751-5823.2004.tb00221.x .