Life expectancy in the open age group.

Get an estimate of life expectancy in the open age group.

lt_a_closeout(
  mx,
  Age,
  extrapLaw = "kannisto",
  extrapFrom = max(Age),
  extrapFit = Age[Age >= 40],
  ...
)

Arguments

mx
Vector or matrix of age specific death-rates.

Age
integer. A vector of ages of the lower integer bound of the age classes.

extrapLaw
character. The following options are available:

"kannisto" -- The Kannisto model;

"kannisto_makeham" -- The Kannisto-Makeham model;

"gompertz" -- The Gompertz model;

"ggompertz" -- The Gamma-Gompertz model;

"makeham" -- The Makeham model;

"beard" -- The Beard model;

"beard_makeham" -- The Beard-Makeham model;

"quadratic" -- The Quadratic model.

extrapFrom
integer. Age from which to impute extrapolated mortality.

extrapFit
integer vector. Ages to include in model fitting. Defaults to all ages > =60.

...
Other arguments to be passed on to the MortalityLaw function.

Value

life expectancy in the open age group

Details

This method estimates life expectancy in the open age group by fitting one of several potential old-age parametric mortality models, extrapolating rates to age 130, then backing out the implied remaining life expectancy in the open age group. This function replaces aomegaMORTPAK().

See also

Examples

nMx <- c(0.12846,0.02477,0.00603,0.0034,
     0.00417,0.00513,0.00581,0.00645,0.00725,
     0.00813,0.00913,0.01199,0.01647,
     0.0256,0.04047,0.06624,0.10638,0.19611)
Age <- c(0,1,seq(5,80,by =5))

lt_a_closeout(nMx,Age,"kannisto")
#> [1] 5.039066
lt_a_closeout(nMx,Age,"kannisto_makeham")
#> [1] 3.909895
lt_a_closeout(nMx,Age,"makeham")
#> [1] 3.641941
lt_a_closeout(nMx,Age,"gompertz")
#> [1] 4.64754
lt_a_closeout(nMx,Age,"ggompertz")
#> [1] 4.649295
lt_a_closeout(nMx,Age,extrapLaw ="beard")
#> [1] 4.649174
lt_a_closeout(nMx,Age,"beard_makeham")
#> [1] 3.640968
lt_a_closeout(nMx,Age,"quadratic")
#> [1] 6.913734