wrapper to invoke PAS or UN ax methods given qx or mx

Given either mx or qx, call either the lt_a_un() or lt_a_pas() functions.

lt_id_morq_a(
  nMx,
  nqx,
  axmethod = "pas",
  Age,
  AgeInt,
  a0rule = "ak",
  IMR = NA,
  Sex = "m",
  region,
  OAG = TRUE,
  mod = TRUE,
  SRB = SRB,
  extrapLaw = "kannisto",
  extrapFrom = max(Age),
  extrapFit = Age[Age >= 60],
  ...
)

Arguments

nMx
numeric. Event exposure mortality rates.

nqx
numeric. Vector of age specific death probabilities in standard abridged age groups.

axmethod
character. Either "pas" or "un"

Age
integer. A vector of ages of the lower integer bound of the age classes.

AgeInt
integer. Vector of age interval widths.

a0rule
character. Either "ak" (default) or "cd".

IMR
numeric. Optional. q0, the death probability in first year of life, in case available separately.

Sex
character. "m", "f" or "b" for male, female, or both.

region
character. "n", "e", "s" or "w" for North, East, South, or West.

OAG
logical. Whether or not the last element of nMx is the open age group Default TRUE.

mod
logical. Whether or not to use Gerland's modification for ages 5-14. Default TRUE.

SRB
numeric. The sex ratio at birth (boys/girls), default 1.05.

extrapLaw
character. If extrapolating, which parametric mortality law should be invoked? Options include "Kannisto", "Kannisto_Makeham", "Makeham","Gompertz", "GGompertz", "Beard", "Beard_Makeham", "Quadratic". Default "Kannisto". See details.

extrapFrom
integer. Age from which to impute extrapolated mortality.

extrapFit
integer vector. Ages to include in model fitting. Defaults to all ages > =60.

...
Other arguments to be passed on to the MortalityLaw function.

Value

nax average contribution to exposure of those dying in the interval.

References

Greville TN (1977). “Short methods of constructing abridged life tables.” In Mathematical Demography, 53--60. Springer. United Nations (1982). Model Life Tables for Developing Countries. United Nations Department of International Economic and Social Affairs. Arriaga EE, Johnson PD, Jamison E (1994). Population analysis with microcomputers, volume 1. Bureau of the Census. United Nations (1988). Mortpak- The United Nations Software Package for Mortality Measurement. United Nations Department of International Economic and Social Affairs. United Nations (1983). Manual X: Indirect Techniques for Demographic Estimation, number 81. United Nations Department of International Economic and Social Affairs, New York. United States Census Bureau (2017). “Population Analysis System (PAS) Software.” https://www.census.gov/data/software/pas.html, https://www.census.gov/data/software/pas.html.