#' Derive nqx from nMx and nax. #' @description This is the standard identity to derive nqx from nax and nMx. #' @details qx values calculated as greater than 1 are imputed with 1. #' @param nMx numeric. Vector of age-specific death rates. #' @param nax numeric. Vector of average time spent in interval by those dying in interval. #' @param AgeInt integer. Vector of age class widths. #' @references #' Preston S, Heuveline P, Guillot M (2000). Demography: measuring and modeling population processes. Wiley-Blackwell. #' @return nqx vector of age specific death probabilities derived via identity. #' @export lt_id_m_q <- function(nMx, nax, AgeInt = inferAgeIntAbr(vec = nMx)) if (as.character(match.call()[[1]]) == "mx2qx") warning("please use lt_id_m_q() instead of mx2qx().", call. = FALSE) qx <- (AgeInt * nMx) / (1 + (AgeInt - nax) * nMx) ind <- qx > 1 | is.na(qx) if (sum(ind) > 0) #cat("at least 1 q(x) > 1, imputed as 1") qx[ind] <- 1 qx #' @export #' @rdname lt_id_ma_q mx2qx <- lt_id_ma_q Derive nax from nqx and nMx.

This is the standard identity to derive nax from nqx and nMx.

lt_id_qm_a(nqx, nMx, AgeInt)

qxmx2ax(nqx, nMx, AgeInt)

Arguments

nqx
numeric. Vector of age specific death probabilities.

nMx
numeric. Vector of age-specific death rates.

AgeInt
integer. Vector of age class widths.

Value

nax numeric vector of average time spent in interval by those dying in interval via identity.

References

Preston S, Heuveline P, Guillot M (2000). Demography: measuring and modeling population processes. Wiley-Blackwell.