Coale-Demeny a(0) as function of m(0), region, and sex.

Coale-Demeny a(0) from Manual X Table 164. This is a rule of thumb. In this and some other older texts, a(0) is known as a 'separation factor'.

lt_rule_1a0_cd(M0, IMR = NA, Sex = "m", region = "w")

Arguments

M0
numeric. Event exposure infant mortality rate.

IMR
numeric. Optional. q0, the death probability in first year of life, in case available separately.

Sex
character. "m" or "f" for male or female,

region
character. "n", "e", "s" or "w" for North, East, South, or West.

Value

The average age at death in the first year of life a(0).

Details

If IMR is not given, then M0 is converted to q(0) using the following approximation:

Find \(\alpha , \beta\). Look up the appropriate slope and intercept for the given sex and region.

calculate \(a\) as: a = M0 * β

calculate \(b\) as: b = 1 + M0 *(1- α)

approximate q0 as: q0 = (b2- √ [b -4*a*M0]) / (2*a)

use q0 as IMR, and applied directly to the Coale-Demeny piecewise linear formula.

If IMR is given, then M0 is disregarded, and transitivity is therefore not guaranteed. In this case, one has the option to use lt_id_qm_a() to derive a(0), however discrepancies between these two parameters could force implausible results in a(0), whereas the CD rule always gives something plausible.

References

United Nations (1983). Manual X: Indirect Techniques for Demographic Estimation, number 81. United Nations Department of International Economic and Social Affairs, New York. United States Census Bureau (2017). “Population Analysis System (PAS) Software.” https://www.census.gov/data/software/pas.html, https://www.census.gov/data/software/pas.html.

Examples

m0 <- seq(.001, .2, by = .001)
if (FALSE) {
plot(m0, sapply(m0, lt_rule_1a0_cd, Sex = "m", region = "e"), ylab = "a0",
    type = 'l', ylim = c(0,.36), lty = 2, col = "blue")
lines(m0,sapply(m0, lt_rule_1a0_cd, Sex = "m", region = "w"), col = "blue")
lines(m0,sapply(m0, lt_rule_1a0_cd, Sex = "f", region = "e"), lty = 2, col = "red")
lines(m0,sapply(m0, lt_rule_1a0_cd, Sex = "f", region = "w"), col = "red")
text(.15, lt_rule_1a0_cd(.15,Sex = "m", region = "e"),"males E",font=2)
text(.15, lt_rule_1a0_cd(.15,Sex = "m", region = "w"),"males N,W,S",font=2)
text(.15, lt_rule_1a0_cd(.15,Sex = "f", region = "e"),"females E",font=2)
text(.15, lt_rule_1a0_cd(.15,Sex = "f", region = "w"),"females N,W,S",font=2)

# compare with the Preston approximation
# constants identical after m0 = .107
m0 <- seq(.001,.107,by =.001)
a0CDm0 <- sapply(m0, lt_rule_1a0_cd, Sex = "m", region = "w")
a0CDpr <- 0.045 + 2.684 * m0
plot(m0, a0CDm0, type = 'l', lty = 2, col = "red")
lines(m0, a0CDpr)
plot(m0, (a0CDm0 - a0CDpr) * 365, main = "difference (days)", ylab = "days")
}