calculate a single age lifetable from qx

Computes lifetable columns from single age qx by first computing 1ax, then computing 1mx from 1qx and 1ax, and finally passing the 1mx to the lt_single_mx() function

lt_single_qx(
  nqx,
  Age = 1:length(nqx) - 1,
  radix = 1e+05,
  a0rule = "ak",
  Sex = "m",
  region = "w",
  IMR = NA,
  mod = TRUE,
  SRB = 1.05,
  OAG = TRUE,
  OAnew = max(Age),
  extrapLaw = NULL,
  extrapFrom = max(Age),
  extrapFit = NULL,
  ...
)

Arguments

nqx
numeric. Vector of conditional death probabilities in abridged age classes.

Age
integer. A vector of ages of the lower integer bound of the age classes.

radix
numeric. Lifetable radix, l0. Default 100000.

a0rule
character. Either "ak" (default) or "cd".

Sex
character. Either male "m", female "f", or both "b".

region
character. North, East, South, or West: code"n", code"e", code"s", code"w". Default code"w".

IMR
numeric. Infant mortality rate q0, in case available and nqx is not specified. Default NA.

mod
logical. If "un" specified for axmethod, whether or not to use Nan Li's modification for ages 5-14. Default TRUE.

SRB
the sex ratio at birth (boys / girls), default 1.05

OAG
logical. Whether or not the last element of nMx (or nqx or lx) is an open age group. Default TRUE.

OAnew
integer. Desired open age group (5-year ages only). Default max(Age). If higher then rates are extrapolated.

extrapLaw
character. If extrapolating, which parametric mortality law should be invoked? Options include "Kannisto", "Kannisto_Makeham", "Makeham", "Gompertz", "GGompertz", "Beard", "Beard_Makeham", "Quadratic". Default "Kannisto" if the highest age is at least 90, otherwise "makeham". See details.

extrapFrom
integer. Age from which to impute extrapolated mortality.

extrapFit
integer vector. Ages to include in model fitting. Defaults to all ages > =60.

...
Other arguments to be passed on to the MortalityLaw function.

Value

Lifetable in data.frame with columns

Ageinteger. Lower bound of abridged age class,

AgeIntinteger. Age class widths.

nMxnumeric. Age-specific central death rates.

nAxnumeric. Average time spent in interval by those deceased in interval.

nqxnumeric. Age-specific conditional death probabilities.

lxnumeric. Lifetable survivorship

ndxnumeric. Lifetable deaths distribution.

nLxnumeric. Lifetable exposure.

Sxnumeric. Survivor ratios in uniform 5-year age groups.

Txnumeric. Lifetable total years left to live above age x.

exnumeric. Age-specific remaining life expectancy.

Details

Similar to lt_abridged() details, forthcoming