Estimate net migration using residual methods: stock change, time even flow and cohort even flow

Estimate net migration using residual methods: stock change, time even flow and cohort even flow

mig_resid(
  pop_m_mat,
  pop_f_mat,
  sr_m_mat,
  sr_f_mat,
  asfr_mat,
  srb_vec,
  ages = NULL,
  ages_asfr = NULL,
  years_pop = NULL,
  years_sr = NULL,
  years_asfr = NULL,
  years_srb = NULL,
  verbose = TRUE,
  method = c("stock", "cohort", "time")
)

mig_resid_stock(
  pop_m_mat,
  pop_f_mat,
  sr_m_mat,
  sr_f_mat,
  asfr_mat,
  srb_vec,
  ages = NULL,
  ages_asfr = NULL,
  years_pop = NULL,
  years_sr = NULL,
  years_asfr = NULL,
  years_srb = NULL,
  verbose = TRUE
)

mig_resid_cohort(
  pop_m_mat,
  pop_f_mat,
  sr_m_mat,
  sr_f_mat,
  asfr_mat,
  srb_vec,
  ages = NULL,
  ages_asfr = NULL,
  years_pop = NULL,
  years_sr = NULL,
  years_asfr = NULL,
  years_srb = NULL,
  verbose = TRUE
)

mig_resid_time(
  pop_m_mat,
  pop_f_mat,
  sr_m_mat,
  sr_f_mat,
  asfr_mat,
  srb_vec,
  ages = NULL,
  ages_asfr = NULL,
  years_pop = NULL,
  years_sr = NULL,
  years_asfr = NULL,
  years_srb = NULL,
  verbose = TRUE
)

Arguments

pop_m_mat
A numeric matrix with population counts. Rows should be ages and columns should be years. Only five year age groups are supported. See examples.

pop_f_mat
A numeric matrix with population counts. Rows should be ages and columns should be years. Only five year age groups are supported. See examples.

sr_m_mat
A numeric matrix with survival rates for males. Rows should be ages and columns should be years. ** This matrix should have one column less than pop_m_mat and pop_f_mat. For example, if the last year in these matrices is 2050, then the last year in sr_m_mat should be 2045. **

sr_f_mat
A numeric matrix with survival rates for females. Rows should be ages and columns should be years. ** This matrix should have one column less than pop_m_mat and pop_f_mat. For example, if the last year in these matrices is 2050, then the last year in sr_f_mat should be 2045. **.

asfr_mat
A numeric matrix with age specific fertility rates. Rows should be ages and columns should be years. ** This matrix should have one column less than pop_m_mat and pop_f_mat. For example, if the last year in these matrices is 2050, then the last year in asfr_mat should be 2045**. This row will usually have fewer age groups (rows) than in the population matrices or survival matrices, so the user needs to supply the specific ages in the ages_asfr argument.

srb_vec
A numeric vector of sex ratios at birth for every year. The years should be the same as the years in sr_m_mat, sr_f_mat, and asfr_mat.

ages
A numeric vector of ages used in the rows in pop_m_mat, pop_f_mat, sr_m_mat, sr_f_mat.

ages_asfr
A numeric vector of ages used in the rows in asfr_mat.

years_pop
Years used in the column names of population. If pop_m_mat or pop_f_mat doesn't have column names, these names are used. Otherwise ignored.

years_sr
Years used in the column names of survival rates. If sr_r_mat doesn't have column names, these names are used. Otherwise ignored.

years_asfr
Years used in the column names of age-specific fertility rate. If codeasfr_r_mat doesn't have column names, these names are used. Otherwise ignored.

years_srb
Years used in the column names of sex-ratio at birth. If srb_r_mat is not named, these names are used. Otherwise ignored.

verbose
Default set to TRUE. If TRUE, the function prints important operations carried out in the function like if years were trimmed from the data.

method
which residual migration method to use. This only works when using mig_resid and the possible options are 'stock', 'cohort' and 'time', with 'stock' being the default.

Value

A list with two matrices. One is for males (called mig_m) and the other for females (called mig_f). Both matrices contain net migration estimates by age/period using one of the three methods.

Details

The stock method (mig_resid_stock) is the difference in stocks that survive between t and t+5, and the first age group is based on the difference with the surviving births by sex. It provides net migrants by lexis cohort parallelograms, and basically such info gets used as end-period migration since the migrants don't get exposed to mortality within the period.

The time even flow (mig_resid_time) method uses the result from the first option, but splits it back into lexis period squares and assumes that half of the net migrants get exposed to the mortality risk during this period. Such info can get used as evenly distributed migration by period, but the assumptions lead to zig-zag age patterns that are highly implausible.

The cohort even flow (mig_resid_cohort) method provides the most meaningful pattern of net migration by age consistent by cohort and assumes an evenly distribution within the 5-year period, and half of the migrants get exposed both fertility and mortality within this period.

mig_resid is a general function able to call the three methods only by specifying the method argument. By default it is set to the stock method. See the examples section.

Examples

library(DemoTools)

# The data is loaded with DemoTools

################ Stock change method #####################

# Generic mig_resid method which allows to choose either stock,
# cohort or time method for five year ages groups

mig_res <-
 mig_resid(
   pop_m_mat = pop_m_mat_five,
   pop_f_mat = pop_f_mat_five,
   sr_m_mat = sr_m_mat_five,
   sr_f_mat = sr_f_mat_five,
   asfr_mat = asfr_mat_five,
   srb_vec = srb_vec_five,
   ages = ages_five,
   ages_asfr = ages_asfr_five,
   # With the stock method
   method = "stock"
 )

#> Using stock residual migration method

# For single ages

mig_res <-
 mig_resid(
   pop_m_mat = pop_m_mat_single,
   pop_f_mat = pop_f_mat_single,
   sr_m_mat = sr_m_mat_single,
   sr_f_mat = sr_f_mat_single,
   asfr_mat = asfr_mat_single,
   srb_vec = srb_vec_single,
   ages = ages_single,
   ages_asfr = ages_asfr_single,
   # With the stock method
   method = "stock"
 )

#> Using stock residual migration method

# Or directly the mid_resid_stock function
# (works for both single and five year age groups)

mig_res <-
 mig_resid_stock(
   pop_m_mat = pop_m_mat_five,
   pop_f_mat = pop_f_mat_five,
   sr_m_mat = sr_m_mat_five,
   sr_f_mat = sr_f_mat_five,
   asfr_mat = asfr_mat_five,
   srb_vec = srb_vec_five,
   ages = ages_five,
   ages_asfr = ages_asfr_five
 )

# Net migration for males using stock change method
mig_res$mig_m

#>              1950          1955          1960          1965          1970
#> 0    1.7899322803  3.919021e+00 -5.8543069648 12.9092374640  2.921119e+01
#> 5   51.6027045000  5.385321e+01 22.2235544078 22.9335351570  3.470664e+01
#> 10  29.1541335600  2.840815e+01  8.0306470720 19.3499225033  3.062147e+01
#> 15   7.3711371748  4.381831e+00 -5.0199593568 15.8034252562  3.131827e+01
#> 20  21.6150524800  1.566065e+01 -2.0994925197 40.1160914300  5.180611e+01
#> 25  47.8647752800  6.026311e+01 23.3854103582 84.5010463454  8.870085e+01
#> 30  37.0250555000  4.779935e+01 15.9943014370 54.4810744403  5.262150e+01
#> 35  37.0572080800  3.955049e+01  9.4550107496 26.9198029920  2.392796e+01
#> 40  26.6996119446  1.316362e+01  0.9599276814  8.9668772707  1.721723e+01
#> 45  19.3238383059  8.535687e+00 -5.7535831006  7.5537018836  4.394683e+00
#> 50  10.8706376948  5.802709e+00  0.6788420915  0.5444807024  1.053791e+01
#> 55  -0.4710838276  2.382475e+00 -3.2403068092  6.2785759004 -7.931068e-01
#> 60  -2.6739240018 -5.165887e-01  0.4397308595  6.8077825892  9.082899e+00
#> 65   8.2682526723  9.950218e+00  4.2594462540  9.7035767336  8.412578e+00
#> 70   5.1050824363  5.938504e+00  5.7341530320  1.5588869257  6.650109e+00
#> 75  -0.5308153600  3.023748e-01 -0.9342732742 -1.0616547642  1.687302e-01
#> 80   2.6458591680  2.923921e+00  2.3863401360  3.0909371135 -6.546424e-01
#> 85  -0.0002322246  6.833100e-05 -0.0016322983  0.0003876803 -2.982705e-04
#> 90  -0.0007090603 -4.370166e-04  0.0007138143  0.0007593676  4.629872e-05
#> 95  -0.0004184120  8.371022e-05 -0.0002732981 -0.0013148423 -9.538320e-05
#> 100  0.0003631314 -3.905702e-04  0.0004240167 -0.0001150535  1.791806e-04
#>              1975          1980          1985          1990          1995
#> 0   14.4819380766 15.0050326629 30.6160862966  2.279134e+01 33.4094386250
#> 5   24.2440961290 14.9144048249 39.4586215291  3.366381e+01 27.6300354900
#> 10  19.8909763056 12.9976172834 31.3140753898  3.926617e+01 35.2017492389
#> 15  34.5412167950 31.1408714290 53.7835608961  5.806982e+01 56.4028387060
#> 20  35.7746574910 48.8141184045 46.6091830651  3.355813e+01 46.6090142025
#> 25  12.2320674097 15.7490655447 36.0248896021  2.994211e+01 35.2341965234
#> 30  15.2768180385 15.1454438950 48.9912652841  2.939288e+01 43.0299390765
#> 35   3.4133017642  2.6512718093 35.7081057973  2.377502e+01 38.0465234759
#> 40  11.3648881127  0.9157143211 26.0002282136  1.815304e+01 25.2147939559
#> 45  -0.9025899903 -6.1414794104  9.4697258728  9.669273e+00 15.3801489305
#> 50   4.1228516807  1.9180957545 11.4602048293  7.420501e-01  9.1744221357
#> 55   2.8324768102  8.0029005255 13.9249297918  9.461663e+00  7.0314765200
#> 60   5.9587573287  6.1901407708 12.8552522823  1.254064e+01  5.9398053623
#> 65   5.0265223258  6.1334063238 15.3105496329  6.693441e+00  6.2767061297
#> 70   3.0667904029  2.1040520686  8.0829480782  9.051541e-01  4.9118127171
#> 75   1.1085509966  0.4440585403  4.6902499048 -1.740263e+00  1.6956094777
#> 80  -0.3545171554 -1.0788234380 -0.8912981728 -3.259234e+00  1.0782154117
#> 85  -0.0007250264 -0.0003707507  0.0009264276 -1.502028e-03  0.0006504005
#> 90  -0.0010096296 -0.0013578629 -0.0001364608 -1.536187e-03  0.0002672308
#> 95  -0.0005597711  0.0002593028  0.0006724129  4.940180e-06  0.0000220428
#> 100 -0.0002663077  0.0002379727  0.0001025889 -1.467865e-04 -0.0004672003
#>              2000          2005          2010          2015          2020
#> 0   35.2063149299  2.574073e+01  1.686618e+01  3.726188e+01  3.596241e+01
#> 5   39.7815636351  5.198246e+01  3.545773e+01  1.621746e+01  1.572724e+01
#> 10  43.6134320305  4.087724e+01  2.529691e+01  1.243047e+01  1.205461e+01
#> 15  69.4862261899  6.151655e+01  1.053807e+02  6.337355e+01  6.148319e+01
#> 20  61.4795683305  1.081416e+02  1.462701e+02  1.140761e+02  1.106660e+02
#> 25  40.9816130294  9.212637e+01  2.026667e+01  1.096321e+02  1.063574e+02
#> 30  54.2000383773  6.484557e+01  5.491794e+01  8.303648e+01  8.055508e+01
#> 35  54.3621106693  6.210615e+01  4.419383e+01  5.759037e+01  5.586794e+01
#> 40  46.1009769761  4.050389e+01  2.708597e+01  3.866853e+01  3.751901e+01
#> 45  24.3645150956  3.725019e+01  8.129041e+00  2.565676e+01  2.489228e+01
#> 50  19.0808001159  1.891404e+01  1.625201e+01  1.695107e+01  1.644513e+01
#> 55   0.6483630675  1.268029e+01  3.792549e+01  1.118371e+01  1.085143e+01
#> 60  15.0435641770  1.036964e+01  1.188830e+01  7.376952e+00  7.159785e+00
#> 65   6.2874481579  8.037788e+00  3.314290e+01  4.860257e+00  4.720100e+00
#> 70   1.6845365642  1.544022e+00  1.088919e+00  3.201408e+00  3.110899e+00
#> 75   1.7851618553  2.240903e+00  2.708415e+00  2.114225e+00  2.050721e+00
#> 80   4.4904038435  1.151357e+01 -2.168209e+00  1.393847e+00  1.351674e+00
#> 85   0.0003877412 -3.414889e-04  1.892320e-03  1.947218e-03 -5.815492e-05
#> 90  -0.0002732069  3.774664e-04 -1.778366e-03 -4.196944e-04  6.462397e-04
#> 95   0.0003672265 -2.385004e-04 -1.187608e-03  9.547959e-04 -4.276800e-04
#> 100  0.0004257843  1.591630e-04 -5.068183e-05 -3.603256e-04 -1.501552e-04
#>             2025          2030          2035          2040          2045
#> 0   3.713937e+01  3.872296e+01  38.710244991  3.875942e+01  3.884919e+01
#> 5   1.624346e+01  1.692868e+01  16.922542527  1.694990e+01  1.698616e+01
#> 10  1.245485e+01  1.298092e+01  12.972254791  1.299493e+01  1.303085e+01
#> 15  6.348393e+01  6.617035e+01  66.152791816  6.626290e+01  6.641678e+01
#> 20  1.142679e+02  1.191112e+02 119.075586720  1.192639e+02  1.195424e+02
#> 25  1.098159e+02  1.144742e+02 114.442603943  1.146180e+02  1.148904e+02
#> 30  8.317144e+01  8.669744e+01  86.667686177  8.681930e+01  8.701090e+01
#> 35  5.769732e+01  6.012598e+01  60.113368482  6.020745e+01  6.034099e+01
#> 40  3.874399e+01  4.037813e+01  40.365452745  4.042687e+01  4.052742e+01
#> 45  2.570411e+01  2.679773e+01  26.784212118  2.683642e+01  2.688482e+01
#> 50  1.699112e+01  1.771062e+01  17.697703877  1.773515e+01  1.777363e+01
#> 55  1.120040e+01  1.167745e+01  11.672268085  1.169656e+01  1.172089e+01
#> 60  7.386918e+00  7.707266e+00   7.697176849  7.715443e+00  7.730054e+00
#> 65  4.878769e+00  5.078950e+00   5.072805619  5.084954e+00  5.100982e+00
#> 70  3.212492e+00  3.348749e+00   3.347115246  3.356025e+00  3.363517e+00
#> 75  2.116043e+00  2.206720e+00   2.205534484  2.206645e+00  2.212107e+00
#> 80  1.392679e+00  1.453848e+00   1.454611333  1.461609e+00  1.459596e+00
#> 85  5.095878e-04  1.592002e-03  -0.003113428 -1.789957e-04 -1.280066e-03
#> 90  6.580606e-04 -6.258823e-04   0.001700642 -2.324909e-03  1.481708e-03
#> 95  5.002212e-05 -3.570606e-04  -0.001343537  1.088385e-03  9.059035e-05
#> 100 6.132923e-05 -2.312750e-04   0.000230856 -1.040497e-04  4.127133e-05

# Net migration for females using stock change method
mig_res$mig_f

#>              1950          1955          1960          1965          1970
#> 0   -8.339360e+02 -9.668708e+02 -1.000674e+03 -8.279100e+02 -7.677141e+02
#> 5    9.534564e-02  9.741671e-02  7.019452e-02  6.587207e-02  5.982639e-02
#> 10   3.938204e-02  3.798494e-02  2.159118e-02  3.006125e-02  4.085119e-02
#> 15   1.893728e-02  1.546635e-02  1.501716e-02  3.573094e-02  4.355191e-02
#> 20   3.506867e-02  3.829911e-02  3.249304e-02  6.196567e-02  7.138232e-02
#> 25   4.652932e-02  5.115055e-02  2.978558e-02  6.940305e-02  7.337267e-02
#> 30   4.926660e-02  5.364078e-02  2.811914e-02  4.128898e-02  4.614464e-02
#> 35   4.294238e-02  4.762602e-02  2.465858e-02  2.585740e-02  2.483667e-02
#> 40   2.940340e-02  2.097116e-02  1.349470e-02  1.116104e-02  2.358218e-02
#> 45   1.857087e-02  1.719342e-02  7.684818e-03  1.601700e-02  1.608625e-02
#> 50   1.371191e-02  1.711940e-02  1.392770e-02  1.362687e-02  2.744808e-02
#> 55   1.000670e-02  1.447415e-02  1.065373e-02  2.255101e-02  1.536314e-02
#> 60   1.150849e-02  1.218200e-02  1.850430e-02  2.764042e-02  2.891679e-02
#> 65   2.250025e-02  2.523334e-02  2.644347e-02  3.500253e-02  3.469564e-02
#> 70   2.900955e-02  3.129857e-02  3.248546e-02  3.231810e-02  3.252892e-02
#> 75   1.682217e-02  1.804474e-02  2.147476e-02  3.018757e-02  3.305470e-02
#> 80   1.908660e-02  2.179088e-02  2.653310e-02  3.116165e-02  3.070524e-02
#> 85   7.007998e-03  9.630182e-03  1.175235e-02  1.601617e-02  1.984793e-02
#> 90   2.737847e-03  3.099530e-03  4.375230e-03  5.830268e-03  8.054336e-03
#> 95   4.775703e-04  6.346433e-04  7.378362e-04  1.166248e-03  1.622854e-03
#> 100  3.766896e-05  4.934962e-05  6.275325e-05  8.664575e-05  1.430509e-04
#>              1975          1980          1985          1990          1995
#> 0   -7.975983e+02 -8.287872e+02 -8.592653e+02 -8.930582e+02 -8.010066e+02
#> 5    4.453593e-02  3.303733e-02  5.436693e-02  4.787678e-02  4.184107e-02
#> 10   2.522668e-02  2.359719e-02  3.554828e-02  4.293321e-02  3.783614e-02
#> 15   4.441858e-02  3.006664e-02  4.922688e-02  5.410500e-02  5.301117e-02
#> 20   6.041144e-02  5.187299e-02  6.652346e-02  5.334868e-02  5.718992e-02
#> 25   2.885065e-02  2.482477e-02  5.927889e-02  4.934692e-02  4.800412e-02
#> 30   3.061833e-02  3.040210e-02  6.323398e-02  4.490459e-02  5.458493e-02
#> 35   1.290396e-02  1.375009e-02  4.613172e-02  4.222216e-02  5.424249e-02
#> 40   1.950625e-02  1.681778e-02  4.246180e-02  3.563240e-02  4.460592e-02
#> 45   1.249823e-02  1.067645e-02  2.401290e-02  3.167980e-02  3.630227e-02
#> 50   1.800190e-02  1.912011e-02  2.168568e-02  2.561224e-02  2.817995e-02
#> 55   2.173224e-02  2.716630e-02  2.597592e-02  3.239668e-02  2.543971e-02
#> 60   2.914986e-02  3.141690e-02  3.233337e-02  3.366166e-02  2.944627e-02
#> 65   3.663175e-02  3.852460e-02  4.607660e-02  3.789778e-02  3.743619e-02
#> 70   3.874703e-02  4.133303e-02  4.552331e-02  4.536883e-02  4.730059e-02
#> 75   3.906288e-02  4.385171e-02  5.226274e-02  4.896906e-02  5.998545e-02
#> 80   3.626655e-02  3.916344e-02  4.429909e-02  5.069815e-02  6.082249e-02
#> 85   2.487058e-02  3.024550e-02  3.437834e-02  4.097165e-02  4.756603e-02
#> 90   1.168350e-02  1.539064e-02  1.751129e-02  2.060835e-02  2.280022e-02
#> 95   2.876118e-03  4.418639e-03  5.167235e-03  5.999656e-03  6.022099e-03
#> 100  2.940806e-04  5.528349e-04  7.229901e-04  8.275142e-04  6.994879e-04
#>              2000          2005          2010          2015          2020
#> 0   -7.729746e+02 -8.694947e+02 -9.084079e+02 -9.169866e+02 -9.390321e+02
#> 5    4.644909e-02  5.301343e-02  4.401259e-02  2.165495e-02  1.571378e-02
#> 10   4.891408e-02  4.109456e-02  2.849520e-02  1.586824e-02  1.285681e-02
#> 15   6.111895e-02  5.930546e-02  9.381769e-02  7.230964e-02  6.796569e-02
#> 20   8.081864e-02  1.082549e-01  1.505999e-01  1.253307e-01  1.182762e-01
#> 25   6.792642e-02  1.111587e-01  8.238454e-02  1.185424e-01  1.095158e-01
#> 30   7.408341e-02  1.030107e-01  1.000555e-01  8.960123e-02  7.987071e-02
#> 35   6.423241e-02  7.975420e-02  6.620362e-02  6.432198e-02  5.333906e-02
#> 40   6.325721e-02  5.906802e-02  5.147925e-02  4.609590e-02  3.449321e-02
#> 45   4.115209e-02  5.746749e-02  2.546582e-02  3.451742e-02  2.203458e-02
#> 50   3.785057e-02  3.924090e-02  3.640836e-02  2.921954e-02  1.402299e-02
#> 55   3.564659e-02  3.957208e-02  4.665738e-02  3.240938e-02  8.899044e-03
#> 60   4.230244e-02  4.510012e-02  3.750044e-02  3.850735e-02  5.651936e-03
#> 65   4.003094e-02  4.446262e-02  5.318551e-02  4.781480e-02  3.591765e-03
#> 70   4.805561e-02  4.562773e-02  5.570568e-02  6.238537e-02  2.279530e-03
#> 75   6.194085e-02  6.305435e-02  6.597926e-02  7.158288e-02  1.447269e-03
#> 80   7.637618e-02  8.619569e-02  7.519138e-02  7.840854e-02  9.191045e-04
#> 85   5.400946e-02  7.035971e-02  7.558003e-02  7.366929e-02  2.023718e-06
#> 90   2.778770e-02  3.433981e-02  4.413133e-02  4.577276e-02 -4.575784e-08
#> 95   6.946771e-03  9.556226e-03  1.103225e-02  1.304860e-02 -6.680146e-07
#> 100  6.407134e-04  8.698355e-04  9.192660e-04  8.355002e-04 -3.823113e-09
#>              2025          2030          2035          2040          2045
#> 0   -9.450575e+02 -9.530888e+02 -9.745826e+02 -1.006847e+03 -1.042720e+03
#> 5    1.622971e-02  1.691831e-02  1.691457e-02  1.693440e-02  1.697581e-02
#> 10   1.327584e-02  1.383276e-02  1.383380e-02  1.385223e-02  1.388592e-02
#> 15   7.018092e-02  7.315433e-02  7.313045e-02  7.324795e-02  7.341229e-02
#> 20   1.221251e-01  1.272877e-01  1.272581e-01  1.274533e-01  1.277532e-01
#> 25   1.130743e-01  1.178537e-01  1.178278e-01  1.180086e-01  1.182870e-01
#> 30   8.246815e-02  8.595619e-02  8.594408e-02  8.607590e-02  8.627359e-02
#> 35   5.508070e-02  5.740907e-02  5.739172e-02  5.748318e-02  5.762230e-02
#> 40   3.560806e-02  3.711957e-02  3.711233e-02  3.716508e-02  3.724629e-02
#> 45   2.275099e-02  2.371591e-02  2.371210e-02  2.374261e-02  2.380269e-02
#> 50   1.447923e-02  1.509269e-02  1.509005e-02  1.510669e-02  1.514986e-02
#> 55   9.189698e-03  9.580706e-03  9.583997e-03  9.598664e-03  9.619318e-03
#> 60   5.833683e-03  6.079587e-03  6.081444e-03  6.084633e-03  6.102345e-03
#> 65   3.699831e-03  3.864664e-03  3.859534e-03  3.866878e-03  3.879600e-03
#> 70   2.350295e-03  2.453483e-03  2.451742e-03  2.453048e-03  2.460003e-03
#> 75   1.494431e-03  1.554029e-03  1.554556e-03  1.562182e-03  1.559817e-03
#> 80   9.422448e-04  9.851964e-04  9.835399e-04  9.827943e-04  9.950495e-04
#> 85  -1.309560e-06 -3.205581e-06  2.799884e-06  3.560486e-06 -4.231530e-06
#> 90  -1.760394e-06  5.530703e-08 -7.542806e-07  2.292086e-06 -1.102522e-06
#> 95   1.115547e-06  2.261978e-07  7.982875e-07  2.063124e-06 -4.913987e-07
#> 100 -6.037463e-07  1.228691e-08 -5.497503e-07 -2.754639e-07 -2.854814e-08

################ cohort even flow method  #####################

# We reuse the same data from before
# Either use the generic mig_resid choosing 'cohort'

# Five year age groups
mig_res <-
 mig_resid(
   pop_m_mat = pop_m_mat_five,
   pop_f_mat = pop_f_mat_five,
   sr_m_mat = sr_m_mat_five,
   sr_f_mat = sr_f_mat_five,
   asfr_mat = asfr_mat_five,
   srb_vec = srb_vec_five,
   ages = ages_five,
   ages_asfr = ages_asfr_five,
   # With the cohort method
   method = "cohort"
 )

#> Using cohort residual migration method

# Single ages
mig_res <-
 mig_resid(
   pop_m_mat = pop_m_mat_single,
   pop_f_mat = pop_f_mat_single,
   sr_m_mat = sr_m_mat_single,
   sr_f_mat = sr_f_mat_single,
   asfr_mat = asfr_mat_single,
   srb_vec = srb_vec_single,
   ages = ages_single,
   ages_asfr = ages_asfr_single,
   # With the stock method
   method = "cohort"
 )

#> Using cohort residual migration method

# Or directly the mid_resid_cohort function

mig_res <-
  mig_resid_cohort(
    pop_m_mat = pop_m_mat_five,
    pop_f_mat = pop_f_mat_five,
    sr_m_mat = sr_m_mat_five,
    sr_f_mat = sr_f_mat_five,
    asfr_mat = asfr_mat_five,
    srb_vec = srb_vec_five,
    ages = ages_five,
    ages_asfr = ages_asfr_five
  )

# Net migration for males using the cohort even flow method
mig_res$mig_m

#>              1950          1955          1960          1965          1970
#> 0   27.7345854720 31.0019371482  5.1942053540 24.5624659219  4.688525e+01
#> 5   40.5102116592 41.2343357598 15.1606536999 21.1793513869  3.271567e+01
#> 10  18.3013151966 16.4222220410  1.5058793365 17.6053516384  3.102412e+01
#> 15  14.5468170638 10.0564010777 -3.5687147753 28.0614898551  4.172147e+01
#> 20  34.8888314231 38.1140374410 10.6856647744 62.5658663403  7.055654e+01
#> 25  42.6329532877 54.2516150063 19.7676547862 69.7666525943  7.093797e+01
#> 30  37.2248714152 43.8717655890 12.7781665228 40.8677327075  3.842664e+01
#> 35  32.0814535293 26.5020660760  5.2333131522 18.0380639382  2.068785e+01
#> 40  23.2334561570 10.9459192982 -2.4258656763  8.3347645192  1.089106e+01
#> 45  15.3265554535  7.2748524613 -2.5633574346  4.0973563278  7.587504e+00
#> 50   5.3028125043  4.1864884603 -1.3245914367  3.5133073257  4.957193e+00
#> 55  -1.6476700692  0.9576589788 -1.4395096532  6.8074763188  4.347471e+00
#> 60   3.0476956367  5.0897985122  2.5236522427  8.7981552818  9.271884e+00
#> 65   7.2975977560  8.6761911905  5.4904630231  6.0964521738  8.217822e+00
#> 70   2.5311742841  3.4940993902  2.6452918966  0.2448806986  3.800742e+00
#> 75   1.3981785084  2.0660658863  0.9794162929  1.3492354084 -3.171552e-01
#> 80   1.7122472632  1.8870772095  1.5306732316  1.9744450576 -4.164296e-01
#> 85  -0.0008180112 -0.0003418182 -0.0005657794  0.0009453795 -1.747616e-04
#> 90  -0.0011282773 -0.0002951569  0.0003245206 -0.0008184162 -6.416606e-05
#> 95  -0.0009690561  0.0001945259 -0.0006136342 -0.0029657259 -2.079006e-04
#> 100  0.0003631314 -0.0003905702  0.0004240167 -0.0001150535  1.791806e-04
#>              1975          1980          1985          1990          1995
#> 0   26.7290464322 22.5514135124 50.5005203859  3.971947e+01  4.733699e+01
#> 5   22.0962653344 13.9694166410 35.4141109601  3.648698e+01  3.143147e+01
#> 10  27.2613524622 22.0997423995 42.5971117509  4.871322e+01  4.583846e+01
#> 15  35.2720807678 40.0886284255 50.3067733763  4.589406e+01  5.158882e+01
#> 20  24.1056789950 32.3931085097 41.4420067581  3.183571e+01  4.101680e+01
#> 25  13.8048985375 15.4971051546 42.6358411330  2.975410e+01  3.922889e+01
#> 30   9.3790926968  8.9270470232 42.4876305010  2.667472e+01  4.065530e+01
#> 35   7.4318376300  1.7905172492 30.9784476453  2.105322e+01  3.174649e+01
#> 40   5.2620681370 -2.6361050610 17.8294359966  1.398769e+01  2.039907e+01
#> 45   1.6409798784 -2.1235510680 10.5680414394  5.242261e+00  1.236784e+01
#> 50   3.5528032225  5.0697335208 12.9059761124  5.190501e+00  8.201545e+00
#> 55   4.5669024604  7.3126011011 13.7574122211  1.127484e+01  6.618932e+00
#> 60   5.7964848148  6.4725021367 14.7392106038  9.973541e+00  6.324542e+00
#> 65   4.3823473498  4.4122945905 12.4941899906  4.003111e+00  5.912938e+00
#> 70   2.3424884610  1.4118350206  7.0917545044 -4.958813e-01  3.587037e+00
#> 75   0.4223807796 -0.4160797352  2.1344354343 -2.976618e+00  1.602920e+00
#> 80  -0.2237692930 -0.6727819681 -0.5508956693 -1.992855e+00  6.527169e-01
#> 85  -0.0013472520 -0.0013649720  0.0005310532 -2.288892e-03  6.637464e-04
#> 90  -0.0014153765 -0.0008447875  0.0005782918 -1.248544e-03  2.421791e-04
#> 95  -0.0011599514  0.0005239992  0.0013807440  1.016482e-05  4.658246e-05
#> 100 -0.0002663077  0.0002379727  0.0001025889 -1.467865e-04 -4.672003e-04
#>             2000          2005          2010          2015          2020
#> 0   5.521348e+01  5.181794e+01  3.464569e+01  4.546868e+01  4.391070e+01
#> 5   4.171498e+01  4.644763e+01  3.038654e+01  1.432799e+01  1.389523e+01
#> 10  5.658804e+01  5.122663e+01  6.537600e+01  3.792246e+01  3.678382e+01
#> 15  6.557516e+01  8.495273e+01  1.259713e+02  8.882648e+01  8.614305e+01
#> 20  5.133396e+01  1.003256e+02  8.339817e+01  1.120286e+02  1.086440e+02
#> 25  4.769132e+01  7.864393e+01  3.765750e+01  9.649354e+01  9.359811e+01
#> 30  5.441186e+01  6.361715e+01  4.964878e+01  7.043648e+01  6.832703e+01
#> 35  5.038857e+01  5.144900e+01  3.572532e+01  4.823902e+01  4.679176e+01
#> 40  3.538553e+01  3.904144e+01  1.766614e+01  3.227328e+01  3.130150e+01
#> 45  2.187529e+01  2.825565e+01  1.226915e+01  2.142061e+01  2.076887e+01
#> 50  9.953746e+00  1.595878e+01  2.736801e+01  1.418917e+01  1.375317e+01
#> 55  8.024718e+00  1.171475e+01  2.523316e+01  9.404806e+00  9.113674e+00
#> 60  1.095858e+01  9.441914e+00  2.309601e+01  6.246623e+00  6.051291e+00
#> 65  4.157159e+00  4.963837e+00  1.761464e+01  4.165787e+00  4.032837e+00
#> 70  1.879553e+00  2.032313e+00  2.029089e+00  2.804870e+00  2.709220e+00
#> 75  3.633292e+00  7.856819e+00  2.269679e-01  1.921612e+00  1.849407e+00
#> 80  2.648544e+00  6.637323e+00 -1.231256e+00  7.883184e-01  7.562531e-01
#> 85  4.033675e-05  7.105796e-05 -1.498956e-04  9.007669e-04  4.340415e-04
#> 90  1.617833e-04  5.104875e-05 -2.514495e-03  6.246262e-04  5.782513e-05
#> 95  7.570578e-04 -4.775804e-04 -2.347887e-03  1.876386e-03 -8.238661e-04
#> 100 4.257843e-04  1.591630e-04 -5.068183e-05 -3.603256e-04 -1.501552e-04
#>             2025          2030          2035          2040         2045
#> 0   4.533636e+01  4.725714e+01  4.723415e+01  4.729110e+01 4.739403e+01
#> 5   1.435297e+01  1.495834e+01  1.495057e+01  1.497529e+01 1.501112e+01
#> 10  3.798272e+01  3.958804e+01  3.957368e+01  3.963905e+01 3.973308e+01
#> 15  8.893813e+01  9.269935e+01  9.266739e+01  9.281208e+01 9.302437e+01
#> 20  1.121642e+02  1.169091e+02  1.168658e+02  1.170394e+02 1.173074e+02
#> 25  9.662625e+01  1.007130e+02  1.006725e+02  1.008273e+02 1.010516e+02
#> 30  7.054235e+01  7.351530e+01  7.348607e+01  7.360189e+01 7.375815e+01
#> 35  4.831215e+01  5.033959e+01  5.031996e+01  5.039165e+01 5.050329e+01
#> 40  3.231319e+01  3.367316e+01  3.365320e+01  3.370409e+01 3.377327e+01
#> 45  2.144079e+01  2.234327e+01  2.232287e+01  2.236153e+01 2.239942e+01
#> 50  1.419322e+01  1.478717e+01  1.477057e+01  1.479497e+01 1.482056e+01
#> 55  9.393798e+00  9.787974e+00  9.772460e+00  9.787043e+00 9.800488e+00
#> 60  6.235955e+00  6.491396e+00  6.475070e+00  6.483358e+00 6.492494e+00
#> 65  4.154458e+00  4.317637e+00  4.305202e+00  4.308499e+00 4.313789e+00
#> 70  2.783655e+00  2.892001e+00  2.881492e+00  2.878664e+00 2.878182e+00
#> 75  1.892981e+00  1.963852e+00  1.953794e+00  1.949436e+00 1.943300e+00
#> 80  7.721803e-01  8.009601e-01  7.934284e-01  7.943901e-01 7.884978e-01
#> 85  7.786621e-04  5.130995e-04 -6.648082e-04 -1.703315e-03 2.604560e-04
#> 90  5.174199e-04 -7.749870e-04 -5.907210e-05 -6.034064e-04 1.089173e-03
#> 95  9.482005e-05 -6.681631e-04 -2.483761e-03  1.989245e-03 1.637868e-04
#> 100 6.132923e-05 -2.312750e-04  2.308560e-04 -1.040497e-04 4.127133e-05

# Net migration for females using the cohort even flow method
mig_res$mig_f

#>              1950          1955          1960          1965          1970
#> 0   -9.021018e+02 -1.035306e+03 -1.062560e+03 -8.722800e+02 -8.035905e+02
#> 5    6.920081e-02  6.928239e-02  4.684641e-02  4.875619e-02  5.099224e-02
#> 10   2.946478e-02  2.696618e-02  1.844417e-02  3.310559e-02  4.243743e-02
#> 15   2.729601e-02  2.713734e-02  2.395068e-02  4.919164e-02  5.781832e-02
#> 20   4.133304e-02  4.522977e-02  3.144274e-02  6.624574e-02  7.291473e-02
#> 25   4.861860e-02  5.307862e-02  2.928367e-02  5.588984e-02  6.027166e-02
#> 30   4.688562e-02  5.138378e-02  2.673501e-02  3.395471e-02  3.584118e-02
#> 35   3.686299e-02  3.486860e-02  1.936135e-02  1.875110e-02  2.449680e-02
#> 40   2.451040e-02  1.945753e-02  1.077529e-02  1.381137e-02  2.011370e-02
#> 45   1.656536e-02  1.757029e-02  1.105335e-02  1.511775e-02  2.217903e-02
#> 50   1.226265e-02  1.629328e-02  1.264430e-02  1.859840e-02  2.191343e-02
#> 55   1.130080e-02  1.392818e-02  1.523128e-02  2.610027e-02  2.299397e-02
#> 60   1.840401e-02  2.013441e-02  2.401143e-02  3.327825e-02  3.364233e-02
#> 65   2.900225e-02  3.155144e-02  3.265725e-02  3.693669e-02  3.667127e-02
#> 70   2.717509e-02  2.894144e-02  3.142784e-02  3.627915e-02  3.779607e-02
#> 75   2.383314e-02  2.611712e-02  3.107521e-02  3.900073e-02  3.994370e-02
#> 80   1.913202e-02  2.282359e-02  2.736889e-02  3.337685e-02  3.552237e-02
#> 85   8.570651e-03  1.090765e-02  1.364066e-02  1.813801e-02  2.285592e-02
#> 90   3.418159e-03  3.919393e-03  5.219671e-03  7.058262e-03  9.571480e-03
#> 95   1.252216e-03  1.635147e-03  1.862978e-03  2.893642e-03  3.943889e-03
#> 100  3.766896e-05  4.934962e-05  6.275325e-05  8.664575e-05  1.430509e-04
#>              1975          1980          1985          1990          1995
#> 0   -8.302725e+02 -8.576678e+02 -8.852852e+02 -9.175961e+02 -8.201743e+02
#> 5    3.528539e-02  2.857301e-02  4.531243e-02  4.571595e-02  4.006480e-02
#> 10   3.499371e-02  2.693969e-02  4.253696e-02  4.868456e-02  4.555738e-02
#> 15   5.270390e-02  4.115812e-02  5.810836e-02  5.393132e-02  5.529366e-02
#> 20   4.492143e-02  3.855642e-02  6.321548e-02  5.159850e-02  5.284309e-02
#> 25   2.996825e-02  2.779777e-02  6.162342e-02  4.740512e-02  5.160834e-02
#> 30   2.195259e-02  2.224223e-02  5.506241e-02  4.387312e-02  5.482470e-02
#> 35   1.638255e-02  1.542823e-02  4.467217e-02  3.926111e-02  4.987627e-02
#> 40   1.620701e-02  1.390226e-02  3.357915e-02  3.401322e-02  4.090281e-02
#> 45   1.551274e-02  1.513240e-02  2.316572e-02  2.903585e-02  3.268978e-02
#> 50   2.033159e-02  2.364304e-02  2.430273e-02  2.956797e-02  2.731674e-02
#> 55   2.632446e-02  3.021525e-02  3.001972e-02  3.395843e-02  2.820850e-02
#> 60   3.466570e-02  3.671803e-02  4.107359e-02  3.737014e-02  3.489756e-02
#> 65   4.093090e-02  4.313549e-02  4.922441e-02  4.467055e-02  4.534973e-02
#> 70   4.444454e-02  4.830363e-02  5.512297e-02  5.284069e-02  5.994264e-02
#> 75   4.661403e-02  5.073629e-02  5.833635e-02  5.996094e-02  7.213774e-02
#> 80   4.236772e-02  4.759447e-02  5.319997e-02  6.129814e-02  7.167610e-02
#> 85   2.954270e-02  3.636650e-02  4.063070e-02  4.742013e-02  5.337645e-02
#> 90   1.417711e-02  1.903697e-02  2.137876e-02  2.459435e-02  2.611270e-02
#> 95   6.802691e-03  1.023439e-02  1.170030e-02  1.330593e-02  1.311551e-02
#> 100  2.940806e-04  5.528349e-04  7.229901e-04  8.275142e-04  6.994879e-04
#>              2000          2005          2010          2015          2020
#> 0   -7.885975e+02 -8.838746e+02 -9.206427e+02 -9.274007e+02 -9.409568e+02
#> 5    4.789851e-02  4.723318e-02  3.636445e-02  1.880751e-02  1.428867e-02
#> 10   5.515970e-02  5.031059e-02  6.126959e-02  4.415891e-02  4.042136e-02
#> 15   7.119899e-02  8.401840e-02  1.224974e-01  9.902303e-02  9.315949e-02
#> 20   7.470320e-02  1.101303e-01  1.168419e-01  1.222617e-01  1.139634e-01
#> 25   7.144286e-02  1.076456e-01  9.159961e-02  1.044229e-01  9.476476e-02
#> 30   6.970228e-02  9.200237e-02  8.355655e-02  7.728825e-02  6.666796e-02
#> 35   6.436625e-02  6.999775e-02  5.922643e-02  5.550694e-02  4.397429e-02
#> 40   5.279877e-02  5.886760e-02  3.877960e-02  4.058715e-02  2.832219e-02
#> 45   4.006024e-02  4.895953e-02  3.127957e-02  3.216736e-02  1.808882e-02
#> 50   3.743268e-02  4.007877e-02  4.216363e-02  3.123129e-02  1.152124e-02
#> 55   4.002258e-02  4.337739e-02  4.297337e-02  3.617621e-02  7.334320e-03
#> 60   4.281333e-02  4.643919e-02  4.692849e-02  4.453795e-02  4.679762e-03
#> 65   4.694410e-02  4.773597e-02  5.743028e-02  5.798596e-02  2.994345e-03
#> 70   6.107370e-02  5.995542e-02  6.652919e-02  7.280986e-02  1.925118e-03
#> 75   8.209279e-02  8.781068e-02  8.190888e-02  8.621005e-02  1.251706e-03
#> 80   8.468009e-02  1.006441e-01  9.614670e-02  9.552677e-02  4.973036e-04
#> 85   6.095627e-02  7.635063e-02  8.639627e-02  8.484706e-02  1.139063e-06
#> 90   3.057028e-02  3.787123e-02  4.646163e-02  4.864233e-02 -5.916063e-07
#> 95   1.465609e-02  1.962123e-02  2.218324e-02  2.551671e-02 -1.122828e-06
#> 100  6.407134e-04  8.698355e-04  9.192660e-04  8.355002e-04 -3.823113e-09
#>              2025          2030          2035          2040          2045
#> 0   -9.467816e+02 -9.546117e+02 -9.759720e+02 -1.008130e+03 -1.043906e+03
#> 5    1.475587e-02  1.537836e-02  1.537671e-02  1.539557e-02  1.543288e-02
#> 10   4.173772e-02  4.350211e-02  4.348980e-02  4.355698e-02  4.365529e-02
#> 15   9.618904e-02  1.002545e-01  1.002246e-01  1.003781e-01  1.006076e-01
#> 20   1.176635e-01  1.226309e-01  1.225980e-01  1.227812e-01  1.230660e-01
#> 25   9.783964e-02  1.019700e-01  1.019459e-01  1.020974e-01  1.023309e-01
#> 30   6.883477e-02  7.174004e-02  7.172074e-02  7.182816e-02  7.199256e-02
#> 35   4.539983e-02  4.731691e-02  4.730032e-02  4.736846e-02  4.747491e-02
#> 40   2.923507e-02  3.047038e-02  3.046052e-02  3.049816e-02  3.056507e-02
#> 45   1.867232e-02  1.945852e-02  1.945084e-02  1.947030e-02  1.951809e-02
#> 50   1.189181e-02  1.239102e-02  1.238686e-02  1.239838e-02  1.242642e-02
#> 55   7.567618e-03  7.883037e-03  7.881202e-03  7.886047e-03  7.901435e-03
#> 60   4.821689e-03  5.024070e-03  5.018025e-03  5.019264e-03  5.030757e-03
#> 65   3.080767e-03  3.211853e-03  3.203963e-03  3.204237e-03  3.210327e-03
#> 70   1.981168e-03  2.059526e-03  2.054404e-03  2.054788e-03  2.053114e-03
#> 75   1.283795e-03  1.332189e-03  1.326888e-03  1.326015e-03  1.327060e-03
#> 80   5.053310e-04  5.244282e-04  5.245498e-04  5.223082e-04  5.222588e-04
#> 85  -1.900901e-06 -1.787040e-06  1.098362e-06  3.448556e-06 -3.048916e-06
#> 90  -2.214464e-07  2.219107e-07  1.682787e-07  3.117486e-06 -1.084673e-06
#> 95   1.857040e-06  3.723792e-07  1.301420e-06  3.330889e-06 -7.856163e-07
#> 100 -6.037463e-07  1.228691e-08 -5.497503e-07 -2.754639e-07 -2.854814e-08

################ time even flow method  #####################

# We reuse the same data from before
# Either use the generic mig_resid with the 'time' method

# For five year age groups
mig_res <-
 mig_resid(
   pop_m_mat = pop_m_mat_five,
   pop_f_mat = pop_f_mat_five,
   sr_m_mat = sr_m_mat_five,
   sr_f_mat = sr_f_mat_five,
   asfr_mat = asfr_mat_five,
   srb_vec = srb_vec_five,
   ages = ages_five,
   ages_asfr = ages_asfr_five,
   # With the time method
   method = "time"
 )

#> Using time residual migration method

# For single ages
mig_res <-
 mig_resid(
   pop_m_mat = pop_m_mat_single,
   pop_f_mat = pop_f_mat_single,
   sr_m_mat = sr_m_mat_single,
   sr_f_mat = sr_f_mat_single,
   asfr_mat = asfr_mat_single,
   srb_vec = srb_vec_single,
   ages = ages_single,
   ages_asfr = ages_asfr_single,
   # With the stock method
   method = "stock"
 )

#> Using stock residual migration method

# Or directly the mid_resid_time function
# (works for both five and single year ages)

mig_res <-
  mig_resid_time(
    pop_m_mat = pop_m_mat_five,
    pop_f_mat = pop_f_mat_five,
    sr_m_mat = sr_m_mat_five,
    sr_f_mat = sr_f_mat_five,
    asfr_mat = asfr_mat_five,
    srb_vec = srb_vec_five,
    ages = ages_five,
    ages_asfr = ages_asfr_five
  )

# Net migration for males using the time even flow method
mig_res$mig_m

#>             1950         1955          1960         1965         1970
#> 0     3.57986456   7.83804281 -11.708613930  25.81847493  58.42237710
#> 5    99.65372764  99.91545521  56.096782636  20.16120060  11.21457792
#> 10  -40.94162774 -42.78430686 -39.882708344  18.59078430  50.05512859
#> 15   55.48348586  51.36112198  29.686148299  13.09179503  12.81047107
#> 20  -11.79757463 -19.64383338 -33.657655806  67.25140324  90.92287133
#> 25  107.42501258 140.01220683  80.160934509 102.29400345  87.21131258
#> 30  -32.40904222 -43.26549135 -47.553510864   7.44391246  18.67159066
#> 35  106.17661163 121.93971456  66.014834078  46.46533877  29.35435108
#> 40  -51.15631402 -93.89843855 -63.174578114 -27.88927758   5.48775806
#> 45   88.57046150 108.82847303  50.267264262  42.37000677   3.42290940
#> 50  -63.32934115 -93.13047934 -47.057179615 -39.73615654  17.77493878
#> 55   58.46883834  92.26143114  37.785755257  49.95450413 -18.35527636
#> 60  -58.32986589 -84.64085443 -33.370638099 -31.75995974  34.90416044
#> 65   66.83013145  92.79117747  37.313414670  46.77483495 -13.45378292
#> 70  -43.57046784 -62.56497815 -18.527162169 -34.44289291  24.15556419
#> 75   30.01886407  45.21821705  11.378053697  22.72777192 -17.18309176
#> 80  -12.62450034 -21.29332857  -2.130173274  -7.74682943   9.31779578
#> 85    5.60441174   9.70724986   0.986839238   3.68891047  -4.50288791
#> 90   -1.70123799  -3.01062248  -0.313951952  -1.20302297   1.52466156
#> 95    0.31632108   0.55768479   0.061728916   0.23383116  -0.32111739
#> 100  -0.02811154  -0.05411331  -0.005447638  -0.02448547   0.03644719
#>             1975         1980          1985         1990        1995
#> 0    28.96387615  30.01006533  61.232172593  45.58267324  66.8188773
#> 5    19.61458145  -0.11384016  17.798839229  21.81186415 -11.4789958
#> 10   20.20595285  26.10889922  44.850683328  56.74184507  81.8729514
#> 15   48.95919987  36.25809415  62.845864529  59.53667644  31.0999449
#> 20   23.01801705  61.62328797  30.756451497   7.89400063  62.2618417
#> 25    1.62298038 -29.72231189  41.475065426  52.03339643   8.4963877
#> 30   28.94199434  59.82581388  56.757960851   7.07303243  77.6072932
#> 35  -21.87239151 -54.10090181  15.060691363  40.53058132  -1.0157575
#> 40   44.33117340  55.38825229  37.079303424  -3.83519751  51.4366294
#> 45  -45.24366939 -66.76114668 -17.615872177  23.12250876 -20.0512376
#> 50   51.97198921  68.73233751  40.118018493 -21.15026980  38.0260225
#> 55  -43.51296896 -49.55790353 -10.657202225  39.33641676 -22.7965318
#> 60   51.82676172  58.29271858  35.664479258 -11.98652737  33.5229966
#> 65  -35.33901983 -39.41621873  -1.325981088  24.25817390 -18.2014075
#> 70   34.99220822  36.88719713  17.281657343 -18.97573621  25.6182594
#> 75  -23.57247670 -26.76288480  -3.793379475  11.33350586 -16.9839990
#> 80   14.15059062  15.05938901   0.694028281 -14.10486407  13.7584471
#> 85   -7.09942975  -7.76410472  -0.358176585   7.41969081  -7.3510661
#> 90    2.59035597   2.91884576   0.132464878  -2.78840546   2.7397113
#> 95   -0.60437431  -0.71424802  -0.030070852   0.65864048  -0.6134260
#> 100   0.07855023   0.09881163   0.004110768  -0.08173701   0.0688248
#>            2000         2005          2010        2015        2020        2025
#> 0    70.4126299  51.48145831   33.73236083  74.5237684  71.9248264  74.2787331
#> 5     9.2223352  52.52998713   37.20561118 -42.0424248 -40.4151595 -41.7427286
#> 10   78.0107215  29.25548089   13.40764281  66.8830873  64.5071365  66.6372408
#> 15   61.1002538  93.82266121  197.37105948  59.9439586  58.5164254  60.3817971
#> 20   62.1021505 122.80643192   95.77213448 168.3817044 162.9313907 168.2596094
#> 25   20.1148330  61.93362372  -54.90300355  51.4450008  50.2569949  51.8140624
#> 30   88.3718114  68.01002222  164.54991665 114.7947184 111.0135390 114.6788458
#> 35   20.8205517  56.52957864  -75.48650810   0.8280661   1.1274237   1.0935709
#> 40   71.5336598  24.86139074  129.21554946  76.5135937  73.9162539  76.3993357
#> 45  -22.0122498  49.89176029 -111.77480201 -24.5285913 -23.5568882 -24.4557117
#> 50   59.7886258 -11.24947223  142.64920256  58.0869502  56.1548785  58.1644701
#> 55  -56.8405940  36.31951173  -63.55416721 -34.4574904 -33.3626521 -34.7478059
#> 60   84.3838898 -14.13565389   85.10506293  48.0687406  46.6862333  48.5890735
#> 65  -65.7712511  29.32707902  -14.15647172 -35.8655365 -35.0816217 -36.8046551
#> 70   61.6773024 -23.35409488   15.10886419  39.3393827  38.7196217  40.7791816
#> 75  -47.0754634  24.20349394   -7.59220972 -29.9265975 -29.9667998 -32.0866992
#> 80   42.8165215   4.82076390    1.53813111  26.2495781  26.6336803  28.8995750
#> 85  -24.3487769  -2.93691216   -0.96750858 -16.8699992 -17.6523987 -19.6928288
#> 90    9.6710926   1.24605503    0.42211034   7.5546157   8.3531350   9.6521764
#> 95   -2.2748115  -0.31123505   -0.11037382  -1.9541809  -2.2518481  -2.6861664
#> 100   0.2597073   0.03683747    0.01288714   0.2285256   0.2509391   0.3042474
#>            2030        2035        2040        2045
#> 0    77.4459176  77.4204900  77.5188470  77.6983814
#> 5   -43.5429804 -43.5345068 -43.5819780 -43.6922370
#> 10   69.4908108  69.4665111  69.5605646  69.7436785
#> 15   62.8975960  62.8819987  63.0042277  63.1255567
#> 20  175.4243967 175.3598544 175.6067308 176.0358055
#> 25   53.9474549  53.9155099  53.9903843  54.0801592
#> 30  119.5914528 119.5528521 119.7715458 120.0562085
#> 35    1.0225270   1.0072712   0.9519375   0.9121235
#> 40   79.7379684  79.7274719  79.9051533  80.1456864
#> 45  -25.6296697 -25.6871280 -25.7956559 -25.9708380
#> 50   60.7878345  60.8398404  61.0409161  61.3084604
#> 55  -36.4591281 -36.5987410 -36.8176156 -37.0954370
#> 60   50.9769737  51.1657131  51.4807825  51.8403704
#> 65  -38.8691838 -39.2202149 -39.6397621 -40.0819810
#> 70   43.1850851  43.6969678  44.2773104  44.8632888
#> 75  -34.3995882 -35.1814702 -35.9946429 -36.7846096
#> 80   31.3094023  32.3339309  33.3781480  34.3711942
#> 85  -21.7940080 -22.9599143 -24.1187034 -25.2468632
#> 90   10.9779695  11.8694772  12.7578752  13.6605184
#> 95   -3.1357372  -3.4757375  -3.8169784  -4.1788256
#> 100   0.3588661   0.4034127   0.4472692   0.4953120

# Net migration for females using the time even flow method
mig_res$mig_f

#>              1950          1955          1960          1965          1970
#> 0   -1667.8720158 -1933.7415780 -2001.3487358 -1655.8199772 -1535.4281731
#> 5    1560.3534801  1827.2836870  1906.3952200  1589.1956239  1482.6533712
#> 10  -1526.1621588 -1793.0405757 -1875.3732529 -1567.1984417 -1464.6613036
#> 15   1498.7614534  1765.3676616  1850.0874230  1549.0672578  1449.8795184
#> 20  -1463.9625194 -1730.0290288 -1817.5727683 -1525.3498678 -1430.4469894
#> 25   1423.7169982  1688.8409997  1779.4235385  1497.5099942  1407.6797414
#> 30  -1379.2166786 -1642.8207086 -1736.5512266 -1465.8557333 -1381.6441828
#> 35   1330.8014089  1591.9498040  1688.6342434  1430.2029061  1352.1517092
#> 40  -1277.6543034 -1535.2285712 -1634.4726710 -1389.3103305 -1317.8076085
#> 45   1218.9911076  1471.7821128  1572.7440892  1341.7300154  1277.1764368
#> 50  -1149.8542506 -1395.5926602 -1497.6473210 -1282.6946202 -1225.7147649
#> 55   1063.0468901  1298.3999061  1400.7577598  1205.7347372  1157.5649508
#> 60   -946.8224703 -1166.8757849 -1268.2788570 -1099.7813179 -1062.7167837
#> 65    791.2468907   988.3286136  1086.0702222   952.3405608   929.2125194
#> 70   -597.1725636  -760.2635838  -849.1207834  -756.5776987  -748.7198179
#> 75    384.7177031   502.3888291   574.1068894   523.2694491   528.2393171
#> 80   -196.4580707  -264.3918257  -311.2417411  -292.1390324  -303.0904616
#> 85     72.2498036   100.6239523   122.8901134   119.4262092   128.2844068
#> 90    -17.3624732   -25.1126688   -31.9737425   -32.2189063   -36.0085001
#> 95      2.5263395     3.7842897     5.0167879     5.2359603     6.1002220
#> 100    -0.2086657    -0.3200464    -0.4384098    -0.4703675    -0.5672467
#>              1975         1980         1985         1990         1995
#> 0   -1595.1965611 -1657.574475 -1718.530552 -1786.116495 -1602.013262
#> 5    1548.0902994  1617.278650  1682.802981  1751.901206  1576.560695
#> 10  -1531.6800940 -1603.371588 -1670.137147 -1739.207434 -1566.955052
#> 15   1517.5376814  1591.365996  1659.097514  1728.082387  1558.079754
#> 20  -1498.9984027 -1575.257281 -1644.031338 -1712.954293 -1545.312069
#> 25   1477.3889358  1556.162464  1626.149896  1694.596665  1528.899062
#> 30  -1452.8620655 -1534.179889 -1605.302604 -1672.776143 -1508.318703
#> 35   1424.8856104  1508.682537  1581.150033  1647.110268  1483.554599
#> 40  -1392.0123678 -1478.123792 -1551.802429 -1616.149937 -1453.896424
#> 45   1352.6142265  1440.787949  1515.642736  1578.586817  1418.880461
#> 50  -1302.2546132 -1392.171524 -1468.138893 -1529.853528 -1374.458019
#> 55   1234.9757669  1326.114209  1403.055934  1463.880136  1315.670147
#> 60  -1140.5179222 -1231.904568 -1309.519825 -1369.808191 -1233.166118
#> 65   1006.3228969  1096.213521  1173.300600  1233.209532  1114.479906
#> 70   -822.2738412  -907.321840  -981.564126 -1040.136335  -946.685530
#> 75    592.3609243   666.255158   733.137756   787.466091   724.683291
#> 80   -349.6196813  -403.887004  -455.494897  -499.670862  -467.591407
#> 85    153.5543207   183.382798   213.655146   241.390641   231.309319
#> 90    -45.1052201   -56.015642   -67.889186   -79.581114   -78.568087
#> 95      8.0684159    10.450330    13.251406    16.191770    16.576283
#> 100    -0.7899315    -1.058186    -1.395044    -1.775917    -1.879536
#>             2000         2005         2010         2015       2020        2025
#> 0   -1545.949170 -1738.989371 -1816.815772 -1833.973184 -1878.0643 -1890.11502
#> 5    1525.995860  1721.719536  1803.051661  1822.662475  1876.9583  1889.12787
#> 10  -1517.834683 -1713.994445 -1796.385472 -1817.020307 -1876.3501 -1888.58311
#> 15   1510.207214  1706.643454  1789.931096  1811.375625  1875.4795  1887.81656
#> 20  -1498.757234 -1695.546188 -1780.080691 -1802.762676 -1873.3787 -1885.86689
#> 25   1483.236896  1680.609260  1767.525627  1792.097887  1871.0050  1883.67872
#> 30  -1462.550059 -1660.303746 -1751.072460 -1778.313139 -1867.7028 -1880.57728
#> 35   1437.043417  1635.016327  1730.717017  1761.357875  1863.6846  1876.84786
#> 40  -1406.844847 -1604.602042 -1705.392696 -1739.701099 -1857.4390 -1871.03646
#> 45   1372.312196  1569.311337  1674.414850  1712.261744  1847.5270  1861.82366
#> 50  -1329.601855 -1524.925684 -1633.698618 -1675.126820 -1831.5866 -1846.98995
#> 55   1274.308012  1466.723377  1578.415843  1623.484964  1807.2878  1824.39166
#> 60  -1197.723939 -1385.359632 -1499.165211 -1548.242958 -1770.7526 -1790.41759
#> 65   1088.428785  1268.112995  1382.860306  1436.292928  1715.0856  1738.54719
#> 70   -933.276625 -1099.585094 -1212.756482 -1270.530977 -1628.8455 -1657.71854
#> 75    725.610038   869.496908   975.358839  1035.611668  1494.0611  1530.25531
#> 80   -479.570910  -589.048243  -677.013847  -733.966411 -1281.8552 -1326.13916
#> 85    245.989770   312.766166   371.484681   415.533110   959.9758  1008.26686
#> 90    -87.847992  -117.032411  -144.490515  -168.936507  -551.2452  -590.69751
#> 95     19.750008    27.779566    35.759028    44.203699   195.1147   213.15651
#> 100    -2.401081    -3.537057    -4.710151    -6.119166   -30.9816   -34.16778
#>            2030        2035        2040        2045
#> 0   -1906.17754 -1949.16521 -2013.69489 -2085.43908
#> 5    1905.30964  1948.37566  2012.96792  2084.76838
#> 10  -1904.82730 -1947.93534 -2012.56097 -2084.39106
#> 15   1904.16750  1947.34277  2012.02262  2083.90183
#> 20  -1902.37304 -1945.65978 -2010.42943 -2082.39169
#> 25   1900.38870  1943.80720  2008.68514  2080.75176
#> 30  -1897.50691 -1941.07938 -2006.08336 -2078.27207
#> 35   1894.09254  1937.87550  2003.05196  2075.40649
#> 40  -1888.75077 -1932.85415 -1998.29242 -2070.89748
#> 45   1880.29671  1924.91301  1990.77065  2063.77771
#> 50  -1866.66440 -1912.09552 -1978.61913 -2052.26496
#> 55   1845.91007  1892.58883  1960.13127  2034.75332
#> 60  -1814.70895 -1863.26264 -1932.33424 -2008.41973
#> 65   1766.96940  1818.31315  1889.65998  1967.93217
#> 70  -1692.09452 -1747.42932 -1822.01073 -1903.42069
#> 75   1572.82419  1633.53728  1712.39558  1798.01460
#> 80  -1378.09576 -1444.44315 -1527.36061 -1617.10857
#> 85   1065.54883  1132.84506  1214.38231  1302.97484
#> 90   -638.59154  -692.31777  -756.48230  -827.19556
#> 95    235.62928   260.48889   290.22137   323.63483
#> 100   -38.19169   -42.64172   -47.98895   -54.06898