blend the Sprague upper boundary age estimates into monotonic spline estimates

A simple monotonic spline on the cumulative sum of population counts may return more convincing single age count estimates than the Sprague or other splitting methods. This function blends the given single age population estimates starting at pivotAge.

monoCloseout(popmat, pops, pivotAge = 90, splitfun = sprague,
  OAG = TRUE, ...)

Arguments

popmat
a numeric matrix of population counts in 5-year age groups, with integer-labeled margins (age in rows and year in columns).

pops
optional numeric matrix of single age population counts derived from popmat.

pivotAge
integer (default 90). Age at which to switch to spline-based estimates.

splitfun
optional. The function used to create pops. Default sprague. Could also be grabill, beersModSimple, or any other function that similarly transforms.

OAG
logical (default FALSE). Would we like to re-impute the last element of Value as the open age group?

...
arguments to be optionally passed to splitfun().

Value

numeric matrix of age by year estimates of single-age counts.

Details

The pivotAge must be at least 10 years below the maximum age detected from rownames(popmat), but not lower than 75. In the exact pivotAge, we may either take the Sprague estimates or the spline estimates, depending on which is larger, then the single-age estimates for this 5-year age group are rescaled to sum to the original total in popmat. Higher ages are taken from the spline-based age splits. The spline results are derive from the "monoH.FC" method of splinefun() on the cumulative sum of the original age grouped data. One could use this function to perform the same closeout to Grabill estimates, if these are given via the pops argument. See examples. Note that the Grabill split method mixed with this closeout will not necessarily preserve the annual totals, and this function performs to rescaling. The open age group is preserved (and must be included in popmat).

Examples

closed.out           <- monoCloseout(pop5_mat)
colSums(closed.out) - colSums(pop5_mat)

#> 1950 1951 1952 1953 1954 
#>    0    0    0    0    0 

monoCloseout(pop5_mat, pivotAge = 85)

#>           1950       1951       1952       1953       1954
#> 0   12172.4912 13335.0352 14267.9920 14942.2848 15353.2592
#> 1   11380.4480 12243.3920 12963.2560 13534.8000 13962.3360
#> 2   10717.4480 11325.9760 11865.3520 12345.5680 12775.6240
#> 3   10171.0000 10567.3840 10956.4960 11356.1680 11776.4880
#> 4    9728.6128  9952.2128 10218.9040 10548.1792 10948.2928
#> 5    9377.7952  9465.0592  9634.7920  9903.1808 10274.4032
#> 6    9106.0560  9090.5200  9186.3760  9402.7520  9738.1840
#> 7    8900.9040  8813.1920  8855.8720  9028.4720  9323.0000
#> 8    8749.8480  8617.6720  8625.4960  8761.9200  9012.2160
#> 9    8640.3968  8488.5568  8477.4640  8584.6752  8789.1968
#> 10   8573.6144  8420.4928  8403.9360  8488.4160  8645.7408
#> 11   8550.5648  8408.1264  8397.0720  8464.8208  8573.6464
#> 12   8490.9808  8385.8064  8389.3680  8444.9728  8514.1104
#> 13   8355.2608  8318.0304  8343.1520  8390.2528  8433.6304
#> 14   8171.5792  8219.5440  8270.4720  8312.5376  8340.8720
#> 15   8008.3808  8146.1408  8228.0816  8268.6640  8282.4496
#> 16   7852.7792  8093.3168  8213.0704  8254.8736  8254.3760
#> 17   7694.2592  7991.1728  8133.2864  8179.4336  8176.7760
#> 18   7534.6752  7812.5568  7951.1424  8004.6976  8016.1520
#> 19   7373.9056  7584.8128  7703.4192  7766.3312  7802.2464
#> 20   7210.8208  7362.4544  7464.0976  7537.2688  7598.2816
#> 21   7046.6624  7134.6000  7217.9168  7302.4704  7391.5920
#> 22   6881.6064  6923.5360  6994.1408  7087.2064  7196.2960
#> 23   6715.9664  6745.8640  6814.6448  6911.2544  7025.7680
#> 24   6550.9440  6590.5456  6664.2000  6760.8000  6871.0624
#> 25   6386.5216  6428.8224  6504.9152  6601.9952  6710.1056
#> 26   6221.6160  6265.1040  6343.4240  6441.3024  6547.6080
#> 27   6065.6560  6110.6160  6188.7600  6284.4544  6386.6960
#> 28   5922.9920  5967.6480  6041.3360  6130.8304  6226.8240
#> 29   5789.2144  5832.8096  5899.5648  5980.4176  6068.7664
#> 30   5656.4784  5700.8880  5761.4016  5833.3472  5913.0928
#> 31   5527.4512  5573.4864  5627.2016  5689.0208  5758.7888
#> 32   5393.8752  5442.6944  5492.8656  5548.1648  5610.8208
#> 33   5251.2032  5304.2224  5356.1216  5410.9648  5471.4288
#> 34   5103.9920  5161.7088  5218.4096  5276.5024  5337.8688
#> 35   4958.4544  5021.6048  5083.0608  5144.0304  5205.8992
#> 36   4809.8784  4880.8480  4948.9520  5014.1264  5077.2592
#> 37   4676.3904  4748.0400  4816.9760  4882.3664  4944.7312
#> 38   4566.5184  4627.5360  4687.7600  4746.6064  4804.5792
#> 39   4469.7584  4513.9712  4560.2512  4608.8704  4660.5312
#> 40   4372.0480  4399.6592  4432.5680  4471.6320  4517.5792
#> 41   4282.1632  4288.5776  4304.8448  4332.3632  4371.7584
#> 42   4161.7152  4163.6416  4176.1408  4200.8432  4238.3904
#> 43   3992.0912  4016.3536  4045.7168  4081.5312  4124.6464
#> 44   3793.9824  3855.7680  3913.7296  3968.6304  4021.6256
#> 45   3600.3712  3697.6000  3782.5808  3855.5168  3917.8320
#> 46   3397.0752  3536.5024  3653.5968  3747.8736  3821.0560
#> 47   3237.6592  3390.3504  3517.8608  3619.4256  3696.8000
#> 48   3149.6752  3268.9184  3371.7568  3457.8256  3528.4960
#> 49   3104.2192  3162.6288  3220.2048  3277.3584  3334.8160
#> 50   3052.7728  3054.4800  3069.8640  3100.0864  3145.2992
#> 51   3014.5648  2950.6928  2917.1504  2916.1136  2947.1968
#> 52   2923.2688  2832.3648  2778.7744  2764.7536  2789.4208
#> 53   2744.0448  2689.2848  2663.0064  2666.0416  2697.1008
#> 54   2513.3488  2532.1776  2561.2048  2599.0048  2643.9824
#> 55   2294.2640  2379.8880  2459.1184  2528.3760  2585.7136
#> 56   2064.3184  2226.1376  2362.7904  2468.0976  2539.6032
#> 57   1897.9424  2091.2976  2251.0384  2370.2496  2446.5632
#> 58   1835.3744  1986.6016  2112.9104  2209.6416  2275.4272
#> 59   1836.1008  1901.0752  1960.1424  2012.6352  2058.6928
#> 60   1826.9504  1814.3872  1812.7776  1825.1984  1853.2624
#> 61   1831.6544  1732.8560  1663.6336  1631.4368  1639.6416
#> 62   1779.1824  1638.9920  1537.3616  1483.4048  1481.0976
#> 63   1629.7344  1522.7680  1447.2336  1409.4928  1412.2496
#> 64   1422.4784  1393.9968  1379.9936  1381.4672  1398.7488
#> 65   1231.8448  1272.1248  1312.1264  1347.9424  1377.6976
#> 66   1037.5024  1152.7904  1251.9504  1325.4816  1368.9600
#> 67    891.8304  1045.2544  1173.8384  1265.7136  1315.4080
#> 68    826.6464   955.9344  1063.8304  1141.4096  1184.6160
#> 69    813.1760   879.8960   936.2544   979.4528  1008.3184
#> 70    792.0960   805.2080   817.1408   830.5920   846.9968
#> 71    776.7936   733.5616   698.5744   680.6416   684.1408
#> 72    745.8176   667.7616   604.2624   567.0256   561.7328
#> 73    683.0176   607.7856   547.7584   512.1856   505.3408
#> 74    601.2752   552.6832   516.2640   495.5552   491.7888
#> 75    529.6992   501.8128   484.1152   475.1312   473.3952
#> 76    462.5600   455.6464   457.2704   460.5200   460.9712
#> 77    401.2800   409.4624   424.0144   435.1200   436.8832
#> 78    349.5120   360.9504   376.1984   386.8320   387.9712
#> 79    304.9488   312.1280   320.4016   325.3968   324.7792
#> 80    261.4336   267.2976   271.3008   272.5056   270.7904
#> 81    219.9616   225.5680   225.8992   223.6416   221.0432
#> 82    183.1696   187.9200   185.6352   181.2096   178.2752
#> 83    151.6816   155.2640   152.6752   148.5216   146.1152
#> 84    124.7536   126.9504   125.4896   123.1216   121.7760
#> 85    108.4960   108.2320   107.4960   106.8960   106.7840
#> 86     77.9920    75.7040    74.7920    74.9520    75.7280
#> 87     56.3440    53.0880    52.1440    52.7040    53.9360
#> 88     43.5520    40.3840    39.5520    40.1520    41.4080
#> 89     39.6160    37.5920    37.0160    37.2960    38.1440
#> 90     32.7520    32.0640    31.8240    31.7520    32.1600
#> 91     19.9840    19.9280    19.9280    19.7440    20.0000
#> 92     11.6080    11.8960    12.0160    11.7680    11.9200
#> 93      7.6240     7.9680     8.0880     7.8240     7.9200
#> 94      8.0320     8.1440     8.1440     7.9120     8.0000
#> 95      7.4320     7.2960     7.1600     6.9600     7.0240
#> 96      3.9040     3.7120     3.5200     3.3200     3.3280
#> 97      1.8880     1.6640     1.4400     1.2400     1.2160
#> 98      1.3840     1.1520     0.9200     0.7200     0.6880
#> 99      2.3920     2.1760     1.9600     1.7600     1.7440
#> 100     0.0000     1.0000     1.0000     1.0000     2.0000

# giving a different single-age split to close out this way:
popg                 <- grabill(pop5_mat)
grabill.closed.out   <- monoCloseout(pop5_mat, popg)
# totals not necessarily preserved if mixed w Grabill
# I wouldn't recommend a rescale of the total, since the
# only part we mess with here is the old age section. Ergo,
# one may wish to instead rescale results colSums() of
# popg at age pivotAge and higher.
colSums(grabill.closed.out) - colSums(pop5_mat)

#>      1950      1951      1952      1953      1954 
#> -22.01899 -22.33631 -22.11829 -21.81176 -21.41604 

# also works on an age-labelled vector of data
popvec               <- pop5_mat[,1]
closed.vec           <- monoCloseout(popvec)
# let's compare this one with sprague()
simple.vec           <- sprague(popvec)
# and with a simple monotonic spline
mono.vec             <- splitMono(popvec)
if (FALSE) {
plot(85:100,simple.vec[86:101], type = 'l', main = "In this case sprague() is the smoothest")
lines(85:100,closed.vec[86:101], col = "red", lwd = 2)
lines(85:100,mono.vec[86:101], col = "blue", lty = 2)
legend("topright",lty=c(1,2,2), col = c("black","red","blue"),lwd = c(1,2,1),
                legend = c("sprague()","monoCloseout()", "splitMono()"))
}