wrapper to invoke PAS or UN ax methods given qx or mx

Given either mx or qx, call either the axUN() or axPAS() functions.

mxorqx2ax(nMx, nqx, axmethod = c("pas", "un")[1], Age, AgeInt,
  IMR = NA, Sex, region, OAG = TRUE, mod = TRUE,
  extrapLaw = c("Kannisto", "Kannisto_Makeham", "Makeham", "Gompertz",
  "GGompertz", "Beard", "Beard_Makeham", "Quadratic")[1],
  extrapFrom = max(Age), extrapFit = Age[Age >= 60], ...)

Arguments

nMx
numeric. Vector of mortality rates in abridged age classes.

nqx
numeric. Vector of conditional death probabilities in abridged age classes.

axmethod
character. Either "pas" or "un".

Age
integer. A vector of ages of the lower integer bound of the age classes.

AgeInt
integer. Vector of age class widths. Default inferAgeIntAbr(Age = Age).

IMR
numeric. Infant mortality rate q0, in case available and nqx is not specified. Default NA.

Sex
character. Either male "m", female "f", or both "b".

region
character. North, East, South, or West: code"n", code"e", code"s", code"w". Default code"w".

OAG
logical. Whether or not the last element of nMx is the open age group Default TRUE.

mod
logical. If "un" specified for axmethod, whether or not to use Nan Li's modification for ages 5-14. Default TRUE.

extrapLaw
character. If extrapolating, which parametric mortality law should be invoked? Options include "Kannisto", "Kannisto_Makeham", "Makeham", "Gompertz", "GGompertz", "Beard", "Beard_Makeham", "Quadratic". Default "Kannisto". See details.

extrapFrom
integer. Age from which to impute extrapolated mortality.

extrapFit
integer vector. Ages to include in model fitting. Defaults to all ages > =60.

...
Other arguments to be passed on to the MortalityLaw function.

Value

nax average contribution to exposure of those dying in the interval.

References

Greville TN (1977). “Short methods of constructing abridged life tables.” In Mathematical Demography, 53--60. Springer. United Nations (1982). Model Life Tables for Developing Countries. United Nations Department of International Economic and Social Affairs. Arriaga EE, Johnson PD, Jamison E (1994). Population analysis with microcomputers, volume 1. Bureau of the Census. United Nations (1988). Mortpak- The United Nations Software Package for Mortality Measurement. United Nations Department of International Economic and Social Affairs. United Nations (1983). Manual X: Indirect Techniques for Demographic Estimation, number 81. United Nations Department of International Economic and Social Affairs, New York. United States Census Bureau (2017). “Population Analysis System (PAS) Software.” https://www.census.gov/data/software/pas.html, https://www.census.gov/data/software/pas.html.