convert strings to concatenation of lower case alphabet

all characters are converteusd to lower case, and all non-alphabet characters are removed. This is useful for reducing checks on names of user-specified strings, like method arguments. For example, "Carrier-Farrag", "Carrier Farrag", or "carrier_farrag" all get converted to "carrierfarrag".

simplify.text(string)

Arguments

string
character string

Value

character string

Examples

simplify.text(c("carrier_farrag","CarrierFarrag","Carrier-Farrag"))

#> [1] "carrierfarrag" "carrierfarrag" "carrierfarrag"