Split age groups using a monotonic spline.

Take the cumulative sum of Value and then run a monotonic spline through it. The first differences split back single-age estimates of Value. Optionally keep the open age group untouched.

splitMono(Value, AgeInt, Age, OAG = FALSE)

Arguments

Value
numeric. Vector of (presumably) counts in grouped ages.

AgeInt
integer or numeric. Vector of age intervals.

Age
numeric. Vector of ages corresponding to the lower integer bound of the age range.

OAG
boolean. Argument that determines whether the final age group (assumed open ended) is kept as it is or has the same length as the rest of the age groups. Default is FALSE, i.e. use the same length for the final age group.

Value

Numeric. vector of single smoothed age counts.

Details

The "monoH.FC" method of stats::splinefun() is used to fit the spline because 1) it passes exactly through the points, 2) it is monotonic and therefore guarantees positive counts, and 3) it seems to be a bit less wiggly (lower average first differences of split counts). Single-age data is returned as-is. If you want to use this function as a smoother you first need to group to non-single ages.

References

Fritsch FN, Carlson RE (1980). “Monotone piecewise cubic interpolation.” SIAM Journal on Numerical Analysis, 17(2), 238--246.

Examples

Value                <- structure(c(88623, 90842, 93439, 96325, 99281, 102051, 104351,
                         106555, 109170, 112188, 113582, 112614, 108904, 102622, 95867,
                         80874, 60196, 37523, 17927, 5642, 1110), .Names = c("0", "5",
                         "10", "15", "20", "25", "30", "35", "40", "45", "50", "55", "60",
                         "65", "70", "75", "80", "85", "90", "95", "100"))

# overwrite open age group with a single age estimate for that age
# (doesn't extrapolate)
splitMono(Value)

#>         0         1         2         3         4         5         6         7 
#> 17689.096 17653.592 17671.344 17742.352 17866.616 17984.832 18067.544 18159.328 
#>         8         9        10        11        12        13        14        15 
#> 18260.184 18370.112 18475.416 18574.672 18680.864 18793.992 18914.056 19033.000 
#>        16        17        18        19        20        21        22        23 
#> 19147.320 19263.320 19381.000 19500.360 19622.696 19743.912 19860.664 19972.952 
#>        24        25        26        27        28        29        30        31 
#> 20080.776 20196.120 20314.440 20421.480 20517.240 20601.720 20687.736 20781.272 
#>        32        33        34        35        36        37        38        39 
#> 20872.504 20961.432 21048.056 21128.104 21209.688 21301.136 21402.448 21513.624 
#>        40        41        42        43        44        45        46        47 
#> 21618.352 21716.504 21824.328 21941.824 22068.992 22222.144 22368.848 22476.576 
#>        48        49        50        51        52        53        54        55 
#> 22545.328 22575.104 22642.672 22736.224 22773.088 22753.264 22676.752 22644.112 
#>        56        57        58        59        60        61        62        63 
#> 22649.264 22588.608 22462.144 22269.872 22118.752 22011.504 21842.528 21611.824 
#>        64        65        66        67        68        69        70        71 
#> 21319.392 21034.528 20790.816 20535.752 20269.336 19991.568 19845.608 19707.216 
#>        72        73        74        75        76        77        78        79 
#> 19371.112 18837.296 18105.768 17465.200 16956.440 16311.240 15529.600 14611.520 
#>        80        81        82        83        84        85        86        87 
#> 13725.360 12930.160 12087.080 11196.120 10257.280  9269.208  8313.056  7430.752 
#>        88        89        90        91        92        93        94        95 
#>  6622.296  5887.688  5036.104  4135.288  3409.936  2860.048  2485.624  1987.152 
#>        96        97        98        99       100       101       102       103 
#>  1371.704   942.328   699.024   641.792   512.048   258.256   113.232    76.976 
#>       104 
#>   149.488 

# or respect open age group
splitMono(Value, OAG = TRUE)

#>         0         1         2         3         4         5         6         7 
#> 17689.096 17653.592 17671.344 17742.352 17866.616 17984.832 18067.544 18159.328 
#>         8         9        10        11        12        13        14        15 
#> 18260.184 18370.112 18475.416 18574.672 18680.864 18793.992 18914.056 19033.000 
#>        16        17        18        19        20        21        22        23 
#> 19147.320 19263.320 19381.000 19500.360 19622.696 19743.912 19860.664 19972.952 
#>        24        25        26        27        28        29        30        31 
#> 20080.776 20196.120 20314.440 20421.480 20517.240 20601.720 20687.736 20781.272 
#>        32        33        34        35        36        37        38        39 
#> 20872.504 20961.432 21048.056 21128.104 21209.688 21301.136 21402.448 21513.624 
#>        40        41        42        43        44        45        46        47 
#> 21618.352 21716.504 21824.328 21941.824 22068.992 22222.144 22368.848 22476.576 
#>        48        49        50        51        52        53        54        55 
#> 22545.328 22575.104 22642.672 22736.224 22773.088 22753.264 22676.752 22644.112 
#>        56        57        58        59        60        61        62        63 
#> 22649.264 22588.608 22462.144 22269.872 22118.752 22011.504 21842.528 21611.824 
#>        64        65        66        67        68        69        70        71 
#> 21319.392 21034.528 20790.816 20535.752 20269.336 19991.568 19845.608 19707.216 
#>        72        73        74        75        76        77        78        79 
#> 19371.112 18837.296 18105.768 17465.200 16956.440 16311.240 15529.600 14611.520 
#>        80        81        82        83        84        85        86        87 
#> 13725.360 12930.160 12087.080 11196.120 10257.280  9269.208  8313.056  7430.752 
#>        88        89        90        91        92        93        94        95 
#>  6622.296  5887.688  5036.104  4135.288  3409.936  2860.048  2485.624  1914.640 
#>        96        97        98        99       100 
#>  1226.680   833.560   735.280   931.840  1110.000 

# Also accepts single ages:
Value                <- structure(pop1m_ind, .Names = 0:100)

                if (FALSE) {
        ages                <- seq(0,100,5)
 plot(splitMono(Value),xlab = 'Age', ylab = 'Counts', type = 'l',main = 'Ungrouped counts')
                }