The basic Sprague age-splitting method.

This method is used to interpolate counts based on the Sprague formula. It is based on the first stage of the Sprague R script prepared by Thomas Buettner and Patrick Gerland, itself based on the description in Siegel and Swanson, 2004, p. 727.

sprague(popmat, Age = as.integer(rownames(as.matrix(popmat))),
  OAG = TRUE)

Arguments

popmat
numeric. A matrix of population counts in 5-year age groups with integer-labeled margins (age in rows and year in columns).

Age
integer. Vector of lower age bound of age groups. Detected from row names of popmat if missing.

OAG
logical. Whether or not the final age group open. Default TRUE.

Value

An age-period matrix of split population counts with the same number of columns as popmat, and single ages in rows.

Details

Ages should refer to lower age bounds, ending in the open age group in the last row (not a closed terminal age). Dimension labelling is necessary. There must be at least six age groups (including the open group). One year of data will work as well, as long as it's given as or coercible to a single-column matrix. This method may produce negative values, most likely in the youngest or oldest ages.

If the highest age does not end in a 0 or 5, and OAG == TRUE, then the open age will be grouped down to the next highest age ending in 0 or 5. If the highest age does not end in a 0 or 5, and OAG == FALSE, then results extend to single ages covering the entire 5-year age group.

References

Sprague TB (1880). “Explanation of a new formula for interpolation.” Journal of the Institute of Actuaries, 22(4), 270--285. Shryock HS, Siegel JS, Larmon EA (1973). The methods and materials of demography. US Bureau of the Census. Siegel Jacob S, Swanson David A (eds.) (2004). The Methods and Materials of Demography, 2 edition. Elsevier Academic Press, California, USA, San Diego, USA.

Examples

head(pop5_mat) # this is the entire matrix

#>     1950  1951  1952  1953  1954
#> 0  54170 57424 60272 62727 64816
#> 5  44775 44475 44780 45681 47137
#> 10 42142 41752 41804 42101 42508
#> 15 38464 39628 40229 40474 40532
#> 20 34406 34757 35155 35599 36083
#> 25 30386 30605 30978 31439 31940

# the last value is an open age group, preserve as such:
p1                            <- sprague(pop5_mat, OAG = TRUE)
head(p1); tail(p1)

#>        1950      1951      1952      1953     1954
#> 0 12172.491 13335.035 14267.992 14942.285 15353.26
#> 1 11380.448 12243.392 12963.256 13534.800 13962.34
#> 2 10717.448 11325.976 11865.352 12345.568 12775.62
#> 3 10171.000 10567.384 10956.496 11356.168 11776.49
#> 4  9728.613  9952.213 10218.904 10548.179 10948.29
#> 5  9377.795  9465.059  9634.792  9903.181 10274.40

#>       1950    1951    1952    1953    1954
#> 95  5.8576  7.1072  7.3408  6.7776  6.4416
#> 96  4.5040  5.5280  5.6640  5.1440  4.8160
#> 97  3.4240  3.7680  3.6960  3.3440  3.1680
#> 98  2.3200  1.4400  1.0320  1.0000  1.1520
#> 99  0.8944 -1.8432 -2.7328 -2.2656 -1.5776
#> 100 0.0000  1.0000  1.0000  1.0000  2.0000

colSums(p1) - colSums(pop5_mat)

#> 1950 1951 1952 1953 1954 
#>    0    0    0    0    0 

# another case, starting with single ages
Age                           <- 0:100
# notice how this particular case produces a negative value in the last age
# before OAG:
pops                          <- sprague(popmat = pop1m_ind, Age = Age, OAG = TRUE)

if (FALSE) {
yrs                           <- as.character(1950:1954)
matplot(0:100,p1, type = 'l', col = 5:9,lty = 1,xlab = 'Age',ylab = 'Counts',
main= 'Ungrouped counts')
legend('topright',
title = 'Year',
lty=1,
col=5:9,
legend = yrs
)
}