Create the Sprague coefficient matrix.

The resulting coefficient matrix is based on the number of rows in popmat where is assumed that each row of data is a 5-year age group. The final row may be an open or closed age group, as indicated by the OAG argument.

spragueExpand(popmat, OAG = TRUE)

Arguments

popmat
numeric. A matrix of population counts in 5-year age groups with integer-labeled margins (age in rows and year in columns).

OAG
logical. Whether or not the final age group open. Default TRUE.

Details

The popmat matrix is really just a placeholder in this case. This function is a utility called by the Sprague family of functions, where it is most convenient to just pass in the same matrix being used in those calculations to determine the layout of the coefficient matrix.

References

Sprague TB (1880). “Explanation of a new formula for interpolation.” Journal of the Institute of Actuaries, 22(4), 270--285. Shryock HS, Siegel JS, Larmon EA (1973). The methods and materials of demography. US Bureau of the Census. Siegel Jacob S, Swanson David A (eds.) (2004). The Methods and Materials of Demography, 2 edition. Elsevier Academic Press, California, USA, San Diego, USA.

Examples

coefsOA                       <- spragueExpand(pop5_mat, TRUE)
coefsclosed                   <- spragueExpand(pop5_mat, FALSE)
dim(coefsOA)

#> [1] 101  21

dim(coefsclosed)

#> [1] 105  21