Determine the proportion of a year passed as of a particular date.

The fraction returned by this is used e.g. for intercensal estimates.

ypart(Year, Month, Day, detect.mid.year = TRUE,
  detect.start.end = TRUE)

Arguments

Year
string or integer. 4-digit year.

Month
string or integer. Month digits, 1 or 2 characters.

Day
string or integer. Day of month digits, 1 or 2 characters.

detect.mid.year
logical. If TRUE, June 30 or July 1 will always return .5.

detect.start.end
logical. Whether or not Jan 1 always be 0 and Dec 31 always be 1. Default TRUE.

Details

Code inherited from HMD, slightly modified to remove matlab legacy bit.

Examples

ypart(2001,2,14) # general use

#> [1] 0.1150685

ypart(2001,6,30) # mid year options, default detection

#> [1] 0.5

ypart(2001,7,1)  # also

#> [1] 0.5

ypart(2000,6,30) # no change for leap year, still detected as mid year

#> [1] 0.5

ypart(2000,6,30, FALSE) # exact measure

#> [1] 0.4945355

ypart(2000,7,1, FALSE)  # July 1 leap year

#> [1] 0.5

ypart(2001,7,1, FALSE)  # July 1 not leap year

#> [1] 0.4986301

ypart(2002,12,31, detect.start.end = FALSE) # assumes end of day by default.

#> [1] 1

ypart(2002,1,1, detect.start.end = FALSE) # end of day year fraction

#> [1] 0.002739726

ypart(2002,1,1, detect.start.end = TRUE)  # assume very begining of year

#> [1] 0