Calculate Jdanov's old-age heaping index

This is a slightly more flexible implementation of Jdanov's formula, with defaults set to match his parameters. The numerator is the sum of (death counts) in ages 95, 100, and 105. The denominator consists in the sum of the 5-year age groups centered around each of the numerator ages. It probably only makes sense to use this with the default values, however. Used with a single age in the numerator, it is almost the same as Noumbissi(), except here we pick out particular ages, whereas Noumbissi picks out terminal digits.

check_heaping_jdanov(Value, Age, Agei = seq(95, 105, by = 5))

Arguments

Value
numeric. A vector of demographic counts by single age.

Age
integer. A vector of ages corresponding to the lower integer bound of the counts.

Agei
integer. A vector of ages to put in the numerator, default c(95,100,105).

Value

The value of the index.

References

Jdanov D (2008). “Beyond the Kannisto-Thatcher database on old age mortality: An assessment of data quality at advanced ages.” MPIDR Working Paper WP-2008-013.

Examples

Value <-c(8904, 592, 354, 299, 292, 249, 222, 216, 181, 169, 151, 167,
        170, 196, 249, 290, 425, 574, 671, 724, 675, 754, 738, 695, 597,
        498, 522, 479, 482, 478, 558, 582, 620, 606, 676, 768, 862, 952,
        1078, 1215, 1215, 1357, 1470, 1605, 1723, 1782, 1922, 2066, 2364,
        2561, 2476, 1674, 1664, 1616, 1808, 3080, 3871, 4166, 4374, 4707,
        5324, 5678, 6256, 6382, 6823, 7061, 7344, 8149, 8439, 8308, 8482,
        8413, 8157, 7945, 7503, 7164, 7289, 7016, 6753, 6906, 6797, 6624,
        6416, 5811, 5359, 4824, 4277, 3728, 3136, 2524, 2109, 1657, 1235,
        924, 667, 465, 287, 189, 125, 99, 80, 24, 10, 7, 3, 1, 0, 1,
        1, 0, 0)
Age <- 0:110
check_heaping_jdanov(Value, Age, Agei = c(95,100,105))

#> [1] 94.72588