blend the Sprague upper boundary age estimates into monotonic spline estimates

A simple monotonic spline on the cumulative sum of population counts may return more convincing single age count estimates than the Sprague or other splitting methods. This function blends the given single age population estimates starting at pivotAge.

graduate_mono_closeout(
  Value,
  Age,
  pops,
  pivotAge = 90,
  splitfun = graduate_sprague,
  OAG = TRUE,
  ...
)

Arguments

Value
numeric vector, presumably counts in grouped ages

Age
integer vector, lower bounds of age groups

pops
optional numeric vector of single age population counts derived from Value.

pivotAge
integer (default 90). Age at which to switch to spline-based estimates.

splitfun
optional. The function used to create pops. Default graduate_sprague. Could also be graduate_grabill, or any other function that similarly transforms.

OAG
logical (default FALSE). Would we like to re-impute the last element of Value as the open age group?

...
arguments to be optionally passed to splitfun().

Value

numeric matrix of age by year estimates of single-age counts.

Details

The pivotAge must be at least 10 years below the maximum age detected from rownames(popmat), but not lower than 75. In the exact pivotAge, we may either take the Sprague estimates or the spline estimates, depending on which is larger, then the single-age estimates for this 5-year age group are rescaled to sum to the original total in Value. Higher ages are taken from the spline-based age splits. The spline results are derive from the "monoH.FC" method of splinefun() on the cumulative sum of the original age grouped data. One could use this function to perform the same closeout to Grabill estimates, if these are given via the pops argument. See examples. Note that the Grabill split method mixed with this closeout will not necessarily preserve the annual totals, and this function performs to rescaling. The open age group is preserved (and must be included in Value).

Examples

a5 <- as.integer(rownames(pop5_mat))
popvec               <- pop5_mat[,1]
closed.out           <- graduate_mono_closeout(Value = popvec, Age = a5, OAG = TRUE)
sum(closed.out) - sum(popvec)

#> [1] 0

graduate_mono_closeout(Value = popvec, pivotAge = 85, Age = a5, OAG = TRUE)

#>          0          1          2          3          4          5          6 
#> 12172.4912 11380.4480 10717.4480 10171.0000  9728.6128  9377.7952  9106.0560 
#>          7          8          9         10         11         12         13 
#>  8900.9040  8749.8480  8640.3968  8573.6144  8550.5648  8490.9808  8355.2608 
#>         14         15         16         17         18         19         20 
#>  8171.5792  8008.3808  7852.7792  7694.2592  7534.6752  7373.9056  7210.8208 
#>         21         22         23         24         25         26         27 
#>  7046.6624  6881.6064  6715.9664  6550.9440  6386.5216  6221.6160  6065.6560 
#>         28         29         30         31         32         33         34 
#>  5922.9920  5789.2144  5656.4784  5527.4512  5393.8752  5251.2032  5103.9920 
#>         35         36         37         38         39         40         41 
#>  4958.4544  4809.8784  4676.3904  4566.5184  4469.7584  4372.0480  4282.1632 
#>         42         43         44         45         46         47         48 
#>  4161.7152  3992.0912  3793.9824  3600.3712  3397.0752  3237.6592  3149.6752 
#>         49         50         51         52         53         54         55 
#>  3104.2192  3052.7728  3014.5648  2923.2688  2744.0448  2513.3488  2294.2640 
#>         56         57         58         59         60         61         62 
#>  2064.3184  1897.9424  1835.3744  1836.1008  1826.9504  1831.6544  1779.1824 
#>         63         64         65         66         67         68         69 
#>  1629.7344  1422.4784  1231.8448  1037.5024   891.8304   826.6464   813.1760 
#>         70         71         72         73         74         75         76 
#>   792.0960   776.7936   745.8176   683.0176   601.2752   529.6992   462.5600 
#>         77         78         79         80         81         82         83 
#>   401.2800   349.5120   304.9488   261.4336   219.9616   183.1696   151.6816 
#>         84         85         86         87         88         89         90 
#>   124.7536   108.4960    77.9920    56.3440    43.5520    39.6160    32.7520 
#>         91         92         93         94         95         96         97 
#>    19.9840    11.6080     7.6240     8.0320     7.4320     3.9040     1.8880 
#>         98         99        100 
#>     1.3840     2.3920     0.0000 

# giving a different single-age split to close out this way:
popg                 <- graduate_grabill(Value = popvec, Age = a5, OAG = TRUE)
grabill.closed.out   <- graduate_mono_closeout(Value = popvec, Age = a5, pops = popg)
# totals not necessarily preserved if mixed w Grabill
# I wouldn't recommend a rescale of the total, since the
# only part we mess with here is the old age section. Ergo,
# one may wish to instead rescale results colSums() of
# popg at age pivotAge and higher.
sum(grabill.closed.out) - sum(popvec)

#> [1] -22.01899

# also works on an age-labelled vector of data
closed.vec           <- graduate_mono_closeout(popvec, Age = a5, OAG = TRUE)
# let's compare this one with sprague()
simple.vec           <- graduate_sprague(popvec, Age = a5, OAG = TRUE)
# and with a simple monotonic spline
mono.vec             <- graduate_mono(popvec, Age = a5, OAG = TRUE)
if (FALSE) {
plot(85:100,simple.vec[86:101], type = 'l',
 main = "In this case graduate_sprague() is the smoothest")
lines(85:100,closed.vec[86:101], col = "red", lwd = 2)
lines(85:100,mono.vec[86:101], col = "blue", lty = 2)
legend("topright",lty=c(1,2,2), col = c("black","red","blue"),lwd = c(1,2,1),
                legend = c("graduate_sprague()","monoCloseout()", "graduate_mono()"))
}