Estimate Wilmoth Model Life Table

Construct model life tables based on the Log-Quadratic (wilmoth) estimates with various choices of 2 input parameters: q0_5, q0_1, q15_45, q15_35 and e0. There are 8 possible combinations (see examples below).

lt_model_lq(Sex = c("b", "f", "m"), fitted_logquad = NULL,
  q0_5 = NULL, q0_1 = NULL, q15_45 = NULL, q15_35 = NULL,
  e0 = NULL, radix = 1e+05, tol = 1e-09, maxit = 200, ...)

Arguments

Sex
Choose the sex of the population. This choice defines the use of a corresponding Log-Quadratic (wilmoth) model fitted for the whole Human Mortality Database (as of Dec 2019, there are 968 life tables for each sex). The following options are available:

"b" -- Both sex;

"f" -- Females;

"m" -- Males.

fitted_logquad
Optional, defaults to NULL. An object of class wilmoth. If full HMD is not enough, one can fit a Log-Quadratic (https://github.com/mpascariu/MortalityEstimate) model based on any other collection of life tables;

q0_5
5q0. The probability that a new-born will die during the subsequent 5 years;

q0_1
1q0. The probability that a life aged 0 will die during the following year;

q15_45
45q15. The probability that a life aged 15 will die during the subsequent 45 years;

q15_35
35q15. The probability that a life aged 15 will die during the subsequent 35 years;

e0
Life expectancy at birth;

radix
Life table radix. Default: 10^5;

tol
Tolerance level for convergence. The tolerance level, is relevant for case 7 and 8 (e0 and 45q15 or 35q15 are known);

maxit
Maximum number of iterations allowed. Default: 100;

...
Additional arguments affecting the predictions produced.

Value

The output is of class lt_model_lq with the components:

lt

Life table matching given inputs

values

Associated values of q0_5, q0_1, q15_45, q15_35 and e0.

Details

Due to limitations of the R language the notation for probability of dying nqx is written qx_n, where x and n are integers. For example 45q15 is represented as q45_15.

Note

This function is ported from MortalityEstimate::wilmothLT experimental package by Marius Pascariu. The package is no longe maintained. The latest version can be found here: https://github.com/mpascariu/MortalityEstimate

Examples

# Build life tables with various choices of 2 input parameters

# case 1: Using 5q0 and e0
L1 <- lt_model_lq(Sex = "b", q0_5 = 0.05, e0 = 65)
L1

#> $lt
#>     Age AgeInt          nMx       nAx         nqx           lx          ndx
#> 0     0      1 0.0410512359 0.1565727 0.039677454 1.000000e+05 3.967745e+03
#> 1     1      4 0.0027005846 1.5334518 0.010730859 9.603225e+04 1.030509e+03
#> 5     5      5 0.0011222400 2.5000000 0.005595502 9.500175e+04 5.315824e+02
#> 10   10      5 0.0008951101 2.5000000 0.004465558 9.447016e+04 4.218620e+02
#> 15   15      5 0.0017662507 2.5000000 0.008792430 9.404830e+04 8.269131e+02
#> 20   20      5 0.0024174684 2.5000000 0.012014729 9.322139e+04 1.120030e+03
#> 25   25      5 0.0025631985 2.5000000 0.012734391 9.210136e+04 1.172855e+03
#> 30   30      5 0.0029668727 2.5000000 0.014725145 9.092850e+04 1.338935e+03
#> 35   35      5 0.0038311611 2.5000000 0.018974074 8.958957e+04 1.699879e+03
#> 40   40      5 0.0052115491 2.5000000 0.025722609 8.788969e+04 2.260752e+03
#> 45   45      5 0.0074262211 2.5000000 0.036454311 8.562894e+04 3.121544e+03
#> 50   50      5 0.0108422621 2.5000000 0.052780656 8.250739e+04 4.354794e+03
#> 55   55      5 0.0159230416 2.5000000 0.076567249 7.815260e+04 5.983930e+03
#> 60   60      5 0.0242979520 2.5000000 0.114532497 7.216867e+04 8.265658e+03
#> 65   65      5 0.0373089177 2.5000000 0.170629577 6.390301e+04 1.090374e+04
#> 70   70      5 0.0590075983 2.5000000 0.257109462 5.299927e+04 1.362661e+04
#> 75   75      5 0.0932173019 2.5000000 0.377996885 3.937265e+04 1.488274e+04
#> 80   80      5 0.1439053113 2.5000000 0.529155749 2.448991e+04 1.295898e+04
#> 85   85      5 0.2168351663 2.5000000 0.703057083 1.153094e+04 8.106906e+03
#> 90   90      5 0.3080513346 2.5000000 0.870138419 3.424030e+03 2.979380e+03
#> 95   95      5 0.4244002008 1.6757931 0.880208315 4.446499e+02 3.913845e+02
#> 100 100      5 0.5495844154 1.4773291 0.935939164 5.326536e+01 4.985314e+01
#> 105 105      5 0.6700558318 1.3106654 0.964925439 3.412223e+00 3.292541e+00
#> 110 110     NA 0.7610553490 1.3137601 1.000000000 1.196822e-01 1.196822e-01
#>              nLx         Sx           Tx        ex
#> 0   9.665350e+04 0.95648143 6.500007e+06 65.000066
#> 1   3.815872e+05 0.99046309 6.403353e+06 66.679192
#> 5   4.736798e+05 0.99496789 6.021766e+06 63.385845
#> 10  4.712962e+05 0.99337585 5.548086e+06 58.728449
#> 15  4.681742e+05 0.98960354 5.076790e+06 53.980666
#> 20  4.633069e+05 0.98762762 4.608616e+06 49.437322
#> 25  4.575747e+05 0.98627661 4.145309e+06 45.008119
#> 30  4.512952e+05 0.98316615 3.687734e+06 40.556416
#> 35  4.436981e+05 0.97768398 3.236439e+06 36.125177
#> 40  4.337966e+05 0.96898145 2.792741e+06 31.775524
#> 45  4.203408e+05 0.95553407 2.358944e+06 27.548448
#> 50  4.016500e+05 0.93564842 1.938604e+06 23.496119
#> 55  3.758032e+05 0.90520578 1.536954e+06 19.666058
#> 60  3.401792e+05 0.85912277 1.161150e+06 16.089397
#> 65  2.922557e+05 0.79016357 8.209711e+05 12.847143
#> 70  2.309298e+05 0.69136342 5.287154e+05  9.975901
#> 75  1.596564e+05 0.56403696 2.977856e+05  7.563260
#> 80  9.005212e+04 0.41517524 1.381292e+05  5.640249
#> 85  3.738741e+04 0.25868864 4.807710e+04  4.169401
#> 90  9.671698e+03 0.09535102 1.068969e+04  3.121961
#> 95  9.222063e+02 0.09836260 1.017988e+03  2.289415
#> 100 9.071061e+01 0.05417041 9.578167e+01  1.798198
#> 105 4.913831e+00 0.03100606 5.071065e+00  1.486147
#> 110 1.572338e-01 0.00000000 1.572338e-01  1.313760
#> 
#> $values
#>          k       q0_1 q0_5    q15_35    q15_45       e0
#>  0.5530788 0.03967745 0.05 0.1227126 0.2326425 65.00007
#> 
#> attr(,"class")
#> [1] "lt_model_lq"

ls(L1)

#> [1] "lt"     "values"

L1f <- lt_model_lq(Sex = "f", q0_5 = 0.05, e0 = 65)
L1m <- lt_model_lq(Sex = "m", q0_5 = 0.05, e0 = 65)

# case 2: Using 5q0 and 45q15
L2 <- lt_model_lq(Sex = "b", q0_5 = 0.05, q15_45 = 0.2)

# case 3: Using 5q0 and 35q15
L3 <- lt_model_lq(Sex = "b", q0_5 = 0.05, q15_35 = 0.125)

# case 4: Using 1q0 and e0
L4 <- lt_model_lq(Sex = "b", q0_1 = 0.01, e0 = 65)

# case 5: Using 1q0 and 45q15
L5 <- lt_model_lq(Sex = "b", q0_1 = 0.05, q15_45 = 0.2)

#> Error in lt_model_lq(fitted_logquad = fitted_logquad, q0_1 = NULL, q15_35 = NULL,     e0 = NULL, q0_5 = q05root, q15_45 = q15_45, Sex = Sex, radix = radix,     tol = tol): object 'q05root' not found

# case 6: Using 1q0 and 35q15
L6 <- lt_model_lq(Sex = "b", q0_1 = 0.05, q15_35 = 0.125)

#> Error in lt_model_lq(fitted_logquad = fitted_logquad, q0_1 = NULL, q15_45 = NULL,     e0 = NULL, q0_5 = q05root, q15_35 = q15_35, Sex = Sex, radix = radix,     tol = tol): object 'q05root' not found

# case 7: Using 45q15 and e0
L7 <- lt_model_lq(Sex = "b", q15_45 = 0.125, e0 = 65)

#> Error in lthat.logquad(coefs = cf, x = x, q0_5 = q0_5, k = k, q0_1 = NULL,     q15_45 = NULL, q15_35 = NULL, radix, Sex = Sex): unused arguments (q0_1 = NULL, q15_45 = NULL, q15_35 = NULL)

# case 8: Using 35q15 and e0
L8 <- lt_model_lq(Sex = "b", q15_35 = 0.15, e0 = 65)

#> Error in lthat.logquad(coefs = cf, x = x, q0_5 = q0_5, k = k, q0_1 = NULL,     q15_45 = NULL, q15_35 = NULL, radix, Sex = Sex): unused arguments (q0_1 = NULL, q15_45 = NULL, q15_35 = NULL)