Extrapolate old-age human mortality curve using mortality laws

Extrapolate old-age human mortality curve using mortality laws

lt_rule_m_extrapolate(
  mx,
  x,
  x_fit = x,
  x_extr,
  law = "kannisto",
  opt.method = "poissonL",
  ...
)

Arguments

mx
Vector or matrix of age specific death-rates.

x
Vector of ages at the beginning of the age interval.

x_fit
Ages to be considered in estimating the mortality model parameters. x_fit can be a subset of x. However, after the model is identifies fitted values and residuals are computed for all ages in x.

x_extr
Ages for which to extrapolate the death-rates.

law
The name of the mortality law/model to be used. The following options are available:

"kannisto" -- The Kannisto model;

"kannisto_makeham" -- The Kannisto-Makeham model;

"gompertz" -- The Gompertz model;

"ggompertz" -- The Gamma-Gompertz model;

"makeham" -- The Makeham model;

"beard" -- The Beard model;

"beard_makeham" -- The Beard-Makeham model;

"quadratic" -- The Quadratic model.

opt.method
How would you like to find the parameters? Specify the function to be optimize. Available options: the Poisson likelihood function poissonL; the Binomial likelihood function -binomialL; and 6 other loss functions. For more details, check the availableLF function.

...
Other arguments to be passed on to the MortalityLaw function.

Value

An object of class lt_rule_m_extrapolate with the following components:

input

List with arguments provided in input. Saved for convenience.

call

An unevaluated function call, that is, an unevaluated expressionwhich consists of the named function applied to the given arguments.

fitted.model

An object of class MortalityLaw. Here one can find fitted values, residuals, goodness of fit measures etc.

values

A vector or matrix containing the complete mortality data, that is the modified input data following the extrapolation procedure.

See also

Author

Marius D. Pascariu rpascariu@outlook.com

Examples

# Example 1 - abridged data

# Age-specific death rates
mx <- c(.0859, .0034, .0009, .0007, .0016, .0029, .0036, .0054,
        .0053, .0146, .0127, .0269, .0170, .0433, .0371, .0784,
        .0930, .1399, .1875, .2250, .2500, .3000)
# Vector of ages
x <- c(0, 1, seq(5, 100, by = 5))
names(mx) <- x

# Fit the models / Extrapolate the mortality curve
x_fit  = c(80, 85, 90, 95, 100)
x_extr = 90:110
f1 <- lt_rule_m_extrapolate(mx, x, x_fit, x_extr, law = "kannisto")
f2 <- lt_rule_m_extrapolate(mx, x, x_fit, x_extr, law = "kannisto_makeham")
f3 <- lt_rule_m_extrapolate(mx, x, x_fit, x_extr, law = "gompertz")
f4 <- lt_rule_m_extrapolate(mx, x, x_fit, x_extr, law = "ggompertz")
f5 <- lt_rule_m_extrapolate(mx, x, x_fit, x_extr, law = "makeham")
f6 <- lt_rule_m_extrapolate(mx, x, x_fit, x_extr, law = "beard")
f7 <- lt_rule_m_extrapolate(mx, x, x_fit, x_extr, law = "beard_makeham")
f8 <- lt_rule_m_extrapolate(mx, x, x_fit, x_extr, law = "quadratic")

# Plot the results
if (FALSE) {
par(mfrow = c(1, 2))
plot(x, mx, pch = 16, xlim = c(60, 110), ylim = c(0, 0.6), cex = 1.5)
points(x_fit, mx[paste(x_fit)], pch = 16, col = 4, cex = 1.5)
lines(x_extr, f1$values[paste(x_extr)], lty = 1, col = 2, lwd = 2)
lines(x_extr, f2$values[paste(x_extr)], lty = 2, col = 3, lwd = 2)
lines(x_extr, f3$values[paste(x_extr)], lty = 3, col = 4, lwd = 2)
lines(x_extr, f4$values[paste(x_extr)], lty = 4, col = 5, lwd = 2)
lines(x_extr, f5$values[paste(x_extr)], lty = 5, col = 6, lwd = 2)
lines(x_extr, f6$values[paste(x_extr)], lty = 6, col = 7, lwd = 2)
lines(x_extr, f7$values[paste(x_extr)], lty = 7, col = 8, lwd = 2)
lines(x_extr, f8$values[paste(x_extr)], lty = 8, col = 9, lwd = 2)

legend("topleft", bty = "n",
       legend = c("Obs. Values", "Obs. Values used in fitting",
                  "Kannisto", "Kannisto-Makeham", "Gompertz", "Gamma-Gompertz",
                  "Makeham", "Beard", "Beard-Makeham", "Quadratic"),
       lty = c(NA, NA, 1:8), pch = c(16, 16, rep(NA, 8)),
       col = c(1, 4, 2:9), lwd = 2, pt.cex = 2)
}

# ----------------------------------------------
# Example 2 - 1-year age data

# Age-specific death rates
mx1 <- c(.0070, .0082, .0091, .0096, .0108, .0122, .0141, .0150, .0165, .0186,
         .0205, .0229, .0259, .0294, .0334, .0379, .0426, .0482, .0550, .0628,
         .0716, .0806, .0897, .1003, .1149, .1264, .1558, .1563, .1812, .2084,
         .2298, .2536, .2813, .3143, .3352, .3651, .4128)
# Vector of ages
x1 <- 65:101
names(mx1) <- x1

# Fit the models / Extrapolate the mortality curve
x_fit = 80:95
x_extr = 80:125
g1 <- lt_rule_m_extrapolate(mx1, x1, x_fit, x_extr, law = "kannisto")
g2 <- lt_rule_m_extrapolate(mx1, x1, x_fit, x_extr, law = "kannisto_makeham")
g3 <- lt_rule_m_extrapolate(mx1, x1, x_fit, x_extr, law = "gompertz")
g4 <- lt_rule_m_extrapolate(mx1, x1, x_fit, x_extr, law = "ggompertz")
g5 <- lt_rule_m_extrapolate(mx1, x1, x_fit, x_extr, law = "makeham")
g6 <- lt_rule_m_extrapolate(mx1, x1, x_fit, x_extr, law = "beard")
g7 <- lt_rule_m_extrapolate(mx1, x1, x_fit, x_extr, law = "beard_makeham")
g8 <- lt_rule_m_extrapolate(mx1, x1, x_fit, x_extr, law = "quadratic")

# Plot
if (FALSE) {
plot(x1, mx1, log = "y", ylim = c(0.001, 5),
     pch = 16, xlim = c(65, 125), cex = 1.3)
points(x_fit, mx1[paste(x_fit)], pch = 16, col = 4, cex = 1.5)
lines(x_extr, g1$values[paste(x_extr)], lty = 1, col = 2, lwd = 2)
lines(x_extr, g2$values[paste(x_extr)], lty = 2, col = 3, lwd = 2)
lines(x_extr, g3$values[paste(x_extr)], lty = 3, col = 4, lwd = 2)
lines(x_extr, g4$values[paste(x_extr)], lty = 4, col = 5, lwd = 2)
lines(x_extr, g5$values[paste(x_extr)], lty = 5, col = 6, lwd = 2)
lines(x_extr, g6$values[paste(x_extr)], lty = 6, col = 7, lwd = 2)
lines(x_extr, g7$values[paste(x_extr)], lty = 7, col = 8, lwd = 2)
lines(x_extr, g8$values[paste(x_extr)], lty = 8, col = 9, lwd = 2)

legend("topleft", bty = "n",
       legend = c("Obs. Values", "Obs. Values used in fitting",
                  "Kannisto", "Kannisto-Makeham", "Gompertz", "Gamma-Gompertz",
                  "Makeham", "Beard", "Beard-Makeham", "Quadratic"),
       lty = c(NA, NA, 1:8), pch = c(16, 16, rep(NA, 8)),
       col = c(1, 4, 2:9), lwd = 2, pt.cex = 2)
}
# ----------------------------------------------
# Example 3 - Extrapolate mortality for multiple years at once

# Create some data
mx_matrix <- matrix(rep(mx1, 3), ncol = 3) %*% diag(c(1, 1.05, 1.1))
dimnames(mx_matrix) <- list(age = x1, year = c("year1", "year2", "year3"))

F1 <- lt_rule_m_extrapolate(mx_matrix, x = x1, x_fit, x_extr, law = "kannisto")

#> Warning: the condition has length > 1 and only the first element will be used

F1

#> $input
#> $input$opt.choices
#> [1] "poissonL"  "LF2"       "LF1"       "LF3"       "LF4"       "LF5"      
#> [7] "LF6"       "binomialL"
#> 
#> $input$all_the_laws_we_care_about
#> [1] "kannisto"         "kannisto_makeham" "makeham"          "gompertz"        
#> [5] "ggompertz"        "beard"            "beard_makeham"    "quadratic"       
#> 
#> $input$mx
#>      year
#> age    year1    year2   year3
#>   65  0.0070 0.007350 0.00770
#>   66  0.0082 0.008610 0.00902
#>   67  0.0091 0.009555 0.01001
#>   68  0.0096 0.010080 0.01056
#>   69  0.0108 0.011340 0.01188
#>   70  0.0122 0.012810 0.01342
#>   71  0.0141 0.014805 0.01551
#>   72  0.0150 0.015750 0.01650
#>   73  0.0165 0.017325 0.01815
#>   74  0.0186 0.019530 0.02046
#>   75  0.0205 0.021525 0.02255
#>   76  0.0229 0.024045 0.02519
#>   77  0.0259 0.027195 0.02849
#>   78  0.0294 0.030870 0.03234
#>   79  0.0334 0.035070 0.03674
#>   80  0.0379 0.039795 0.04169
#>   81  0.0426 0.044730 0.04686
#>   82  0.0482 0.050610 0.05302
#>   83  0.0550 0.057750 0.06050
#>   84  0.0628 0.065940 0.06908
#>   85  0.0716 0.075180 0.07876
#>   86  0.0806 0.084630 0.08866
#>   87  0.0897 0.094185 0.09867
#>   88  0.1003 0.105315 0.11033
#>   89  0.1149 0.120645 0.12639
#>   90  0.1264 0.132720 0.13904
#>   91  0.1558 0.163590 0.17138
#>   92  0.1563 0.164115 0.17193
#>   93  0.1812 0.190260 0.19932
#>   94  0.2084 0.218820 0.22924
#>   95  0.2298 0.241290 0.25278
#>   96  0.2536 0.266280 0.27896
#>   97  0.2813 0.295365 0.30943
#>   98  0.3143 0.330015 0.34573
#>   99  0.3352 0.351960 0.36872
#>   100 0.3651 0.383355 0.40161
#>   101 0.4128 0.433440 0.45408
#> 
#> $input$x
#>  [1]  65  66  67  68  69  70  71  72  73  74  75  76  77  78  79  80  81  82  83
#> [20]  84  85  86  87  88  89  90  91  92  93  94  95  96  97  98  99 100 101
#> 
#> $input$x_fit
#>  [1] 80 81 82 83 84 85 86 87 88 89 90 91 92 93 94 95
#> 
#> $input$x_extr
#>  [1]  80  81  82  83  84  85  86  87  88  89  90  91  92  93  94  95  96  97  98
#> [20]  99 100 101 102 103 104 105 106 107 108 109 110 111 112 113 114 115 116 117
#> [39] 118 119 120 121 122 123 124 125
#> 
#> $input$law
#> [1] "kannisto"
#> 
#> $input$opt.method
#> [1] "poissonL"
#> 
#> 
#> $call
#> lt_rule_m_extrapolate(mx = mx_matrix, x = x1, x_fit = x_fit, 
#>     x_extr = x_extr, law = "kannisto")
#> 
#> $fitted.model
#> Kannisto model: mu[x] = A exp(Bx) / [1 + A exp(Bx)]
#> Fitted values: mx 
#> 
#> $values
#>          year1      year2      year3
#> 65  0.00700000 0.00735000 0.00770000
#> 66  0.00820000 0.00861000 0.00902000
#> 67  0.00910000 0.00955500 0.01001000
#> 68  0.00960000 0.01008000 0.01056000
#> 69  0.01080000 0.01134000 0.01188000
#> 70  0.01220000 0.01281000 0.01342000
#> 71  0.01410000 0.01480500 0.01551000
#> 72  0.01500000 0.01575000 0.01650000
#> 73  0.01650000 0.01732500 0.01815000
#> 74  0.01860000 0.01953000 0.02046000
#> 75  0.02050000 0.02152500 0.02255000
#> 76  0.02290000 0.02404500 0.02519000
#> 77  0.02590000 0.02719500 0.02849000
#> 78  0.02940000 0.03087000 0.03234000
#> 79  0.03340000 0.03507000 0.03674000
#> 80  0.03724804 0.03900214 0.04079927
#> 81  0.04241843 0.04444163 0.04651936
#> 82  0.04827054 0.05059980 0.05299710
#> 83  0.05488373 0.05755989 0.06031979
#> 84  0.06234361 0.06541139 0.06858100
#> 85  0.07074155 0.07424951 0.07787986
#> 86  0.08017400 0.08417424 0.08831999
#> 87  0.09074133 0.09528903 0.10000789
#> 88  0.10254620 0.10769888 0.11305071
#> 89  0.11569142 0.12150786 0.12755347
#> 90  0.13027708 0.13681592 0.14361546
#> 91  0.14639719 0.15371509 0.16132605
#> 92  0.16413553 0.17228500 0.18075969
#> 93  0.18356096 0.19258776 0.20197059
#> 94  0.20472230 0.21466258 0.22498690
#> 95  0.22764298 0.23852020 0.24980507
#> 96  0.25231569 0.26413755 0.27638462
#> 97  0.27869757 0.29145317 0.30464381
#> 98  0.30670638 0.32036380 0.33445690
#> 99  0.33621796 0.35072270 0.36565336
#> 100 0.36706573 0.38234006 0.39801947
#> 101 0.39904231 0.41498594 0.43130259
#> 102 0.43190364 0.44839570 0.46521802
#> 103 0.46537551 0.48227782 0.49945816
#> 104 0.49916214 0.51632369 0.53370337
#> 105 0.53295642 0.55021870 0.56763386
#> 106 0.56645097 0.58365377 0.60094146
#> 107 0.59934920 0.61633649 0.63334060
#> 108 0.63137561 0.64800092 0.66457751
#> 109 0.66228447 0.67841558 0.69443716
#> 110 0.69186649 0.70738894 0.72274761
#> 111 0.71995298 0.73477246 0.74938170
#> 112 0.74641769 0.76046110 0.77425636
#> 113 0.77117635 0.78439176 0.79732984
#> 114 0.79418423 0.80653988 0.81859750
#> 115 0.81543237 0.82691489 0.83808660
#> 116 0.83494279 0.84555495 0.85585058
#> 117 0.85276319 0.86252134 0.87196337
#> 118 0.86896156 0.87789306 0.88651392
#> 119 0.88362106 0.89176173 0.89960128
#> 120 0.89683528 0.90422706 0.91133035
#> 121 0.90870412 0.91539302 0.92180827
#> 122 0.91933030 0.92536462 0.93114159
#> 123 0.92881655 0.93424540 0.93943404
#> 124 0.93726341 0.94213555 0.94678492
#> 125 0.94476754 0.94913048 0.95328799
#> 
#> attr(,"class")
#> [1] "lt_rule_m_extrapolate"

ls(F1)

#> [1] "call"         "fitted.model" "input"        "values"      

coef(F1)

#> NULL